% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{aggregate_payoff}
\alias{aggregate_payoff}
\title{Aggregate payoff}
\usage{
aggregate_payoff(df_month_day, r = 0.03)
}
\description{
Aggregate payoff
}
\examples{
nmonth = 1:12
t_now = "2012-01-01"
t_hor = "2013-01-01"
}
\keyword{internal}
