% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMA_model-internals.R
\name{AR_params_to_phi}
\alias{AR_params_to_phi}
\title{AR reparametrization}
\usage{
AR_params_to_phi(zeta)
}
\arguments{
\item{zeta}{Numeric vector of length p, unconstraint parameters.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Map unconstraint AR parameters into contraint parameters
}
\note{
Version 1.0.2
}
\seealso{
Other ARMA-params: 
\code{\link{ARMA_companion_matrix}()},
\code{\link{ARMA_params_to_zeta}()},
\code{\link{ARMA_vector_b}()},
\code{\link{AR_params_to_zeta}()}
}
\concept{ARMA-params}
\keyword{ARMA}
