% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_internals.R
\name{AR_variance}
\alias{AR_variance}
\title{Long term variance AR process}
\usage{
AR_variance(phi, sigma2 = 1)
}
\arguments{
\item{phi}{AR parameters}

\item{sigma2}{variance}
}
\description{
Long term variance AR process
}
\examples{
AR_variance(0.4)
AR_variance(c(0.4, 0.2))
AR_variance(c(0.4, 0.2, 0.1))
AR_variance(c(0.3, 0.2, 0.1, 0.05))
}
