% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMA_model-internals.R
\name{ARMA_forecast}
\alias{ARMA_forecast}
\title{Matrix form of an ARMA model}
\usage{
ARMA_forecast(h, X0, A, b, intercept = 0)
}
\arguments{
\item{h}{integer(1), number of steps ahead.}

\item{X0}{numeric vector of length p + q, state vector of past values.}

\item{A}{numeric matrix of size p+q x p+q. See the function \code{\link{ARMA_companion_matrix}} for more details.}

\item{b}{mumeric vector of length p+q. See the function \code{\link{ARMA_vector_b}} for more details.}

\item{intercept}{Numeric scalar, intercept parameter.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Fast ARMA state-space h-step forecast and weights (C/BLAS)
}
\note{
Version 1.0.2
}
\examples{
h <- 1000
X0 <- c(0.2, 0.1)
A <- ARMA_companion_matrix(0.2, 0.1)
b <- ARMA_vector_b(1,1)
ARMA_forecast(h, X0, A, b, intercept = 0)

}
\seealso{
Other ARMA-moments: 
\code{\link{ARMA_covariance}()},
\code{\link{ARMA_expectation}()},
\code{\link{ARMA_filter}()},
\code{\link{ARMA_next_step}()},
\code{\link{ARMA_variance}()}
}
\concept{ARMA-moments}
\keyword{ARMA}
