% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMA_model-internals.R
\name{ARMA_next_step}
\alias{ARMA_next_step}
\title{Next-step value of an ARMA process}
\usage{
ARMA_next_step(h = 1, X0, A, b, intercept = 0, eps = 0)
}
\arguments{
\item{h}{integer(1), number of steps ahead.}

\item{X0}{numeric vector of length p + q, state vector of past values.}

\item{A}{numeric matrix of size p+q x p+q. See the function \code{\link{ARMA_companion_matrix}} for more details.}

\item{b}{mumeric vector of length p+q. See the function \code{\link{ARMA_vector_b}} for more details.}

\item{intercept}{Numeric scalar, intercept parameter.}

\item{eps}{optional numeric vector of length h. Next step new residuals.}
}
\description{
Next-step value of an ARMA process
}
\note{
Version 1.0.2
}
\examples{
# Companion matrix and vector b
A <- ARMA_companion_matrix(c(0.4), c())
b <- ARMA_vector_b(1,0)
# Initial value
X0 <- c(0.9)

# Next step forecast
ARMA_next_step(1, X0, A, b, intercept = 0.2, eps = 0)
# 3-steps ahead forecast
ARMA_next_step(3,X0, A, b, intercept = 0.2, eps = 0)
# 10-step ahead forecast
ARMA_next_step(10, X0, A, b, intercept = 0.2, eps = 0)

# Next step simulation
eps <- rnorm(1)
ARMA_next_step(1, X0, A, b, intercept = 0.2, eps = eps)
# 3-steps ahead simulation
eps <- rnorm(3)
ARMA_next_step(3, X0, A, b, intercept = 0.2, eps = eps)
# 10-step ahead simulation
eps <- rnorm(10)
ARMA_next_step(10, X0, A, b, intercept = 0.2, eps = eps)

}
\keyword{ARMA}
