% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMA_model-internals.R
\name{ARMA_params_to_zeta}
\alias{ARMA_params_to_zeta}
\title{ARMA reparametrization}
\usage{
ARMA_params_to_zeta(phi, theta)
}
\arguments{
\item{zeta_phi}{Numeric vector of length p, AR parameters.}

\item{zeta_theta}{Numeric vector of length q, MA parameters.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Map constraint ARMA parameters into uncontraint parameters
}
\note{
Version 1.0.2
}
\examples{
phi <- c(0.3, 0.4, 0.1)
theta <- c(0.1, 0.8)
zeta <- ARMA_params_to_zeta(phi, theta)

}
\seealso{
Other ARMA-params: 
\code{\link{ARMA_companion_matrix}()},
\code{\link{ARMA_vector_b}()},
\code{\link{AR_params_to_phi}()},
\code{\link{AR_params_to_zeta}()}
}
\concept{ARMA-params}
\keyword{ARMA}
