% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMA_model-internals.R
\name{ARMA_variance}
\alias{ARMA_variance}
\title{Compute the long-term variance of an ARMA model}
\usage{
ARMA_variance(h = 1, A, b, sigma2 = 1)
}
\arguments{
\item{h}{integer(1), number of steps ahead.}

\item{A}{numeric matrix of size p+q x p+q. See the function \code{\link{ARMA_companion_matrix}} for more details.}

\item{b}{mumeric vector of length p+q. See the function \code{\link{ARMA_vector_b}} for more details.}

\item{sigma2}{integer(1), std. deviation of the residuals.}
}
\description{
Compute the long-term variance of an ARMA model
}
\note{
Version 1.0.2
}
\examples{
A <- ARMA_companion_matrix(c(0.4, 0.1), c(0.1, 0.05))
b <- ARMA_vector_b(2,2)
ARMA_variance(h = 3, A, b, sigma2 = 1)
}
\keyword{ARMA}
