% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarScenario.R
\name{as_solarScenario}
\alias{as_solarScenario}
\title{Extract and structure simulations from a `solarScenarioSpec`}
\usage{
as_solarScenario(simSpec)
}
\arguments{
\item{simSpec}{object with the class `solarScenario_spec`. See the function \code{\link{solarScenario_spec}} for details.}
}
\description{
Extract and structure simulations from a `solarScenarioSpec`
}
\keyword{internal}
\keyword{solarScenario}
