% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bologna}
\alias{Bologna}
\title{Bologna model specification}
\format{
An R6 object of class \code{solarModel_spec}.
}
\usage{
data(Bologna)
}
\description{
A saved \code{solarModel_spec} object for Bologna. The data file
\code{Bologna.RData} loads an object named \code{spec}.
}
\details{
The object is included as an example package specification. It has
\code{place} set to \code{"Bologna"} and \code{target} set to \code{"GHI"}.
}
\keyword{datasets}
