% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnorm_mix.R
\name{boot_dnorm_mix}
\alias{boot_dnorm_mix}
\title{Bootstrap Parameters of Normal Mixture}
\usage{
boot_dnorm_mix(
  x,
  params,
  B = 50,
  ci = 0.95,
  sample_perc = 0.8,
  loss = "kl",
  seed = 1,
  na.rm = TRUE
)
}
\arguments{
\item{x}{vector}

\item{params}{initial parameters}

\item{B}{number of bootstraps}

\item{ci}{confidence interval for empirical quantiles}

\item{loss}{loss type. Can be `ml` for maximum likelihood or `kl` for kl_dist.}

\item{seed}{random seed}

\item{na.rm}{logical.}
}
\description{
Bootstrap the parameter for the density function for a normal mixture with two components.
}
\examples{
params <- c(mu1 = -2, mu2 = 2, sd1 = 3, sd2 = 1, p = 0.5)
n1 <- rnorm(t_bar, mean = params[1], sd = params[3])
n2 <- rnorm(t_bar, mean = params[2], sd = params[4])
Z <- rbinom(t_bar, 1, params[5])
x <- Z * n1 + (1-Z)*n2
boot_dnorm_mix(x, params = init_params,  B = 50, ci = 0.95, sample_perc = 0.8, loss = "ml")$par
boot_dnorm_mix(x, params = init_params,  B = 50, ci = 0.95, sample_perc = 0.8, loss = "kl")$par

}
