% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundTransform-internals.R
\name{boundTransform_link.gumbel}
\alias{boundTransform_link.gumbel}
\title{Link function (Gumbel)}
\usage{
boundTransform_link.gumbel()
}
\description{
Link function (Gumbel)
}
\note{
Version 1.0.0
}
\keyword{utils}
