% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundTransform-internals.R
\name{boundTransform_link.identity}
\alias{boundTransform_link.identity}
\title{Link function (identity)}
\usage{
boundTransform_link.identity()
}
\description{
Link function (identity)
}
\note{
Version 1.0.0
}
\keyword{utils}
