% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundTransform-internals.R
\name{boundTransform_link.invgumbel}
\alias{boundTransform_link.invgumbel}
\title{Link function (invGumbel)}
\usage{
boundTransform_link.invgumbel()
}
\description{
Link function (invGumbel)
}
\note{
Version 1.0.0
}
\keyword{utils}
