% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundTransform-internals.R
\name{boundTransform_link.logis}
\alias{boundTransform_link.logis}
\title{Link function (logit)}
\usage{
boundTransform_link.logis()
}
\description{
Link function (logit)
}
\note{
Version 1.0.0
}
\keyword{utils}
