% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundTransform-internals.R
\name{boundTransform_link.norm}
\alias{boundTransform_link.norm}
\title{Link function (probit)}
\usage{
boundTransform_link.norm()
}
\description{
Link function (probit)
}
\note{
Version 1.0.0
}
\keyword{utils}
