% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiationModel_internals.R
\name{calibrate_dQdP_solar}
\alias{calibrate_dQdP_solar}
\title{Calibration for the best lambda}
\usage{
calibrate_dQdP_solar(data_dQdP, r = 0, nmonths = 1:12)
}
\description{
Calibration for the best lambda
}
