% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAMS.R
\name{CAMS}
\alias{CAMS}
\title{CAMS Data (Location object)}
\usage{
CAMS(place, year_max = lubridate::year(Sys.Date()))
}
\arguments{
\item{place}{place}
}
\description{
CAMS Data (Location object)
}
\examples{
object <- CAMS("Bologna")

place <- "Bologna"
}
