% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CAMS_data}
\alias{CAMS_data}
\title{CAMS data by location}
\format{
A named list. Each element is a tibble with 7,465 rows and 8
columns:
\describe{
\item{date}{Date.}
\item{n}{Numeric day index.}
\item{Year}{Numeric year.}
\item{Month}{Numeric month.}
\item{Day}{Integer day of month.}
\item{GHI}{Numeric global horizontal irradiance value.}
\item{clearsky}{Numeric clear-sky radiation value.}
\item{H0}{Numeric extraterrestrial radiation value.}
}
}
\usage{
data(CAMS_data)
}
\description{
A list of daily solar radiation data sets for selected locations.
}
\details{
The list contains named elements for the locations available to
\code{solarModel_spec$specification()}. Individual elements include \code{place} and
\code{coords} attributes.
}
\keyword{datasets}
