% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalModel.R
\name{clearsky.seasonalModel}
\alias{clearsky.seasonalModel}
\title{Seasonal Model Clear sky Radiation}
\usage{
clearsky.seasonalModel(object, control = control.seasonalModel())
}
\description{
Seasonal Model Clear sky Radiation
}
\examples{
object <- CAMS("Amsterdam")
control = control.seasonalModel()
}
