% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalClearsky-R6-internals.R
\name{clearsky_delta}
\alias{clearsky_delta}
\title{Optimizer for Solar Clear sky}
\usage{
clearsky_delta(x, Ct, lower = 0, upper = 3, by = 0.01, ntol = 0)
}
\arguments{
\item{x}{Numeric vector, time series of solar radiation.}

\item{Ct}{Numeric vector, time series of  clear sky radiation.}

\item{lower}{Numeric scalar, lower bound for grid of delta parameters used for optimization. Default is \code{0}.}

\item{upper}{Numeric scalar, upper bound for grid of delta parameters used for optimization. Default is \code{3}.}

\item{by}{Numeric scalar, step for the grid. Default is \code{0.01}.}

\item{ntol}{Integer scalar, Tolerance for the maximum number of violations admitted of the condition \code{Ct > x}. The default is \code{0}.}
}
\description{
Find the best parameter delta for fitting clear sky radiation.
}
\details{
Detect the best parameter \code{delta} such that the condition \code{x > delta * Ct} is satisfied for all the observations except for \code{ntol} observations.
}
\note{
Version 1.0.1
}
\keyword{clearsky}
