% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalClearsky.R
\name{clearsky_delta_optimizer}
\alias{clearsky_delta_optimizer}
\title{Optimizer for Solar Clear sky}
\usage{
clearsky_delta_optimizer(x, Ct, lower = 0, upper = 3, by = 0.01, ntol = 0)
}
\arguments{
\item{x}{Numeric vector. Time series of solar radiation.}

\item{Ct}{Numeric vector. Time series of  clear sky radiation.}

\item{lower}{Numeric scalar. Lower bound for grid of delta parameters used for optimization. Default is \code{0}.}

\item{upper}{Numeric scalar. Upper bound for grid of delta parameters used for optimization. Default is \code{3}.}

\item{by}{Numeric scalar, step for the grid. Default is \code{0.01}.}

\item{ntol}{Integer scalar. Tolerance for the maximum number of violations admitted of the condition \code{clearsky > GHI}. Default is \code{0}.}
}
\value{
Numeric, scalar the optimal delta parameter.
}
\description{
Find the best parameter delta for fitting clear sky radiation.
}
\keyword{internal}
