% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearskySeasonal.R
\name{clearsky_optimize}
\alias{clearsky_optimize}
\title{Optimizer for Solar Clear sky}
\usage{
clearsky_optimize(x, Ct, lower = 0, upper = 3, by = 0.01, ntol = 30)
}
\arguments{
\item{x}{vector of solar radiation}

\item{Ct}{vector of extraterrestrial radiation}

\item{lower}{numeric, lower bound for delta grid.}

\item{upper}{numeric, upper bound for delta grid.}

\item{by}{numeric, step for delta grid.}

\item{ntol}{integer, tolerance for `clearsky > GHI` condition. Maximum number of violations admitted.}
}
\value{
the optimal delta
}
\description{
Find the best parameter delta for fitting clear sky radiation.
}
