% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalClearsky-R6-internals.R
\name{clearsky_optimizer}
\alias{clearsky_optimizer}
\title{Fit the seasonal model for clear sky radiation.}
\usage{
clearsky_optimizer(
  x,
  dates,
  clearsky,
  lat,
  spec = seasonalClearsky_spec(),
  control = control_seasonalClearsky(),
  seasonal_model_Ct
)
}
\arguments{
\item{x}{Numeric vector, time series of solar radiation.}

\item{dates}{Character or Date vector, time series of dates.}

\item{clearsky}{Numeric vector, time series of target clear sky radiation.}

\item{lat}{Numeric scalar, reference latitude. Overwritten when \code{seasonal_model_Ct} is specified.}

\item{spec}{See the function \code{\link{seasonalClearsky_spec}} for more details.
Overwritten when \code{seasonal_model_Ct} is specified.}

\item{control}{See the function \code{\link{control_seasonalClearsky}} for more details.
Overwritten when \code{seasonal_model_Ct} is specified.}

\item{seasonal_model_Ct}{Optional object of the class \code{\link{seasonalClearsky}}.
When submitted \code{lat}, \code{spec} and \code{control} parameters used are recovered from the specified model.}
}
\description{
Fit the seasonal model for clear sky radiation.
}
\note{
Version 1.0.0
}
\examples{
x  = solarr::spec$data$GHI
dates  = solarr::spec$data$date
lat = solarr::spec$coords$lat
clearsky = solarr::spec$data$clearsky
# Without a clearsky model
seasonal_model_Ct <- clearsky_optimizer(x, dates, clearsky, lat)
# With a clearsky model
clearsky_optimizer(x, dates, clearsky, seasonal_model_Ct = seasonal_model_Ct)
}
\keyword{clearsky}
