% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalClearsky-R6-internals.R
\name{clearsky_optimizer_CLS}
\alias{clearsky_optimizer_CLS}
\title{Fully constrained optimization (Constrained Least Squares)}
\usage{
clearsky_optimizer_CLS(
  seasonal_model_Ct,
  newdata,
  eps = 0.01,
  delta_optimize = TRUE
)
}
\arguments{
\item{seasonal_model_Ct}{An object of the class \code{seasonalClearsky}. See \code{\link{seasonalClearsky}} for more details.}

\item{newdata}{dataset}

\item{eps}{threshold parameter}
}
\description{
Fully constrained optimization (Constrained Least Squares)
}
\note{
Version 1.0.0
}
\keyword{clearsky}
