% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalClearsky-R6-internals.R
\name{clearsky_optimizer_delta}
\alias{clearsky_optimizer_delta}
\title{Delta optimizer on a clear-sky model}
\usage{
clearsky_optimizer_delta(seasonal_model_Ct, newdata)
}
\arguments{
\item{seasonal_model_Ct}{An object of the class \code{seasonalClearsky}. See \code{\link{seasonalClearsky}} for more details.}

\item{newdata}{dataset}
}
\description{
Delta optimizer on a clear-sky model
}
\note{
Version 1.0.0
}
\keyword{clearsky}
