% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalClearsky-R6-internals.R
\name{clearsky_optimizer_RLS}
\alias{clearsky_optimizer_RLS}
\title{Penalized Least-Squares}
\usage{
clearsky_optimizer_RLS(
  seasonal_model_Ct,
  newdata,
  lambda = 0,
  delta_optimize = TRUE
)
}
\arguments{
\item{seasonal_model_Ct}{An object of the class \code{seasonalClearsky}. See \code{\link{seasonalClearsky}} for more details.}

\item{newdata}{dataset}

\item{lambda}{penalization parameter.}
}
\description{
Penalized Least-Squares
}
\note{
Version 1.0.0
}
\keyword{clearsky}
