% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalClearsky.R
\name{clearsky_outliers}
\alias{clearsky_outliers}
\title{Impute clear sky outliers}
\usage{
clearsky_outliers(x, Ct, date, threshold = 1e-04, quiet = FALSE)
}
\arguments{
\item{x}{Numeric vector, realized solar radiation.}

\item{Ct}{Numeric vector, clear sky radiation.}

\item{date}{Vector, optional time series of dates.
It will be used for a more precise imputation when a solar radiation value is \code{NA}.}
}
\description{
Detect and impute outliers with respect to a maximum level of radiation (Ct)
}
\examples{
clearsky_outliers(c(1,2,3), 2)

}
\keyword{internal}
