% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearskyModel.R
\name{clearskyModel}
\alias{clearskyModel}
\alias{clearskyModel_fit}
\title{Seasonal model for clear sky radiation}
\usage{
clearskyModel_fit(data, control = clearskyModel_control())
}
\arguments{
\item{data}{dataset}

\item{control}{list of control parameters. See `clearskyModel_control()` for details.}
}
\description{
Fit a seasonal model for clear sky radiation in a location.
}
