% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearskyModel.R
\name{clearskyModel_control}
\alias{clearskyModel_control}
\title{Control for `clearskyModel`}
\usage{
clearskyModel_control(
  method = "II",
  include.intercept = TRUE,
  order = 1,
  period = 365,
  seed = 1,
  delta_init = 1.1,
  tol = 30,
  lower = 0,
  upper = 1,
  by = 0.001,
  quiet = FALSE
)
}
\arguments{
\item{method}{character, method for clearsky estimate, can be `I` or `II`.}

\item{include.intercept}{logical. When `TRUE`, the default, the intercept will be included in the model.}

\item{order}{numeric, of sine and cosine elements.}

\item{period}{numeric, periodicity. The default is `2*base::pi/365`.}

\item{seed}{numeric, random seed for reproducibility. It is used to random impute the violations.}

\item{delta_init}{Value for delta init in the clearsky model.}

\item{tol}{integer, numeric tolerance for clearsky > GHI condition. Maximum number of violations admitted. Used in `clearskyModel_optimize()`.}

\item{lower}{numeric, lower bound for delta grid, see `clearskyModel_optimize()`.}

\item{upper}{numeric, upper bound for delta grid, see `clearskyModel_optimize()`.}

\item{by}{numeric, step for delta grid, see `clearskyModel_optimize()`.}

\item{quiet}{logical. When `FALSE`, the default, the functions displays warning or messages.}
}
\description{
Control parameters for a `clearskyModel` object.
}
