% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearskyModel.R
\name{clearskyModel_predict}
\alias{clearskyModel_predict}
\title{Predict method}
\usage{
clearskyModel_predict(object, n = 1)
}
\arguments{
\item{object}{an object of the class `cleaskyModel`}

\item{n}{number of day of the year.}
}
\description{
Predict method for the class `cleaskyModel`.
}
