% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearskySeasonal.R
\name{clearskySeasonal}
\alias{clearskySeasonal}
\title{Seasonal model for clear sky radiation}
\usage{
clearskySeasonal(
  data,
  seasonal_data,
  method = "II",
  control = control_clearskyModel()
)
}
\arguments{
\item{data}{dataset}

\item{seasonal_data}{dataset with two columns: `n` with the number of the day in 1:365
and `H0` with the extraterrestrial radiation.}

\item{method}{character, method for clearsky estimate, can be `I` or `II`.}

\item{control}{list of control parameters. See \code{\link{control_clearskyModel}} for details.}
}
\description{
Fit a seasonal model for clear sky radiation in a location.
}
