% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control.seasonalModel}
\alias{control.seasonalModel}
\title{seasonalModel control parameters}
\usage{
\method{control}{seasonalModel}(
  object,
  method = "II",
  include.intercept = TRUE,
  order = 1,
  seed = 1,
  delta_init = 1.1,
  tol = 30,
  lower = 0,
  upper = 1,
  by = 0.001,
  quiet = FALSE
)
}
\description{
seasonalModel control parameters
}
