% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control.solarModel}
\alias{control.solarModel}
\title{solarModel control parameters}
\usage{
\method{control}{solarModel}(
  object,
  loss = "ml",
  clearsky.model = control.seasonalModel(),
  mean.model = list(seasonalOrder = 1, arOrder = 2, include.intercept = FALSE),
  variance.model = list(seasonalOrder = 1, match_moments = FALSE),
  threshold = 0.001,
  quiet = FALSE
)
}
\arguments{
\item{loss}{type of loss function for mixture model, `ml` stands for maximum-likelihood, while
`kl` for KL-distance.}

\item{mean.model}{a list of parameters}

\item{variance.model}{a list of parameters}

\item{threshold}{Threshold for the estimation of alpha and beta}

\item{quiet}{logical, when `TRUE` the function will not display any message.}
}
\description{
solarModel control parameters
}
