% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearskySeasonal.R
\name{control_clearskyModel}
\alias{control_clearskyModel}
\title{Control parameters for a `clearskySeasonal` object}
\usage{
control_clearskyModel(
  method = "II",
  include.intercept = TRUE,
  order = 1,
  period = 365,
  seed = 1,
  delta_init = 1.4,
  ntol = 30,
  lower = 0,
  upper = 2,
  by = 0.001,
  quiet = FALSE
)
}
\arguments{
\item{method}{character, method for clearsky estimate, can be `I` or `II`.}

\item{include.intercept}{logical. When `TRUE`, the default, the intercept will be included in the model.}

\item{order}{numeric, of sine and cosine elements.}

\item{period}{numeric, periodicity. The default is `365`.}

\item{seed}{numeric, random seed for reproducibility. It is used to random impute the violations.}

\item{delta_init}{Value for delta init in the clear sky model.}

\item{ntol}{integer, tolerance for `clearsky > GHI` condition. Maximum number of violations admitted.}

\item{lower}{numeric, lower bound for delta grid.}

\item{upper}{numeric, upper bound for delta grid.}

\item{by}{numeric, step for delta grid.}

\item{quiet}{logical. When `FALSE`, the default, the functions displays warning or messages.}
}
\description{
Control parameters for a `clearskySeasonal` object
}
\details{
The parametes `tol`, `lower`, `upper` and `by` are used exclusively in \code{\link{clearskyModel_optimize}}.
}
