% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption_internals.R
\name{control_hedging}
\alias{control_hedging}
\title{Control for hedging}
\usage{
control_hedging(
  n_panels = 1,
  efficiency = 1,
  PUN = 1,
  tick = 1,
  n_contracts = 1
)
}
\arguments{
\item{n_panels}{numeric, number of meters squared of solar panels.}

\item{efficiency}{numeric, mean efficiency of the solar panels.}

\item{PUN}{numeric, mean efficiency of the solar panels.}

\item{tick}{numeric, conversion tick for the monetary payoff of a contract.}

\item{n_contract}{numeric, number of contracts}
}
\description{
Control for hedging
}
