% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalClearsky-R6-internals.R
\name{control_seasonalClearsky}
\alias{control_seasonalClearsky}
\title{Control parameters for a \code{seasonalClearsky} object}
\usage{
control_seasonalClearsky(
  delta0 = 1.4,
  lower = 0,
  upper = 3,
  by = 0.001,
  ntol = 0,
  eps = 0.01,
  lambda = 100,
  quiet = FALSE
)
}
\arguments{
\item{delta0}{Numeric scalar, initial value for the optimization. The estimated clear sky is increased by delta0.}

\item{lower}{Numeric scalar, lower bound for grid of delta parameters used for optimization. Default is \code{0}.}

\item{upper}{Numeric scalar, upper bound for grid of delta parameters used for optimization. Default is \code{3}.}

\item{by}{Numeric scalar, step for the grid. Default is \code{0.01}.}

\item{ntol}{Integer scalar, Tolerance for the maximum number of violations admitted of the condition \code{Ct > x}. The default is \code{0}.}

\item{eps}{threshold parameter}

\item{lambda}{penalization parameter.}

\item{quiet}{Logical, when \code{FALSE}, the default, the functions displays warning or messages.}
}
\value{
Named list of control parameters.
}
\description{
Control parameters for a \code{seasonalClearsky} object
}
\details{
The parameters \code{ntol}, \code{lower}, \code{upper} and \code{by} are used exclusively in \code{\link{clearsky_delta}}.
}
\note{
Version 1.0.1
}
\examples{
control = control_seasonalClearsky()
}
\keyword{clearsky}
