% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control_solarEsscher}
\alias{control_solarEsscher}
\title{solarEsscher control parameters}
\usage{
control_solarEsscher(
  nsim = 200,
  ci = 0.05,
  seed = 1,
  quiet = FALSE,
  n_key_points = 15,
  init_lambda = 0,
  lower_lambda = -1,
  upper_lambda = 1
)
}
\description{
solarEsscher control parameters
}
