% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control_solarOption}
\alias{control_solarOption}
\title{solarOption control parameters}
\usage{
control_solarOption(
  nyears = c(2010, 2022),
  K = 0,
  put = TRUE,
  B = discount_factor()
)
}
\description{
solarOption control parameters
}
