% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialClearsky-R6-internals.R
\name{control_spatialClearsky}
\alias{control_spatialClearsky}
\title{Control parameters for a \code{spatialClearsky} object}
\usage{
control_spatialClearsky(
  lambda = 0,
  eps = 0.1,
  gamma = NA,
  normalize_loss = FALSE,
  penalty = c("none", "ridge"),
  mask_time = NULL,
  osqp_settings = list(max_iter = 30000L)
)
}
\value{
Named list of control parameters.
}
\description{
Control parameters for a \code{spatialClearsky} object
}
\note{
Version 1.0.1
}
\examples{
control = control_spatialClearsky()
}
\keyword{spatialClearsky}
