% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiationModel_internals.R
\name{create_monthly_sequence}
\alias{create_monthly_sequence}
\title{Create a monthly sequence of dates}
\usage{
create_monthly_sequence(t_now, t_hor, last_day = FALSE)
}
\arguments{
\item{t_now}{Character, today date.}

\item{t_hor}{Character, horizon date.}

\item{last_day}{Logical. When \code{TRUE} the last day will be treated as conditional variance otherwise not.}
}
\description{
Create a monthly sequence of dates
}
\examples{
t_now <- "2022-01-01"
t_hor <- "2022-03-24"
create_monthly_sequence("2022-01-01", "2022-03-24")
create_monthly_sequence("2022-01-01", "2022-03-24", TRUE)

}
\keyword{integral}
\keyword{internal}
\keyword{radiationModel}
