% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmix2norm.R
\name{dmix2norm}
\alias{dmix2norm}
\alias{d2mixnorm}
\alias{p2mixnorm}
\alias{q2mixnorm}
\title{Density of a bivariate Gaussian mixture with 2 components.}
\usage{
dmix2norm(
  x,
  means = matrix(0, 4, 2),
  Sigma = rep(list(diag(1, 2)), 4),
  probs = rep(1/4, 4),
  log = FALSE
)
}
\arguments{
\item{x}{matrix, on the rows the observations, on the columns the locations.}

\item{means}{matrix of means for each state.}

\item{Sigma}{List of variance-covariance matrices for each state.}

\item{probs}{vector of joint probabilities for each state.}

\item{log}{Logical, when \code{TRUE}, probabilities are returned as \code{log(p)}.}

\item{log.p}{Logical, when \code{TRUE}, probabilities p are given as \code{log(p)}.}

\item{lower.tail}{Logical, when \code{TRUE}, the default, the computed probabilities are \eqn{\mathbb{P}(X < x)}, otherwise \eqn{\mathbb{P}(X \ge x)}.}
}
\description{
Density of a bivariate Gaussian mixture with 2 components.
}
\details{
Version 1.0.0.
}
\examples{
library(tidyverse)
x <- matrix(runif(200, -3, 3), ncol = 2)
dmix2norm(x)
pmix2norm(x)
qmix2norm(x)

}
\keyword{distributions}
