% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desscher.R
\name{desscher}
\alias{desscher}
\title{Esscher transform of a density}
\usage{
desscher(pdf, theta = 0, lower = -Inf, upper = Inf)
}
\arguments{
\item{pdf}{density function.}

\item{theta}{Esscher parameter.}

\item{lower}{numeric, lower bound for integration, i.e. the lower bound for the pdf.}

\item{upper}{numeric, lower bound for integration, i.e. the upper bound for the pdf.}
}
\description{
Given a function of \code{x}, i.e. \eqn{f_X(x)}, compute its Esscher transform and return again a function of \code{x}.
}
\details{
Given a pdf \eqn{f_X(x)} the function computes its Esscher transform, i.e.

\deqn{\mathcal{E}_{\theta}\{f_X(x)\} = \frac{e^{\theta x} f_X(x)}{\int_{-\infty}^{\infty} e^{\theta x} f_X(x) dx}}
}
\examples{
# Grid of points
grid <- seq(-3, 3, 0.1)
# Density function of x
pdf <- function(x) dnorm(x, mean = 0)
# Esscher density (no transform)
esscher_pdf <- desscher(pdf, theta = 0)
pdf(grid) - esscher_pdf(grid)
# Esscher density (transform)
esscher_pdf_1 <- function(x) dnorm(x, mean = -0.1)
esscher_pdf_2 <- desscher(pdf, theta = -0.1)
esscher_pdf_1(grid) - esscher_pdf_2(grid)
# Log-probabilities
esscher_pdf(grid, log = TRUE)
esscher_pdf_2(grid, log = TRUE)

}
