% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desscherMixture.R
\name{desscherMixture}
\alias{desscherMixture}
\alias{pesscherMixture}
\title{Esscher-distorted density and distribution of a Gaussian Mixture}
\usage{
desscherMixture(mean = c(0, 0), sd = c(1, 1), alpha = c(0.5, 0.5), theta = 0)

pesscherMixture(mean = c(0, 0), sd = c(1, 1), alpha = c(0.5, 0.5), theta = 0)
}
\arguments{
\item{mean}{Numeric vector of component means.}

\item{sd}{Numeric vector of component standard deviations.}

\item{alpha}{Numeric vector of component probabilities.}

\item{theta}{Numeric, distortion parameter.}
}
\description{
Esscher-distorted density and distribution of a Gaussian Mixture
}
\note{
Version 1.0.0.
}
\examples{
library(ggplot2)
grid <- seq(-5, 5, 0.01)
# Density
pdf_1 <- desscherMixture(mean = c(-3, 3), theta = 0)(grid)
pdf_2 <- desscherMixture(mean = c(-3, 3), theta = -0.5)(grid)
pdf_3 <- desscherMixture(mean = c(-3, 3), theta = 0.5)(grid)
ggplot()+
 geom_line(aes(grid, pdf_1), color = "black")+
 geom_line(aes(grid, pdf_2), color = "green")+
 geom_line(aes(grid, pdf_3), color = "red")
# Distribution
cdf_1 <- pesscherMixture(mean = c(-3, 3), theta = 0)(grid)
cdf_2 <- pesscherMixture(mean = c(-3, 3), theta = -0.2)(grid)
cdf_3 <- pesscherMixture(mean = c(-3, 3), theta = 0.2)(grid)
ggplot()+
  geom_line(aes(grid, cdf_1), color = "black")+
  geom_line(aes(grid, cdf_2), color = "green")+
  geom_line(aes(grid, cdf_3), color = "red")

}
\seealso{
Other distributions: 
\code{\link{desscher}()},
\code{\link{dgumbel}()},
\code{\link{dinvgumbel}()},
\code{\link{dkumaraswamy}()},
\code{\link{dmixnorm}()},
\code{\link{dsnorm}()},
\code{\link{dsolarGHI}()},
\code{\link{dsolarK}()},
\code{\link{dsolarX}()},
\code{\link{dsugeno}()},
\code{\link{dtnorm}()}
}
\concept{distributions}
\keyword{distributions}
