% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalSolarFunctions.R
\name{detect_season}
\alias{detect_season}
\title{Detect the season}
\usage{
detect_season(x, invert = FALSE)
}
\arguments{
\item{x}{vector of dates in the format `YYYY-MM-DD`.}

\item{invert}{logica, when `TRUE` the seasons will be inverted.}
}
\value{
a character vector containing the correspondent season. Can be `spring`, `summer`, `autumn`, `winter`.
}
\description{
Detect the season from a vector of dates
}
\examples{
detect_season("2040-01-31")
detect_season(c("2040-01-31", "2023-04-01", "2015-09-02"))

}
