% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{discount_factor}
\alias{discount_factor}
\title{Discount factor function}
\usage{
discount_factor(r = 0.03)
}
\arguments{
\item{r}{level of yearly constant risk-free rate}
}
\description{
Discount factor function
}
