% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{discountFactor}
\alias{discountFactor}
\title{Discount factor function}
\usage{
discountFactor(r = 0.03, discrete = TRUE)
}
\arguments{
\item{r}{level of yearly constant risk-free rate}

\item{discrete}{logical, when `TRUE`, the default, discrete compounding will be used. Otherwise continuous compounding.}
}
\description{
Discount factor function
}
