% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmultmixnorm.R
\name{dmultmixnorm}
\alias{dmultmixnorm}
\alias{pmultmixnorm}
\alias{qmultmixnorm}
\alias{d2mixnorm}
\alias{p2mixnorm}
\alias{q2mixnorm}
\title{Density of a multinomial Gaussian mixture with 2 components each}
\usage{
dmultmixnorm(
  x,
  means = matrix(0, 4, 2),
  Sigma = rep(list(diag(1, 2)), 4),
  probs = rep(1/4, 4),
  log = FALSE
)

pmultmixnorm(
  x,
  means = matrix(0, 4, 2),
  Sigma = rep(list(diag(1, 2)), 4),
  probs = rep(1/4, 4),
  log = FALSE
)

qmultmixnorm(
  p,
  means = matrix(0, 4, 2),
  Sigma = rep(list(diag(1, 2)), 4),
  probs = rep(1/4, 4),
  x_init,
  log.p = FALSE,
  lower.tail = TRUE
)
}
\arguments{
\item{x}{matrix, on the rows the observations, on the columns the locations.}

\item{means}{matrix of means for each state.}

\item{Sigma}{List of variance-covariance matrices for each state.}

\item{probs}{vector of joint probabilities for each state.}

\item{log}{Logical, when \code{TRUE}, probabilities are returned as \code{log(p)}.}

\item{log.p}{Logical, when \code{TRUE}, probabilities p are given as \code{log(p)}.}

\item{lower.tail}{Logical, when \code{TRUE}, the default, the computed probabilities are \eqn{\mathbb{P}(X < x)}, otherwise \eqn{\mathbb{P}(X \ge x)}.}
}
\description{
Density of a multinomial Gaussian mixture with 2 components each

Distribution of a multinomial Gaussian mixture with 2 components each

Quantile of a multinomial Gaussian mixture with 2 components each
}
\details{
Version 1.0.0.
}
\examples{
library(tidyverse)
x <- matrix(runif(200, -3, 3), ncol = 2)

system.time(dmix2norm(x), gcFirst = F)
system.time(map(1:500, ~dmultmixnorm(x)), gcFirst = F)

system.time(map(1:5, ~pmix2norm(x)), gcFirst = F)
system.time(map(1:5, ~pmultmixnorm(x)), gcFirst = F)

qmix2norm(c(0.1, 0.2), x0 = c(-1, 1))
qmultmixnorm(c(0.1, 0.2))

}
\keyword{distributions}
