% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmvsolarGHI.R
\name{dmvsolarGHI}
\alias{dmvsolarGHI}
\alias{pmvsolarGHI}
\alias{qmvsolarGHI}
\title{Multivariate PDF GHI}
\usage{
dmvsolarGHI(x, Ct, pdf_Yt, st)

pmvsolarGHI(x, Ct, cdf_Yt, st)

qmvsolarGHI(p, Ct, cdf_Yt, st, v = NULL, tol = 1e-08, range_s = c(-100, 100))
}
\arguments{
\item{x}{vector or quantiles.}

\item{Ct}{clear sky radiation}

\item{pdf_Yt}{joint density of Y_t.}

\item{st}{list of solarTransform.}
}
\description{
Multivariate PDF GHI

Distribution function for the GHI

Directional quantile
}
\details{
Version 1.0.0.
}
\keyword{distributions}
