% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARCH-forecast.R
\name{e_sigma2_h}
\alias{e_sigma2_h}
\title{Standard GARCH expected value formula}
\usage{
e_sigma2_h(h, omega, alpha, beta, e_x2 = 1, sigma2_t)
}
\description{
Standard GARCH expected value formula
}
\examples{
# Forecast horizon
h <- 2
# GARCH parameters
alpha <- 0.08
beta <- 0.35
omega <- 1*(1 - alpha - beta)
# Moments
e_x2 = 1
e_x4 = 3
# Initial values for variance
sigma2_t <- 1.2
sigma4_t <- sigma2_t^2

e_sigma2_h(10, omega, alpha, beta, e_x2, sigma2_t)
e_sigma2_h_mix(10, omega, alpha, beta, e_x2[1], sigma2_t)
e_sigma4_h_mix(10, omega, alpha, beta, e_x2, e_x4, sigma4_t)
v_sigma2_h_mix(10, omega, alpha, beta, e_x2, e_x4, sigma4_t)
v_sigma_h_mix(10, omega, alpha, beta, e_x2, e_x4, sigma4_t)
e_sigma12_h_mix(10, omega, alpha, beta, e_x2, e_x4, sigma4_t)
e_sigma32_h_mix(10, omega, alpha, beta, e_x2, e_x4, sigma4_t)
}
