% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARCH-forecast.R
\name{e_sigma2_h_mix}
\alias{e_sigma2_h_mix}
\title{Iterative GARCH expected value formula}
\usage{
e_sigma2_h_mix(h, omega, alpha, beta, e_x2 = 1, sigma2_t)
}
\description{
Iterative GARCH expected value formula
}
