% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARCH-forecast.R
\name{e_sigma32_h_mix}
\alias{e_sigma32_h_mix}
\title{Conditional third moment GARCH std. dev (approximated)}
\usage{
e_sigma32_h_mix(h, omega, alpha, beta, e_x2 = 1, e_x4 = 3, sigma4_t)
}
\description{
Conditional third moment GARCH std. dev (approximated)
}
