% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARCH-forecast.R
\name{e_sigma4_h_mix}
\alias{e_sigma4_h_mix}
\title{Iterative GARCH second moment formula}
\usage{
e_sigma4_h_mix(h, omega, alpha, beta, e_x2 = 1, e_x4 = 3, sigma4_t)
}
\description{
Iterative GARCH second moment formula
}
