% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desscher.R
\name{esscher_probability}
\alias{esscher_probability}
\title{Change probability according to esscher parameters}
\usage{
esscher_probability(params = c(0, 0, 1, 1, 0.5), df_n, theta = 0)
}
\description{
Change probability according to esscher parameters
}
