% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnorm_mix.R
\name{fit_dnorm_mix}
\alias{fit_dnorm_mix}
\title{Fit Normal Mixture Pdf}
\usage{
fit_dnorm_mix(x, params, loss = "ml", na.rm = TRUE)
}
\arguments{
\item{x}{vector}

\item{params}{initial parameters}

\item{loss}{loss type. Can be `ml` for maximum likelihood or `kl` for kl_dist.}
}
\description{
Fit the parameter for the density function for a normal mixture with two components.
}
\examples{
params <- c(mu1 = -2, mu2 = 2, sd1 = 3, sd2 = 1, p = 0.5)
n1 <- rnorm(t_bar, mean = params[1], sd = params[3])
n2 <- rnorm(t_bar, mean = params[2], sd = params[4])
Z <- rbinom(t_bar, 1, params[5])
x <- Z * n1 + (1-Z)*n2
fit_dnorm_mix(x, params = init_params, loss = "ml")$par
fit_dnorm_mix(x, params = init_params, loss = "kl")$par

}
