% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnorm_mix.R
\name{fit_dnorm_mix_em}
\alias{fit_dnorm_mix_em}
\title{Fit the Gaussian mixture parameters with EM algorithm.}
\usage{
fit_dnorm_mix_em(
  x,
  params = NULL,
  abstol = 1e-30,
  maxit = 50000,
  match_moments = FALSE,
  na.rm = FALSE
)
}
\arguments{
\item{x}{vector}

\item{params}{initial parameters}

\item{maxit}{maximum number of iterations.}

\item{match_moments}{logical. When `TRUE`, the parameters of the second distribution will be estimated such that
the empirical first two moments of `x` matches the theoretical Gaussian mixture moments.}

\item{na.rm}{logical. When `TRUE`, the default, `NA` values will be excluded from the computations.}

\item{absotol}{absolute level for convergence.}
}
\description{
Fit the parameters for the density function of a Gaussian mixture with two components.
}
\examples{
t_bar <- 1000
params <- c(mu1 = -2, mu2 = 2, sd1 = 3, sd2 = 1, p = 0.5)
n1 <- rnorm(t_bar, mean = params[1], sd = params[3])
n2 <- rnorm(t_bar, mean = params[2], sd = params[4])
Z <- rbinom(t_bar, 1, params[5])
x <- Z*n1 + (1-Z)*n2
fit_dnorm_mix_em(x, params = params)$par
fit_dnorm_mix_em(x, params = params)$par

}
