% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnorm_mix.R
\name{fit_dnorm_mix_monthly}
\alias{fit_dnorm_mix_monthly}
\title{Fit a monthly Gaussian Mixture Pdf}
\usage{
fit_dnorm_mix_monthly(x, date, algo = c("em", "ml"), ...)
}
\arguments{
\item{x}{vector}

\item{date}{vector of dates}

\item{algo}{character, type of algorithm. Can be `em` for Expectation-maximization or `ml` for maximum likelihood.}

\item{loss}{loss type. Can be `ml` for maximum likelihood or `kl` for kl_dist.}

\item{match_moments}{when true the theoric moments will match the empirical ones.}
}
\description{
Fit the monthly parameters for the density function of a Gaussian mixture with two components.
}
