% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalModel.R
\name{formula_fourier}
\alias{formula_fourier}
\title{Create a custom fourier formula}
\usage{
formula_fourier(
  formula,
  reparam = FALSE,
  order = 1,
  period = 365,
  phi = 0,
  sin = TRUE,
  cos = TRUE,
  t_idx = "n"
)
}
\description{
Create a custom fourier formula
}
\keyword{internal}
