% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{from_list_to_parameters.solarModel}
\alias{from_list_to_parameters.solarModel}
\title{Convert a vector of parameter in a structured list}
\usage{
from_list_to_parameters.solarModel(params_list)
}
\description{
Convert a vector of parameter in a structured list
}
