% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sGARCH-internals.R
\name{GARCH_companion_matrix}
\alias{GARCH_companion_matrix}
\title{Construct the companion matrix of an GARCH model}
\usage{
GARCH_companion_matrix(alpha = 0, beta = 0)
}
\arguments{
\item{phi}{numeric vector of length p, AR parameters.}

\item{theta}{numeric vector of length q, MA parameters.}
}
\value{
numeric matrix of size p+q x p+q.
}
\description{
Construct the companion matrix of an GARCH model
}
\note{
Version 1.0.3
}
\examples{
# ARCH(1) / GARCH(1) ~ Base 2 x 2
GARCH_companion_matrix(alpha = c(0.4))
GARCH_companion_matrix(beta = c(0.4))
# Only ARCH + 1 column of zero
GARCH_companion_matrix(alpha = c(0.4, 0.3, 0.1))
# Only GARCH + 1 column of zero
GARCH_companion_matrix(beta = c(0.4, 0.3, 0.1))
# FULL
GARCH_companion_matrix(c(0.4, 0.2), c(0.3, 0))
}
\keyword{GARCH}
