% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARCH.R
\name{GARCH_pq_next_step}
\alias{GARCH_pq_next_step}
\title{Next step function for a GARCH(p,q)}
\usage{
GARCH_pq_next_step(omega = 1, alpha, beta)
}
\arguments{
\item{omega}{The intercept}

\item{alpha}{ARCH parameters}

\item{beta}{GARCH parameters}
}
\description{
Next step function for a GARCH(p,q)
}
\keyword{internal}
