% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianMixture.R
\name{gaussianMixture_monthly}
\alias{gaussianMixture_monthly}
\title{Fit a monthly Gaussian Mixture Pdf (??NOT USED)}
\usage{
gaussianMixture_monthly(x, date, means, sd, p, components = 2, prior_p, ...)
}
\arguments{
\item{x}{vector}

\item{date}{vector of dates}

\item{means}{matrix of initial means with dimension `12 X components`.}

\item{sd}{matrix of initial std. deviations with dimension `12 X components`.}

\item{p}{matrix of initial p with dimension `12 X components`. The rows must sum up to 1.}

\item{prior_p}{matrix of prior probabilities for the each month. Any element that is different from `NA` will be not optimized and will be considered
as given.}

\item{...}{other parameters for the optimization function. See \code{\link{gaussianMixture}} for more details.}
}
\description{
Fit the monthly parameters for the density function of a Gaussian mixture with two components.
}
