% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianMixture.R
\name{GM_fit_rob}
\alias{GM_fit_rob}
\title{Fit GM robust}
\usage{
GM_fit_rob(
  x,
  B = length(x),
  method = c("mclust", "mixtools"),
  components = 2,
  maxit = 30000
)
}
\description{
Fit GM robust
}
