% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianMixture.R
\name{GM_loglik}
\alias{GM_loglik}
\title{Compute the log-likelihood of a Gaussian Mixture}
\usage{
GM_loglik(means, sd, alpha, x)
}
\arguments{
\item{means}{description}

\item{sd}{description}

\item{alpha}{description}
}
\description{
Compute the log-likelihood of a Gaussian Mixture
}
\examples{
GM_loglik(c(-0.8, 0.8), c(0.4,1), c(0.5, 0.5), rnorm(100))
}
