% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianMixture.R
\name{GM_moments}
\alias{GM_moments}
\title{Moments of a gaussian mixture}
\usage{
GM_moments(means, sd, alpha)
}
\description{
Compute the first fourth moments and statistics for a Gaussian Mixture with K components.
}
\examples{
GM_moments(c(-0.3, 0.8), c(0.4,1), c(0.5, 0.5))
GM_moments(c(-0.8, 0.8), c(0.4,1), c(0.5, 0.5))

}
