% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianMixture.R
\name{GM_moments_match}
\alias{GM_moments_match}
\title{Match the first three moments of a Gaussian Mixture}
\usage{
GM_moments_match(d, m1 = 0, m2 = 1, m3 = 0, p = 0.5)
}
\arguments{
\item{d}{Numeric, distance between the two means.}

\item{m1}{Numeric, first target moment.}

\item{m2}{Numeric, second target moment.}

\item{m3}{Numeric, third target moment.}

\item{p}{Numeric, probability.}
}
\description{
Match the first three moments of a Gaussian Mixture
}
