% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarFunctions.R
\name{is_leap_year}
\alias{is_leap_year}
\title{Is leap year?}
\usage{
is_leap_year(day_date)
}
\arguments{
\item{day_date}{dates vector in the format `%YYYY-%MM-%DD`.}
}
\value{
Boolean. `TRUE` if it is a leap year, `FALSE` otherwise.
}
\description{
Check if an year is leap (366 days) or not (365 days).
}
\examples{
is_leap_year("2024-02-01")
is_leap_year(c("2024-10-01", "2025-10-01"))
is_leap_year("2029-02-01")
}
