% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialParameters.R
\name{kernelRegression}
\alias{kernelRegression}
\title{Kernel regression}
\usage{
kernelRegression(formula, data, ...)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{...}{other parameters to be passed to. See \code{np::npreg}.}
}
\description{
Fit a kernel regression.
}
