% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_ts_test.R
\name{ks_ts_test}
\alias{ks_ts_test}
\title{Two sample Kolmogorov Smirnov test for a time series}
\usage{
ks_ts_test(
  x,
  ci = 0.01,
  min_quantile = 0.015,
  max_quantile = 0.985,
  seed = 1,
  plot = FALSE
)
}
\arguments{
\item{x}{a vector.}

\item{ci}{p.value for rejection.}

\item{min_quantile}{minimum quantile for the grid of values.}

\item{max_quantile}{maximum quantile for the grid of values.}

\item{seed}{random seed.}

\item{plot}{when `TRUE` a plot is returned, otherwise a `tibble`.}
}
\value{
when `plot = TRUE` a plot is returned, otherwise a `tibble`.
}
\description{
Two sample Kolmogorov Smirnov test for a time series
}
