% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkumaraswamy.R
\name{kumaraswamy}
\alias{kumaraswamy}
\alias{dkumaraswamy}
\alias{pkumaraswamy}
\alias{qkumaraswamy}
\alias{rkumaraswamy}
\title{Kumaraswamy Random Variable}
\usage{
dkumaraswamy(x, a = 1, b = 1, log.p = FALSE)

pkumaraswamy(x, a = 1, b = 1, log.p = FALSE, lower.tail = TRUE)

qkumaraswamy(p, a = 1, b = 1, log.p = FALSE, lower.tail = TRUE)

rkumaraswamy(n, a = 1, b = 1)
}
\arguments{
\item{x}{vector of quantiles.}

\item{a}{parameter.}

\item{b}{parameter..}

\item{log.p}{logical; if `TRUE`, probabilities p are given as `log(p)`.}

\item{lower.tail}{logical; if TRUE (default), probabilities are `P[X < x]` otherwise, `P[X > x]`.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If `length(n) > 1`, the length is taken to be the number required.}
}
\description{
Probability functions for a Kumaraswamy random variable
}
