% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Location.R
\name{Location}
\alias{Location}
\title{Generate a location}
\usage{
Location(
  place,
  nsim = 50,
  by = "1 month",
  exact_daily_premium = FALSE,
  measures = c("Q", "Qdw", "Qup"),
  control_model = control_solarModel(),
  control_options = control_solarOption(),
  control_esscher = control_solarEsscher(),
  seed = 1
)
}
\description{
Generate a location
}
\keyword{internal}
