% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{logLik.solarModel}
\alias{logLik.solarModel}
\title{Compute the Log-likelihood for a Solar Model}
\usage{
\method{logLik}{solarModel}(model, params)
}
\description{
Compute the Log-likelihood for a Solar Model
}
