% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiationModel_internals.R
\name{loss_dQdP}
\alias{loss_dQdP}
\title{Evaluate the loss for a set of times to maturity}
\usage{
loss_dQdP(lambda, data_dQdP, r = 0, nmonths = 1:12, quiet = FALSE)
}
\description{
Evaluate the loss for a set of times to maturity
}
