% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiationModel_internals.R
\name{loss_dQdP_tau}
\alias{loss_dQdP_tau}
\title{Evaluate the loss for a specific time to maturity}
\usage{
loss_dQdP_tau(lambda, data, r = 0, nmonths = 1:12, quiet = FALSE)
}
\description{
Evaluate the loss for a specific time to maturity
}
