% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmix2norm.R
\name{mix2norm_M_Yl_given_Yv}
\alias{mix2norm_M_Yl_given_Yv}
\title{Conditional means component l given Yv = yv}
\usage{
mix2norm_M_Yl_given_Yv(y_v, M_Yl, M_Yv, S_Yv, Cv_Yl_Yv)
}
\arguments{
\item{y_v}{conditioning value Yv = yv}

\item{M_Yl}{mean vector for component l (state 1, state0)}

\item{M_Yv}{mean vector for component v (state 1, state0)}

\item{S_Yv}{std. deviation vector for component v (state 1, state0)}

\item{Cv_Yl_Yv}{covariance between Yl and Yv.}
}
\description{
Conditional means component l given Yv = yv
}
\keyword{distributions}
