% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmix2norm.R
\name{mix2norm_probs_given_Yv}
\alias{mix2norm_probs_given_Yv}
\title{Conditional probabilities component l given Yv = yv}
\usage{
mix2norm_probs_given_Yv(y_v, M_Yv, S_Yv, probs)
}
\arguments{
\item{y_v}{conditioning value Yv = yv}

\item{M_Yv}{mean vector for component v (state 1, state0)}

\item{S_Yv}{std. deviation vector for component v (state 1, state0)}

\item{probs}{Joint probabilities}
}
\description{
Conditional probabilities component l given Yv = yv
}
\keyword{distributions}
