% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmix2norm.R
\name{mix2norm_S_Yl_given_Yv}
\alias{mix2norm_S_Yl_given_Yv}
\title{Conditional std. deviation component l given Yv = yv}
\usage{
mix2norm_S_Yl_given_Yv(S_Yl, S_Yv, Cv_Yl_Yv)
}
\arguments{
\item{S_Yl}{std. deviation vector for component l (state 1, state0)}

\item{S_Yv}{std. deviation vector for component v (state 1, state0)}

\item{Cv_Yl_Yv}{covariance between Yl and Yv.}
}
\description{
Conditional std. deviation component l given Yv = yv
}
\keyword{distributions}
