% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnorm_mix.R
\name{norm_mix}
\alias{norm_mix}
\alias{dnorm_mix}
\alias{pnorm_mix}
\alias{qnorm_mix}
\alias{rnorm_mix}
\title{Gaussian mixture density}
\usage{
dnorm_mix(params)

pnorm_mix(params)

qnorm_mix(params)

rnorm_mix(n, params)
}
\arguments{
\item{params}{parameters of the two components, (mu1,mu2,sd1,sd2,p)}
}
\value{
a density function.
}
\description{
Probability density function for a Gaussian mixture with two components.
}
\examples{
params <- c(mu1 = -2, mu2 = 2, sd1 = 3, sd2 = 1, p = 0.5)
# Density function
pdf <- dnorm_mix(params)
# Distribution function
cdf <- dnorm_mix(params)
# Quantile function
q <- qnorm_mix(params)
# Random numbers
rnorm_mix(100, params)

}
