% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{normality_test}
\alias{normality_test}
\title{Perform normality tests}
\usage{
normality_test(x = NULL, p_value = 0.05)
}
\arguments{
\item{x}{vector}

\item{p_value}{p.value}
}
\value{
a tibble
}
\description{
Perform normality tests
}
