% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{optimal_n_contracts}
\alias{optimal_n_contracts}
\title{Compute optimal number of contracts}
\usage{
optimal_n_contracts(
  model,
  type = "model",
  premium = "Qr",
  nyear = 2021,
  tick = 0.06,
  efficiency = 0.2,
  n_panels = 2000,
  pun = 0.06
)
}
\description{
Compute optimal number of contracts
}
