% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{optimize.solarModel}
\alias{optimize.solarModel}
\title{Log-likelihood optimazation for Solar Model}
\usage{
optimize.solarModel(model, maxit = 100, quiet = FALSE)
}
\description{
Log-likelihood optimazation for Solar Model
}
