% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption_internals.R
\name{optionPayoff}
\alias{optionPayoff}
\title{Option payoff function}
\usage{
optionPayoff(x, strike = 0, c0 = 0, put = TRUE)
}
\arguments{
\item{x}{numeric, vector of values at maturity.}

\item{strike}{numeric, option strike.}

\item{put}{logical, when `TRUE`, the default, the payoff function is a put othewise a call.}

\item{v0}{numeric, price of the option.}
}
\description{
Compute the payoffs of an option at maturity.
}
\examples{
optionPayoff(10, 9, 1, put = TRUE)
mean(optionPayoff(seq(0, 20), 9, 1, put = TRUE))

}
