% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIT.R
\name{PIT}
\alias{PIT}
\alias{PIT_psolarGHI}
\title{PIT computation}
\usage{
PIT_psolarGHI(Rt, Ct, alpha, beta, M_Y, S_Y, probs, link = "invgumbel")
}
\arguments{
\item{Rt}{vector of solar radiation.}

\item{Ct}{vector of clear-sky radiation.}
}
\description{
PIT computation
}
\keyword{diagnostic}
