% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIT.R
\name{PIT_test}
\alias{PIT_test}
\title{Perform KS and LB test on PIT grades}
\usage{
PIT_test(u, type = "bp", lag = 5)
}
\arguments{
\item{u}{vector, PIT.}

\item{type}{type of autocorrelation test ("bp" or "lb")}

\item{lag}{vector of lags at which perform autocorrelation test.}
}
\description{
Perform KS and LB test on PIT grades
}
\keyword{diagnostic}
