% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_test.R
\name{pks}
\alias{pks}
\alias{qks}
\title{Kolmogorov distribution and quantile functions}
\usage{
pks(x, k = 100)

qks(p, k = 100)
}
\arguments{
\item{x}{vector of quantiles.}

\item{k}{finite value for approximation of infinite sum.}

\item{p}{vector of probabilities.}
}
\value{
A probability, a numeric vector in 0, 1.
}
\description{
Kolmogorov distribution and quantile functions
}
\keyword{internal}
