% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.solarModelSimulation}
\alias{plot.solarModelSimulation}
\title{Plot solar model simulations}
\usage{
\method{plot}{solarModelSimulation}(data, object, nplot = 1, empiric = TRUE)
}
\description{
Plot solar model simulations
}
