% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnorm_mix.R
\name{pnorm_mix}
\alias{pnorm_mix}
\title{Normal Mixture Cdf}
\usage{
pnorm_mix(params)
}
\arguments{
\item{params}{parameters of the two components, (mu1, mu2, sd1, sd2, p)}
}
\value{
a function of x.
}
\description{
Probability distribution function for a normal mixture with two components.
}
\examples{
params <- c(mu1 = -2, mu2 = 2, sd1 = 3, sd2 = 1, p = 0.5)
# Density function
cdf <- pnorm_mix(params)

}
