% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pow_matrix}
\alias{pow_matrix}
\title{Power of a matrix}
\usage{
pow_matrix(x, n = 0)
}
\arguments{
\item{x}{Matrix}

\item{n}{power, if zero will return the identity matrix.}
}
\description{
Compute the power of a matrix
}
