% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalModel.R
\name{predict.seasonalModel}
\alias{predict.seasonalModel}
\title{predict method for seasonalModel}
\usage{
\method{predict}{seasonalModel}(object, n = 1)
}
\description{
predict method for seasonalModel
}
