% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiationModel_internals.R
\name{prepare_dQdP_calibration}
\alias{prepare_dQdP_calibration}
\title{Preprocess the moments for a specific set of time to maturities to avoid recomputing them each time}
\usage{
prepare_dQdP_calibration(tau = c(10, 40, 50), model_Rt)
}
\description{
Preprocess the moments for a specific set of time to maturities to avoid recomputing them each time
}
