% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiationModel_internals.R
\name{prepare_dQdP_calibration_tau}
\alias{prepare_dQdP_calibration_tau}
\title{Preprocess the moments for a specific tau to avoid recomputing them each time}
\usage{
prepare_dQdP_calibration_tau(tau = 10, model_Rt)
}
\description{
Preprocess the moments for a specific tau to avoid recomputing them each time
}
