% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{print.Location}
\alias{print.Location}
\title{Method Print for Location object}
\usage{
\method{print}{Location}(x)
}
\description{
Method Print for Location object
}
\examples{
object <- Location("Roma")
object
}
