% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUN.R
\name{PUN}
\alias{PUN}
\title{PUN}
\usage{
PUN(nyear = NULL, nmonth = NULL, file = "data/df_GME_day.RData")
}
\arguments{
\item{file}{path}

\item{month}{Reference month.}

\item{year}{Reference year.}
}
\value{
numeric, price in euros of a kwh
}
\description{
Function that computes the mean PUN.
}
