% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnorm_mix.R
\name{qnorm_mix}
\alias{qnorm_mix}
\title{Normal Mixture Quantile}
\usage{
qnorm_mix(params)
}
\arguments{
\item{params}{parameters of the two components, (mu1, mu2, sd1, sd2, p)}
}
\value{
a function of p.
}
\description{
Quantile function for a normal mixture with two components.
}
\examples{
params <- c(mu1 = -2, mu2 = 2, sd1 = 3, sd2 = 1, p = 0.5)
# Density function
quantile_func <- qnorm_mix(params)

}
