% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiationModel.R
\name{radiationModel}
\alias{radiationModel}
\title{Radiation model}
\description{
Radiation model

Radiation model
}
\note{
Version 1.0.0
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{theta}}{Numeric, mean reversion parameter.}

\item{\code{lambda_S}}{Numeric, market risk premium Q-measure.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{model}}{model}

\item{\code{measure}}{Character, reference probability measure actually used.}

\item{\code{lambda}}{Numeric, market risk premium actually used.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-radiationModel-new}{\code{radiationModel$new()}}
\item \href{#method-radiationModel-change_measure}{\code{radiationModel$change_measure()}}
\item \href{#method-radiationModel-Ct}{\code{radiationModel$Ct()}}
\item \href{#method-radiationModel-Yt_bar}{\code{radiationModel$Yt_bar()}}
\item \href{#method-radiationModel-Rt_bar}{\code{radiationModel$Rt_bar()}}
\item \href{#method-radiationModel-sigma_bar}{\code{radiationModel$sigma_bar()}}
\item \href{#method-radiationModel-mu_B}{\code{radiationModel$mu_B()}}
\item \href{#method-radiationModel-sigma_B}{\code{radiationModel$sigma_B()}}
\item \href{#method-radiationModel-mu_Y}{\code{radiationModel$mu_Y()}}
\item \href{#method-radiationModel-sigma_Y}{\code{radiationModel$sigma_Y()}}
\item \href{#method-radiationModel-mu_R}{\code{radiationModel$mu_R()}}
\item \href{#method-radiationModel-sigma_R}{\code{radiationModel$sigma_R()}}
\item \href{#method-radiationModel-integral_expectation}{\code{radiationModel$integral_expectation()}}
\item \href{#method-radiationModel-integral_variance}{\code{radiationModel$integral_variance()}}
\item \href{#method-radiationModel-e_mix_drift}{\code{radiationModel$e_mix_drift()}}
\item \href{#method-radiationModel-e_mix_diffusion}{\code{radiationModel$e_mix_diffusion()}}
\item \href{#method-radiationModel-M_Y}{\code{radiationModel$M_Y()}}
\item \href{#method-radiationModel-S_Y}{\code{radiationModel$S_Y()}}
\item \href{#method-radiationModel-pdf_Y}{\code{radiationModel$pdf_Y()}}
\item \href{#method-radiationModel-cdf_Y}{\code{radiationModel$cdf_Y()}}
\item \href{#method-radiationModel-pdf_R}{\code{radiationModel$pdf_R()}}
\item \href{#method-radiationModel-cdf_R}{\code{radiationModel$cdf_R()}}
\item \href{#method-radiationModel-e_GHI}{\code{radiationModel$e_GHI()}}
\item \href{#method-radiationModel-v_GHI}{\code{radiationModel$v_GHI()}}
\item \href{#method-radiationModel-print}{\code{radiationModel$print()}}
\item \href{#method-radiationModel-clone}{\code{radiationModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{radiationModel} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$new(model, correction = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{solarModel model fitted}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-change_measure"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-change_measure}{}}}
\subsection{Method \code{change_measure()}}{
Change the reference probability measure
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$change_measure(measure)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{measure}}{Character, probability measure. Can be \code{P} or \code{Q}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-Ct"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-Ct}{}}}
\subsection{Method \code{Ct()}}{
Clear sky radiation for a day of the year.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$Ct(t_now)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_now}}{Character, today date.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Clear sky radiation on date t_now.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-Yt_bar"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-Yt_bar}{}}}
\subsection{Method \code{Yt_bar()}}{
Seasonal mean for the transformed variable Yt for a given day of the year.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$Yt_bar(t_now)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_now}}{Character, today date.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Seasonal mean for Yt on date t_now.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-Rt_bar"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-Rt_bar}{}}}
\subsection{Method \code{Rt_bar()}}{
Compute the seasonal mean for the solar radiation for a given day of the year.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$Rt_bar(t_now)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_now}}{Character, today date.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Seasonal mean for Rt.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-sigma_bar"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-sigma_bar}{}}}
\subsection{Method \code{sigma_bar()}}{
Instantaneous seasonal variance for the transformed variable for a given day of the year.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$sigma_bar(t_now)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_now}}{Character, today date.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Seasonal std. deviation for Yt on date t_now.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-mu_B"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-mu_B}{}}}
\subsection{Method \code{mu_B()}}{
Return the mixture drift if B is specified, otherwise it return the average drift.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$mu_B(t_now, B = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_now}}{Character, today date.}

\item{\code{B}}{Integer, 1 for the first component, 0 for the second.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Mixture seasonal drift for Yt on date t_now.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-sigma_B"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-sigma_B}{}}}
\subsection{Method \code{sigma_B()}}{
Return the mixture diffusion with seasonal jump.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$sigma_B(t_now, B)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_now}}{Character, today date.}

\item{\code{B}}{Integer, 1 for the first component, 0 for the second.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Mixture seasonal diffusion for Yt.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-mu_Y"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-mu_Y}{}}}
\subsection{Method \code{mu_Y()}}{
Return the drift for the transformed variable Yt.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$mu_Y(Yt, t_now, B = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Yt}}{Numeric, transformed solar radiation.}

\item{\code{t_now}}{Character, today date.}

\item{\code{B}}{Integer, 1 for the first component, 0 for the second.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Mixture drift for Yt.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-sigma_Y"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-sigma_Y}{}}}
\subsection{Method \code{sigma_Y()}}{
Return the diffusion for solar radiation process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$sigma_Y(t_now, B = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_now}}{Character, today date.}

\item{\code{B}}{Integer, 1 for the first component, 0 for the second.}

\item{\code{Rt}}{Numeric, solar radiation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Diffusion for Rt.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-mu_R"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-mu_R}{}}}
\subsection{Method \code{mu_R()}}{
Return the drift for solar radiation process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$mu_R(Rt, t_now, B = 1, dt = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Rt}}{Numeric, solar radiation.}

\item{\code{t_now}}{Character, today date.}

\item{\code{B}}{Integer, 1 for the first component, 0 for the second.}

\item{\code{dt}}{Numeric, time step.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Drift for Rt.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-sigma_R"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-sigma_R}{}}}
\subsection{Method \code{sigma_R()}}{
Return the diffusion for solar radiation process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$sigma_R(Rt, t_now, B = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Rt}}{Numeric, solar radiation.}

\item{\code{t_now}}{Character, today date.}

\item{\code{B}}{Integer, 1 for the first component, 0 for the second.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Diffusion for Rt.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-integral_expectation"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-integral_expectation}{}}}
\subsection{Method \code{integral_expectation()}}{
Compute the integral for expectation for constant mixture parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$integral_expectation(t_now, t_hor, df_date, last_day = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_now}}{Character, today date.}

\item{\code{t_hor}}{Character, horizon date.}

\item{\code{df_date}}{Optional dataframe. See \code{\link{create_monthly_sequence}} for more details.}

\item{\code{last_day}}{Logical. When \code{TRUE} the last day will be treated as conditional variance otherwise not.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-integral_variance"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-integral_variance}{}}}
\subsection{Method \code{integral_variance()}}{
Compute the integral for variance for constant mixture parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$integral_variance(t_now, t_hor, df_date, last_day = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_now}}{Character, today date.}

\item{\code{t_hor}}{Character, horizon date.}

\item{\code{df_date}}{Optional dataframe. See \code{\link{create_monthly_sequence}} for more details.}

\item{\code{last_day}}{Logical. When \code{TRUE} the last day will be treated as conditional variance otherwise not.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-e_mix_drift"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-e_mix_drift}{}}}
\subsection{Method \code{e_mix_drift()}}{
Return the value of the mixture drift of both component of Yt.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$e_mix_drift(t_now, t_hor, df_date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_now}}{Character, today date.}

\item{\code{t_hor}}{Character, horizon date.}

\item{\code{df_date}}{Optional dataframe. See \code{\link{create_monthly_sequence}} for more details.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Mixture expected value for both component of Yt.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-e_mix_diffusion"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-e_mix_diffusion}{}}}
\subsection{Method \code{e_mix_diffusion()}}{
Return the value of the mixture drift of both component of Yt.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$e_mix_diffusion(t_now, t_hor, df_date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_now}}{Character, today date.}

\item{\code{t_hor}}{Character, horizon date.}

\item{\code{df_date}}{Optional dataframe. See \code{\link{create_monthly_sequence}} for more details.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Mixture expected value for both component of Yt.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-M_Y"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-M_Y}{}}}
\subsection{Method \code{M_Y()}}{
Return the conditional expectation for Yn for YN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$M_Y(Rt, t_now, t_hor, df_date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Rt}}{Numeric, solar radiation.}

\item{\code{t_now}}{Character, today date.}

\item{\code{t_hor}}{Character, horizon date.}

\item{\code{df_date}}{Optional dataframe. See \code{\link{create_monthly_sequence}} for more details.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Conditional mean for Yt
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-S_Y"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-S_Y}{}}}
\subsection{Method \code{S_Y()}}{
Return the conditional variance for Yn for YN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$S_Y(t_now, t_hor, df_date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_now}}{Character, today date.}

\item{\code{t_hor}}{Character, horizon date.}

\item{\code{df_date}}{Optional dataframe. See \code{\link{create_monthly_sequence}} for more details.}

\item{\code{Rt}}{Numeric, solar radiation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Conditional variance for Yt
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-pdf_Y"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-pdf_Y}{}}}
\subsection{Method \code{pdf_Y()}}{
Return the conditional density for Y_N given Yn.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$pdf_Y(Rt, t_now, t_hor, B)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Rt}}{Numeric, solar radiation.}

\item{\code{t_now}}{Character, today date.}

\item{\code{t_hor}}{Character, horizon date.}

\item{\code{B}}{Integer, mixture component, if B is missing will be returned the mixture density, otherwise the component density non weighted.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Conditional density for Y_N
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-cdf_Y"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-cdf_Y}{}}}
\subsection{Method \code{cdf_Y()}}{
Return the conditional distribution for Y_N given Yn.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$cdf_Y(Rt, t_now, t_hor, B)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Rt}}{Numeric, solar radiation.}

\item{\code{t_now}}{Character, today date.}

\item{\code{t_hor}}{Character, horizon date.}

\item{\code{B}}{Integer, mixture component, if B is missing will be returned the mixture distribution, otherwise the component distribution non weighted.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Conditional distribution for Y_N
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-pdf_R"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-pdf_R}{}}}
\subsection{Method \code{pdf_R()}}{
Return the conditional density for R_N given Rn.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$pdf_R(Rt, t_now, t_hor, B)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Rt}}{Numeric, solar radiation.}

\item{\code{t_now}}{Character, today date.}

\item{\code{t_hor}}{Character, horizon date.}

\item{\code{B}}{Integer, mixture component, if B is missing will be returned the mixture density, otherwise the component density non weighted.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Conditional density for R_N
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-cdf_R"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-cdf_R}{}}}
\subsection{Method \code{cdf_R()}}{
Return the conditional distribution for R_N given Rn.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$cdf_R(Rt, t_now, t_hor, B)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Rt}}{Numeric, solar radiation.}

\item{\code{t_now}}{Character, today date.}

\item{\code{t_hor}}{Character, horizon date.}

\item{\code{B}}{Integer, mixture component, if B is missing will be returned the mixture distribution, otherwise the component distribution non weighted.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Conditional distribution for R_N
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-e_GHI"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-e_GHI}{}}}
\subsection{Method \code{e_GHI()}}{
Return the conditional expected value for R_N given Rn.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$e_GHI(Rt, t_now, t_hor, B, moment = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Rt}}{Numeric, solar radiation.}

\item{\code{t_now}}{Character, today date.}

\item{\code{t_hor}}{Character, horizon date.}

\item{\code{B}}{Integer, mixture component, if B is missing will be returned the mixture density, otherwise the component density non weighted.}

\item{\code{moment}}{Integer, scalar. Moment order. The default is 1, i.e. the expectation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Conditional moment for solar radiation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-v_GHI"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-v_GHI}{}}}
\subsection{Method \code{v_GHI()}}{
Return the conditional variance value for R_N given Rn.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$v_GHI(Rt, t_now, t_hor, B)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Rt}}{Numeric, solar radiation.}

\item{\code{t_now}}{Character, today date.}

\item{\code{t_hor}}{Character, horizon date.}

\item{\code{B}}{Integer, mixture component, if B is missing will be returned the mixture density, otherwise the component density non weighted.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Conditional variance for R_N
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-print"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-print}{}}}
\subsection{Method \code{print()}}{
Method print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiationModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-radiationModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{radiationModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
