% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riccati_root.R
\name{riccati_root}
\alias{riccati_root}
\title{Riccati Square Root}
\usage{
riccati_root(x)
}
\arguments{
\item{x}{symmetric matrix.}
}
\description{
Square root of a symmetric matrix.
}
\examples{
x <- matrix(c(1, 0.3, 0.3, 1), nrow = 2, byrow = TRUE)
riccati_root(x)
}
