% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalClearsky-R6-internals.R
\name{seasonalClearsky_spec}
\alias{seasonalClearsky_spec}
\title{Specification for \code{seasonalClearsky} model}
\usage{
seasonalClearsky_spec(
  method = "OLS",
  order = 1,
  order_H0 = 1,
  method_H0 = "spencer",
  period = 365,
  include.intercept = TRUE,
  include.trend = FALSE
)
}
\arguments{
\item{method}{Character, method of fitting: can be:  \code{OLS}, \code{RLS}, \code{CLS}}

\item{order}{Integer scalar, number of combinations of sines and cosines.}

\item{order_H0}{Integer scalar, order of the polynomial for extra-terrestrial radiation.}

\item{method_H0}{Character, method of computation of the extra-terrestrial radiation: can be:  \code{spencer}, \code{cooper}.}

\item{period}{Integer scalar, seasonality period. The default is 365.}

\item{include.intercept}{Logical, when \code{TRUE}, the default, the intercept will be included in the clear sky model.}
}
\value{
Named list of control parameters.
}
\description{
Specification for \code{seasonalClearsky} model
}
\note{
Version 1.0.0
}
\examples{
spec_Ct = spec_seasonalClearsky()
}
\keyword{clearsky}
