% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalModel.R
\name{seasonalModel_fit}
\alias{seasonalModel_fit}
\title{Fit a seasonal model}
\usage{
seasonalModel_fit(formula = "Yt ~ 1", order = 1, period = 365, data)
}
\arguments{
\item{formula}{formula, an object of class `formula` (or one that can be coerced to that class).
It is a symbolic description of the model to be fitted and can be used to include or exclude the intercept or external regressors in `data`.}

\item{order}{numeric, of sine and cosine elements.}

\item{period}{numeric, periodicity. The default is `2*base::pi/365`.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.
If not found in data, the variables are taken from environment(formula), typically the environment from which `lm` is called.}
}
\description{
Fit a seasonal model as a linear combination of sine and cosine functions.
}
