% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalModel.R
\name{seasonalModel_predict}
\alias{seasonalModel_predict}
\title{Predict method for the class `seasonalModel`.}
\usage{
seasonalModel_predict(object, n = 1)
}
\arguments{
\item{object}{object of the class `seasonalModel`.}

\item{n}{integer, number of day of the year.}
}
\description{
Predict method for the class `seasonalModel`.
}
