% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalRadiation.R
\name{seasonalRadiation}
\alias{seasonalRadiation}
\title{Seasonal model for solar radiation radiation}
\usage{
seasonalRadiation(spec)
}
\arguments{
\item{spec}{an object with class `solarModelSpec`. See the function \code{\link{solarModel_spec}} for details.}
}
\description{
Fit a seasonal model for solar radiation
}
\examples{
library(ggplot2)
# Seasonal model for GHI
spec <- solarModel_spec("Oslo", target = "GHI")
model <- seasonalRadiation(spec)
spec$data$GHI_bar <- model$predict(spec$data$n)
ggplot(spec$data)+
 geom_line(aes(n, GHI))+
 geom_line(aes(n, GHI_bar), color = "blue")

# Seasonal model for clear sky
spec <- solarModel_spec("Oslo", target = "clearsky")
model <- seasonalRadiation(spec)
spec$data$Ct_bar <- model$predict(spec$data$n)
ggplot(spec$data)+
 geom_line(aes(n, clearsky))+
 geom_line(aes(n, Ct_bar), color = "blue")
}
