% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalSolarFunctions.R
\name{seasonalSolarFunctions}
\alias{seasonalSolarFunctions}
\title{Solar seasonal functions}
\description{
Solar seasonal functions

Solar seasonal functions
}
\examples{
sf <- seasonalSolarFunctions$new()
sf$angle_minmax("2022-01-01", 44)
sf$H0(1:365, 44)
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{G0}}{solar constant, i,e, `1367`.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{seasonalSolarFunctions$new()}}
\item \href{#method-method}{\code{seasonalSolarFunctions$method()}}
\item \href{#method-B}{\code{seasonalSolarFunctions$B()}}
\item \href{#method-degree}{\code{seasonalSolarFunctions$degree()}}
\item \href{#method-radiant}{\code{seasonalSolarFunctions$radiant()}}
\item \href{#method-time_adjustment}{\code{seasonalSolarFunctions$time_adjustment()}}
\item \href{#method-G0n}{\code{seasonalSolarFunctions$G0n()}}
\item \href{#method-declination}{\code{seasonalSolarFunctions$declination()}}
\item \href{#method-solar_angle}{\code{seasonalSolarFunctions$solar_angle()}}
\item \href{#method-solar_altitude}{\code{seasonalSolarFunctions$solar_altitude()}}
\item \href{#method-sun_hours}{\code{seasonalSolarFunctions$sun_hours()}}
\item \href{#method-angle_minmax}{\code{seasonalSolarFunctions$angle_minmax()}}
\item \href{#method-cosZ}{\code{seasonalSolarFunctions$cosZ()}}
\item \href{#method-H0}{\code{seasonalSolarFunctions$H0()}}
\item \href{#method-solar_hour}{\code{seasonalSolarFunctions$solar_hour()}}
\item \href{#method-omega}{\code{seasonalSolarFunctions$omega()}}
\item \href{#method-clearsky}{\code{seasonalSolarFunctions$clearsky()}}
\item \href{#method-clone}{\code{seasonalSolarFunctions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a `seasonalSolarFunctions` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$new(method = "cooper")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{character, method type for computations. Can be `cooper` or `spencer`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-method"></a>}}
\if{latex}{\out{\hypertarget{method-method}{}}}
\subsection{Method \code{method()}}{
Extract or update the method used for computations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$method(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{character, method type. Can be `cooper` or `spencer`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
When `x` is missing it return a character containing the method that
is actually used.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-B"></a>}}
\if{latex}{\out{\hypertarget{method-B}{}}}
\subsection{Method \code{B()}}{
Seasonal adjustment parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$B(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes
\deqn{B(n) = \frac{2\pi}{365} n}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-degree"></a>}}
\if{latex}{\out{\hypertarget{method-degree}{}}}
\subsection{Method \code{degree()}}{
Convert angles in radiant into an angles in degrees.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$degree(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{numeric vector, angles in radiant.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes:
\deqn{\frac{x 180}{\pi}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiant"></a>}}
\if{latex}{\out{\hypertarget{method-radiant}{}}}
\subsection{Method \code{radiant()}}{
Convert angles in degrees into an angles in radiant
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$radiant(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{numeric vector, angles in degrees.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes:
\deqn{\frac{x \pi}{180}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-time_adjustment"></a>}}
\if{latex}{\out{\hypertarget{method-time_adjustment}{}}}
\subsection{Method \code{time_adjustment()}}{
Compute solar time adjustment in seconds
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$time_adjustment(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes
\deqn{229.2(0.000075 + 0.001868 \cos(B) - 0.032077\sin(B) - 0.014615\cos(2B) - 0.04089\sin(2B))}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-G0n"></a>}}
\if{latex}{\out{\hypertarget{method-G0n}{}}}
\subsection{Method \code{G0n()}}{
Compute solar constant
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$G0n(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If the selected method is `cooper`, the function computes:
\deqn{G_{0,n} = G_0 (1 + 0.033\cos(B))}
otherwise when it is `spencer` it computes:
\deqn{G_{0,n} = G_0 (1.000110 + 0.034221\cos(B) + 0.001280\sin(B) + 0.000719\cos(2B) + 0.000077\sin(2B))}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-declination"></a>}}
\if{latex}{\out{\hypertarget{method-declination}{}}}
\subsection{Method \code{declination()}}{
Compute solar declination
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$declination(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If the selected method was `cooper`, the function computes:
\deqn{\delta(n) = 23.45 \sin \left(\frac{2 \pi (284 + n)}{365}\right)}
otherwise when it is `spencer` it computes:
\deqn{\delta(n) = \frac{180}{\pi}(0.006918 - 0.399912\cos(B) + 0.070257\sin(B) - 0.006758\cos(2B))}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solar_angle"></a>}}
\if{latex}{\out{\hypertarget{method-solar_angle}{}}}
\subsection{Method \code{solar_angle()}}{
Compute solar angle at sunset in degrees
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$solar_angle(n, lat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}

\item{\code{lat}}{latitude in degrees.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes
\deqn{\cos^{-1}(-\tan(\delta(n))\tan(\phi))}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solar_altitude"></a>}}
\if{latex}{\out{\hypertarget{method-solar_altitude}{}}}
\subsection{Method \code{solar_altitude()}}{
Compute solar altitude in degrees
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$solar_altitude(n, lat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}

\item{\code{lat}}{latitude in degrees.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes
\deqn{\sin^{-1}(-\sin(\delta(n))\sin(\phi) + \cos(\delta(n))\cos(\phi))}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sun_hours"></a>}}
\if{latex}{\out{\hypertarget{method-sun_hours}{}}}
\subsection{Method \code{sun_hours()}}{
Compute number of sun hours
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$sun_hours(n, lat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}

\item{\code{lat}}{latitude in degrees.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-angle_minmax"></a>}}
\if{latex}{\out{\hypertarget{method-angle_minmax}{}}}
\subsection{Method \code{angle_minmax()}}{
Compute the solar angle for a latitude in different dates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$angle_minmax(n, lat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}

\item{\code{lat}}{latitude in degrees.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cosZ"></a>}}
\if{latex}{\out{\hypertarget{method-cosZ}{}}}
\subsection{Method \code{cosZ()}}{
Compute the incidence angle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$cosZ(n, lat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}

\item{\code{lat}}{latitude in degrees.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H0"></a>}}
\if{latex}{\out{\hypertarget{method-H0}{}}}
\subsection{Method \code{H0()}}{
Compute the solar extraterrestrial radiation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$H0(n, lat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}

\item{\code{lat}}{latitude in degrees.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solar_hour"></a>}}
\if{latex}{\out{\hypertarget{method-solar_hour}{}}}
\subsection{Method \code{solar_hour()}}{
Compute the solar hour
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$solar_hour(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{datehour}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-omega"></a>}}
\if{latex}{\out{\hypertarget{method-omega}{}}}
\subsection{Method \code{omega()}}{
Compute the solar angle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$omega(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{datehour}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clearsky"></a>}}
\if{latex}{\out{\hypertarget{method-clearsky}{}}}
\subsection{Method \code{clearsky()}}{
Hottel clearsky
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$clearsky(
  cosZ = NULL,
  G0 = NULL,
  altitude = 2.5,
  clime = "No Correction"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cosZ}}{solar incidence angle}

\item{\code{G0}}{solar constant}

\item{\code{altitude}}{altitude in km}

\item{\code{clime}}{clime correction}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
