% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{simulate.solarModel}
\alias{simulate.solarModel}
\alias{scenario.solarModel}
\title{Simulate scenarios of solar model}
\usage{
\method{simulate}{solarModel}(
  object,
  from = "2010-01-01",
  to = "2010-12-31",
  nsim = 1,
  lambda = 0,
  vol = NA,
  rf = FALSE,
  seed = 1,
  quiet = FALSE
)

scenario.solarModel(
  object,
  from = "2010-01-01",
  to = "2010-12-31",
  by = "1 month",
  nsim = 1,
  lambda = 0,
  vol = NA,
  rf = FALSE,
  seed = 1,
  quiet = FALSE
)
}
\arguments{
\item{from}{scalar date, starting date for simulations.}

\item{to}{scalar date, end date for simulations.}

\item{nsim}{scalar integer, number of simulations.}

\item{lambda}{scalar numeric, Esscher parameter. When `rf = FALSE`, the input parameter `lambda` will be transformed in negative.}

\item{vol}{scalar numeric, unconditional mean of GARCH(1,1) standard deviation. If `NA` will be used the estimated one.}

\item{rf}{logical. When `TRUE` the AR(2) component will be set to zero.}

\item{seed}{scalar integer, starting random seed.}

\item{quiet}{logical}
}
\description{
Simulate scenarios of solar model
}
