% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsn.R
\name{snorm}
\alias{snorm}
\alias{dsnorm}
\alias{psnorm}
\alias{qsnorm}
\alias{rsnorm}
\title{Skewed Normal}
\usage{
dsnorm(x, mean = 0, sd = 1, skew = 0, log = FALSE)

psnorm(x, mean = 0, sd = 1, skew = 0, log.p = FALSE, lower.tail = TRUE)

qsnorm(p, mean = 0, sd = 1, skew = 0, log.p = FALSE, lower.tail = TRUE)

rsnorm(n, mean = 0, sd = 1, skew = 0)
}
\arguments{
\item{x}{vector of quantiles.}

\item{mean}{vector of means.}

\item{sd}{vector of standard deviations.}

\item{skew}{vector of skewness.}

\item{log}{logical; if `TRUE`, probabilities are returned as `log(p)`.}

\item{log.p}{logical; if `TRUE`, probabilities p are given as `log(p)`.}

\item{lower.tail}{logical; if TRUE (default), probabilities are `P[X < x]` otherwise, `P[X > x]`.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If `length(n) > 1`, the length is taken to be the number required.}
}
\description{
Probability for a skewed normal random variable.
}
\examples{
x <- seq(-5, 5, 0.01)
# Density function
p <- dsnorm(x, mean = 0, sd = 1)
plot(x, p, type = "l")
# Distribution function
p <- psnorm(x, mean = 0, sd = 1)
plot(x, p, type = "l")
# Quantile function
dsnorm(0.1)
psnorm(qsnorm(0.1))
# Random numbers
rsnorm(1000)
plot(rsnorm(1000), type = "l")

}
