% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarFunctions.R
\name{solar_angle_minmax}
\alias{solar_angle_minmax}
\title{solar_angle_minmax}
\usage{
solar_angle_minmax(
  lat = NULL,
  day_date = Sys.Date(),
  day_end = NULL,
  method = "cooper"
)
}
\arguments{
\item{lat}{integer, latitude.}

\item{day_date}{vector of dates in the format `%YYYY-%MM-%DD`}

\item{day_end}{end date, if it is not NULL will be end date.}

\item{method}{method used for computation of solar declination, can be `cooper` or `spencer`.}
}
\value{
a tibble.
}
\description{
Compute the solar angle for a latitude in different dates.
}
\examples{

solar_angle_minmax(55.3, "2040-01-01", day_end = "2040-12-31")

solar_angle_minmax(55.3, c("2040-01-31", "2023-04-01", "2015-09-02"))

}
