% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarFunctions.R
\name{solar_clearsky_hourly}
\alias{solar_clearsky_hourly}
\title{solar_clearsky_hourly}
\usage{
solar_clearsky_hourly(
  cosZ = NULL,
  G0 = NULL,
  altitude = 2.5,
  clime = c("No Correction", "Summer", "Winter", "Subartic Summer", "Tropical")
)
}
\value{
a numeric vector containing the time adjustment in minutes.
}
\description{
Compute the clearsky radiation for hourly data.
}
\examples{

# detect the season in 2040-01-31
solar_clearsky(cosZ = 1, altitude = 2.5, clime = c("No Correction", "Summer", "Winter", "Subartic Summer", "Tropical"))

}
