% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalModel.R
\name{solar_clearsky_optimizer}
\alias{solar_clearsky_optimizer}
\title{Optimizer for Solar Clear sky}
\usage{
solar_clearsky_optimizer(GHI, G0, control = control.seasonalModel())
}
\arguments{
\item{GHI}{vector of solar radiation}

\item{G0}{vector of extraterrestrial radiation}
}
\value{
a numeric vector containing the fitted clear sky radiation.
}
\description{
Find the best parameter delta for fitting clear sky radiation.
}
