% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarFunctions.R
\name{solar_extraterrestrial_radiation}
\alias{solar_extraterrestrial_radiation}
\title{Solar extraterrestrial radiation}
\usage{
solar_extraterrestrial_radiation(
  lat = NULL,
  day_date = Sys.Date(),
  day_end = NULL,
  method = "spencer"
)
}
\arguments{
\item{lat}{latitude}

\item{day_date}{vector of dates in the format `%YYYY-%MM-%DD`}

\item{day_end}{end date, if it is not NULL will be end date.}

\item{method}{method used for computation of solar declination, can be `cooper` or `spencer`.}
}
\value{
a numeric vector containing the time adjustment in minutes.
}
\description{
Compute the solar angle for a latitude in different times of the day.
}
\examples{
solar_extraterrestrial_radiation(42.23, "2022-05-01", day_end = "2022-05-31")
}
