% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarFunctions.R
\name{solar_extraterrestrial_radiation_hourly}
\alias{solar_extraterrestrial_radiation_hourly}
\title{solar_extraterrestrial_radiation_hourly}
\usage{
solar_extraterrestrial_radiation_hourly(
  lat = NULL,
  lon = NULL,
  day_date_time = NULL,
  day_time_end = NULL,
  altitude = 2.5,
  clime = "No Correction"
)
}
\arguments{
\item{lat}{latitude}

\item{day_date}{vector of dates in the format `%YYYY-%MM-%DD`}

\item{day_end}{end date, if it is not NULL will be end date.}

\item{method}{method used for computation of solar declination, can be `cooper` or `spencer`.}
}
\value{
a numeric vector containing the time adjustment in minutes.
}
\description{
Compute the extraterrestrial hourly total radiation on an horizontal surface. Note that hottel clear sky max model is included in computation.
}
\examples{

# detect the season in 2040-01-31
solar_extraterrestrial_radiation_hourly(44.23, 11.20, day_date_time = "2040-01-01 00:00:00", day_time_end = "2040-01-03 00:00:00")

}
