% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{solar_monthly_mixture}
\alias{solar_monthly_mixture}
\title{Monthly Gaussian mixture with two components}
\usage{
solar_monthly_mixture(x, date, weights, match_moments = FALSE, prior_p, ...)
}
\arguments{
\item{x}{arg}

\item{date}{arg}

\item{weights}{arg}

\item{match_moments}{arg}

\item{...}{arg}
}
\description{
Monthly Gaussian mixture with two components
}
