% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_option.R
\name{solar_option_esscher_calibrator}
\alias{solar_option_esscher_calibrator}
\title{Calibrate Esscher Bounds and parameters}
\usage{
solar_option_esscher_calibrator(
  model,
  sim,
  control_options = control_solarOption(),
  control = control_solarEsscher()
)
}
\arguments{
\item{model}{an object of the class `solarModel`.}

\item{sim}{simulations object.}

\item{control_options}{control function, see `control_solarOption`.}

\item{control}{control function, see `control_solarEsscher`.}
}
\description{
Calibrate Esscher Bounds and parameters
}
