% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_option.R
\name{solar_option_payoff_model}
\alias{solar_option_payoff_model}
\title{Pricing function for a solar model (for all the year)}
\usage{
solar_option_payoff_model(
  model,
  lambda = 0,
  vol = NA,
  nmonths = 1:12,
  control = control_solarOption()
)
}
\arguments{
\item{model}{an object of the class `solarModel`.}

\item{lambda}{Esscher parameter}

\item{vol}{unconditional GARCH variance, when `NA` will be used the fitted one,}

\item{nmonths}{index for the months.}

\item{control}{control function, see `control_solarOption`.}
}
\description{
Pricing function for a solar model (for all the year)
}
\examples{
lambda = 0
vol = 1
nmonth = 3
nday = 1
}
