% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_option.R
\name{solar_option_payoff_scenarios}
\alias{solar_option_payoff_scenarios}
\title{Payoff on Simulated Data}
\usage{
solar_option_payoff_scenarios(
  sim,
  nmonth = 1:12,
  nsim = NULL,
  control = control.solarOption()
)
}
\description{
Payoff on Simulated Data
}
\examples{
sim = Location$model$scenarios$P
nmonth = 1:12
nsim = NULL
control = control.solarOption()
}
