% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_option.R
\name{solar_option_payoff_structure}
\alias{solar_option_payoff_structure}
\title{Structure payoffs}
\usage{
solar_option_payoff_structure(model, type = "sim", exact_daily_premium = TRUE)
}
\arguments{
\item{model}{an object of the class `solarModel`.}

\item{type}{can be `sim` or `model`.}

\item{exact_daily_premium}{when `TRUE` the historical premium is computed as daily average among all the years.
Otherwise the monthly premium is computed and then divided by the number of days of the month.}
}
\description{
Structure payoffs
}
