% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarFunctions.R
\name{solar_time_adjustment}
\alias{solar_time_adjustment}
\title{Solar time adjustment}
\usage{
solar_time_adjustment(day_date = NULL, day_end = NULL)
}
\arguments{
\item{day_date}{vector of dates in the format `%YYYY-%MM-%DD`}

\item{day_end}{end date, if it is not NULL will be end date.}
}
\value{
a numeric vector containing the time adjustment in seconds.
}
\description{
Compute the time adjustment for a date.
}
\examples{
solar_time_adjustment("2040-01-31")
solar_time_adjustment(c("2040-01-31", "2023-04-01", "2015-09-02"))
}
