% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarFunctions.R
\name{solar_time_constant}
\alias{solar_time_constant}
\title{Solar time constant}
\usage{
solar_time_constant(day_date = NULL, day_end = NULL, method = "spencer")
}
\arguments{
\item{day_date}{vector of dates in the format `YYYY-MM-DD`.}

\item{day_end}{end date, if it is not `NULL` will be end date.}

\item{method}{method used for computation, can be `cooper` or `spencer`.}
}
\value{
a numeric vector containing the solar constant.
}
\description{
Compute the solar constant for a date.
}
\examples{
solar_time_constant("2040-01-31")
solar_time_constant(c("2040-01-31", "2023-04-01", "2015-09-02"))
}
