% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarFunctions.R
\name{solar_time_declination}
\alias{solar_time_declination}
\title{solar_time_declination}
\usage{
solar_time_declination(
  day_date = NULL,
  day_end = NULL,
  method = c("cooper", "spencer")
)
}
\arguments{
\item{day_date}{vector of dates in the format `%YYYY-%MM-%DD`}

\item{day_end}{end date, if it is not NULL will be end date.}

\item{method}{method used for computation, can be `cooper` or `spencer`.}
}
\value{
a numeric vector containing the solar declination in minutes.
}
\description{
Compute the solar declination for different dates.
}
\examples{

# detect the season in 2040-01-31
solar_time_declination("2040-01-01", day_end = "2040-12-31")

# detect the season in a vector of dates
solar_time_declination(c("2040-01-31", "2023-04-01", "2015-09-02"))
}
