% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarEsscher.R
\name{solarEsscher_bounds}
\alias{solarEsscher_bounds}
\title{Calibrate Esscher Bounds and parameters}
\usage{
solarEsscher_bounds(
  model,
  control_options = control_solarOption(),
  control_esscher = control_solarEsscher()
)
}
\arguments{
\item{model}{object with the class `solarModel`. See the function \code{\link{solarModel}} for details.}

\item{control_options}{control function. See \code{\link{control_solarOption}} for details.}

\item{control_esscher}{control function. See \code{\link{control_solarEsscher}} for details.}
}
\description{
Calibrate Esscher Bounds and parameters
}
