% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarEsscher.R
\name{solarEsscher_calibrator_month}
\alias{solarEsscher_calibrator_month}
\title{Calibrate monthly Esscher parameter given the expected return}
\usage{
solarEsscher_calibrator_month(
  model,
  nmonth = 1,
  expected_return = 0,
  target_price = NA,
  control_esscher = control_solarEsscher(),
  control_options = control_solarOption()
)
}
\arguments{
\item{model}{solar model}

\item{nmonth}{month}

\item{expected_return}{expected return at maturity. The benchmark for the `target_price` to match will be the
mean cumulated net payoff on the last day of the month plus the model price paid under the Esscher measure.
The return of the `target_price` with respect to the model price will match the parameter `expected_return`. For example `0.01`, `0.02`, ecc.}

\item{target_price}{alternative to the `expected_return` parameter. Submitting a `target_price` will imply that the `expected_return = 0` so that
the model price under the Esscher measure matches the `target_price`}

\item{control_esscher}{control}

\item{control_options}{control}
}
\description{
Calibrator function for the monthly Esscher parameter of a solarOption given a
desired level of expected return at maturity.
}
