% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarMixture.R
\name{solarMixture}
\alias{solarMixture}
\title{Monthly Gaussian mixture with two components}
\description{
Monthly Gaussian mixture with two components

Monthly Gaussian mixture with two components
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{maxit}}{Integer, maximum number of iterations.}

\item{\code{abstol}}{Numeric, absolute level for convergence.}

\item{\code{components}}{Integer, number of components.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{A tibble with the following columns:
#' \describe{
\item{date}{Time series of dates.}
\item{Month}{Vector of Month.}
\item{x}{Time series used for fitting.}
\item{w}{Time series of weights.}
}}

\item{\code{means}}{Matrix of means where a row represents a month and a column a mixture component.}

\item{\code{sd}}{Matrix of std. deviations where a row represents a month and a column a mixture component.}

\item{\code{p}}{Matrix of probabilities where a row represents a month and a column a mixture component.}

\item{\code{model}}{Named List with 12 \code{\link{gaussianMixture}} objects.}

\item{\code{loglik}}{Numeric, total log-likelihood.}

\item{\code{fitted}}{A \code{tibble} with the classified series}

\item{\code{moments}}{A \code{tibble} with the theoric moments. It contains:
#' \describe{
\item{Month}{Month of the year.}
\item{mean}{Theoric monthly expected value of the mixture model.}
\item{variance}{Theoric monthly variance of the mixture model.}
\item{skewness}{Theoric monthly skewness.}
\item{kurtosis}{Theoric monthly kurtosis.}
\item{nobs}{Number of observations used for fitting.}
\item{loglik}{Monthly log-likelihood.}
}}

\item{\code{parameters}}{A \code{tibble} with the fitted parameters.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-solarMixture-new}{\code{solarMixture$new()}}
\item \href{#method-solarMixture-fit}{\code{solarMixture$fit()}}
\item \href{#method-solarMixture-update}{\code{solarMixture$update()}}
\item \href{#method-solarMixture-clone}{\code{solarMixture$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMixture-new"></a>}}
\if{latex}{\out{\hypertarget{method-solarMixture-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{solarMixture} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMixture$new(components = 2, maxit = 100, abstol = 1e-14)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{components}}{Integer, number of components.}

\item{\code{maxit}}{Integer, maximum number of iterations.}

\item{\code{abstol}}{Numeric, absolute level for convergence.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMixture-fit"></a>}}
\if{latex}{\out{\hypertarget{method-solarMixture-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the parameters with mclust package
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMixture$fit(x, date, weights)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector}

\item{\code{date}}{date vector}

\item{\code{weights}}{observations weights, if a weight is equal to zero the observation is excluded, otherwise is included with unitary weight.
When \code{missing} all the available observations will be used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMixture-update"></a>}}
\if{latex}{\out{\hypertarget{method-solarMixture-update}{}}}
\subsection{Method \code{update()}}{
Update means, sd, p and recompute log-likelihood and fitted data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMixture$update(means, sd, p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{means}}{Numeric matrix of means parameters.}

\item{\code{sd}}{Numeric matrix of std. deviation parameters.}

\item{\code{p}}{Numeric matrix of probability parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMixture-clone"></a>}}
\if{latex}{\out{\hypertarget{method-solarMixture-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMixture$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
