% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarMixture.R
\name{solarMixture_moments_match}
\alias{solarMixture_moments_match}
\title{Correct the moments to ensure moments matching}
\usage{
solarMixture_moments_match(
  coefficients,
  e_target,
  v_target,
  sk_target,
  kt_target,
  x
)
}
\description{
Correct the moments to ensure moments matching
}
\keyword{internal}
\keyword{solarMixture}
