% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{solarModel.monthly_mixture}
\alias{solarModel.monthly_mixture}
\title{Solar Normal Mixture Model}
\usage{
solarModel.monthly_mixture(data, loss = "ml", match_moments = FALSE)
}
\arguments{
\item{data}{dataset with at least a column with `Month` and the target variable names `ut`.}

\item{loss}{character, type of loss function. Default is `ml` for maximum likelihood or can be `kl` for KL distance.}

\item{match_moments}{logical.}
}
\description{
Solar Normal Mixture Model
}
