% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{solarModel_calibrator}
\alias{solarModel_calibrator}
\title{Calibrator for solar Options}
\usage{
solarModel_calibrator(
  model,
  nmonths = 1:12,
  control_options = control_solarOption()
)
}
\description{
Calibrator for solar Options
}
