% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{solarModel_control}
\alias{solarModel_control}
\title{Control parameters for a `solarModel` object}
\usage{
solarModel_control(
  clearsky.model = clearskyModel_control(),
  mean.model = list(seasonalOrder = 1, arOrder = 2, include.intercept = FALSE),
  variance.model = list(seasonalOrder = 1, match_moments = FALSE),
  threshold = 0.001,
  algo = "em",
  quiet = FALSE
)
}
\arguments{
\item{clearsky.model}{list with control parameters, see `clearskyModel_control()`.}

\item{mean.model}{a list of parameters.}

\item{variance.model}{a list of parameters.}

\item{threshold}{Threshold for the estimation of alpha and beta.}

\item{quiet}{logical, when `TRUE` the function will not display any message.}
}
\description{
Control parameters for a `solarModel` object
}
