% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_internals.R
\name{solarModel_data}
\alias{solarModel_data}
\title{Extract a dataset from a solarModel object}
\usage{
solarModel_data(model, monthly = TRUE, GM = TRUE)
}
\description{
Extract a dataset from a solarModel object
}
\examples{
model <- Bologna
solarModel_data(model)
# Do not add monthly data
solarModel_data(model, monthly = FALSE)
# Do not add seasonal data
solarModel_data(model, monthly = FALSE, GM = FALSE)
}
