% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{solarModel_empiric_GM}
\alias{solarModel_empiric_GM}
\title{Empiric Gaussian Mixture parameters}
\usage{
solarModel_empiric_GM(model, match_moments = FALSE)
}
\description{
Empiric Gaussian Mixture parameters
}
