% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{solarModel_filter}
\alias{solarModel_filter}
\title{Update the time series inside a `solarModel` object}
\usage{
solarModel_filter(model)
}
\arguments{
\item{model}{`solarModel` object}
}
\description{
Update the time series inside a `solarModel` object
}
\examples{
model <- Bologna
params <- solarModel_parameters(model)
model <- solarModel_update(model, params)
model <- solarModel_filter(model)
}
