% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{solarModel_fit}
\alias{solarModel_fit}
\title{Fit a model for solar radiation}
\usage{
solarModel_fit(object, ...)
}
\arguments{
\item{object}{object with class `solarModel`. See the function `solarModel_spec()`. For example  `solarModel_spec("Bologna")`.}

\item{...}{additional parameters for the function `fit_dnorm_mix_em()`.}
}
\description{
Fit a model for solar radiation
}
