% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_internals.R
\name{solarModel_forecaster}
\alias{solarModel_forecaster}
\title{Produce a forecast from a solarModel object}
\usage{
solarModel_forecaster(
  model,
  date = "2020-01-01",
  ci = 0.1,
  unconditional = FALSE
)
}
\description{
Produce a forecast from a solarModel object
}
\examples{
model <- Bologna
solarModel_forecaster(model, date = "2020-04-01")
object <- solarModel_forecaster(model, date = "2020-04-01", unconditional = TRUE)
object
}
