% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-selection.R
\name{solarModel_grid}
\alias{solarModel_grid}
\title{Fit a grid of solarModels}
\usage{
solarModel_grid(
  spec,
  arOrder = 2,
  maOrder = 2,
  archOrder = 1,
  garchOrder = 1,
  QMLE = FALSE
)
}
\arguments{
\item{spec}{A \code{solarModel_spec} object.}

\item{arOrder}{Numeric. Maximum AR order.}

\item{maOrder}{Numeric. Maximum MA order.}

\item{archOrder}{Numeric. Maximum ARCH order.}

\item{garchOrder}{Numeric. Maximum GARCH order.}

\item{QMLE}{Logical. If \code{TRUE}, each fitted model is refitted with
\code{solarModel_QMLE()}.}
}
\value{
A tibble containing model names, selected orders, cloned
specifications, and fitted \code{solarModel} objects.
}
\description{
Fit a grid of solarModels
}
\note{
Version 1.0.0.
}
\examples{
if (interactive()) {
  spec <- solarModel_spec$new()
  spec$specification("Bologna")
  models <- solarModel_grid(spec, 1, 1, 1, 1)
  models
}

}
\keyword{solarModel}
