% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_internals.R
\name{solarModel_loglik_calibrator}
\alias{solarModel_loglik_calibrator}
\title{Calibrate the parameters}
\usage{
solarModel_loglik_calibrator(
  model,
  abstol = 1e-04,
  reltol = 1e-04,
  quiet = FALSE
)
}
\description{
Calibrate the parameters
}
