% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_internals.R
\name{solarModel_match_params}
\alias{solarModel_match_params}
\title{Match solarModel parameters in vector form}
\usage{
solarModel_match_params(vec_params, params)
}
\description{
Match solarModel parameters in vector form
}
\examples{
model = solarModel$new(spec)
model$fit()
vec_params <- c(theta = 1, alpha1 = 10)
solarModel_match_params(vec_params, model$coefficients)
}
\keyword{internal}
\keyword{solarModel}
