% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_internals.R
\name{solarModel_mixture}
\alias{solarModel_mixture}
\alias{solarModel_mvmixture}
\title{Monthly Gaussian mixture with two components}
\usage{
solarModel_mixture(
  x,
  date,
  weights,
  match_moments = FALSE,
  maxit = 100,
  abstol = 1e-14
)

solarModel_mvmixture(model_Ct, model_GHI)
}
\arguments{
\item{x}{arg}

\item{date}{arg}

\item{weights}{observations weights, if a weight is equal to zero the observation is excluded, otherwise is included with unitary weight.
When `missing` all the available observations will be used.}

\item{match_moments}{arg}

\item{maxit}{maximum number of iterations.}

\item{model_Ct}{arg}

\item{model_GHI}{arg}
}
\description{
Monthly Gaussian mixture with two components

Monthly multivariate Gaussian mixture with two components
}
