% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_internals.R
\name{solarModel_moments}
\alias{solarModel_moments}
\title{Moments from a solarModel object}
\usage{
solarModel_moments(
  t_now,
  t_hor,
  data,
  ARMA,
  GARCH,
  NM_model,
  transform,
  quiet = FALSE
)
}
\description{
Moments from a solarModel object
}
\examples{
model = solarModel$new(spec)
model$fit()
t_now = "2022-01-30"
t_hor = "2022-05-01"
data = model$data
ARMA = model$ARMA
GARCH  = model$GARCH
NM_model = model$NM_model
transform = model$transform
solarModel_moments(t_now, t_hor, data, ARMA, GARCH, NM_model, transform, quiet = FALSE)
}
\keyword{internal}
\keyword{solarModel}
