% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarMixture.R
\name{solarModel_mvmixture}
\alias{solarModel_mvmixture}
\title{Monthly multivariate Gaussian mixture with two components}
\usage{
solarModel_mvmixture(model_Ct, model_GHI)
}
\arguments{
\item{model_Ct}{arg}

\item{model_GHI}{arg}
}
\description{
Monthly multivariate Gaussian mixture with two components
}
