% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{solarModel_parameters}
\alias{solarModel_parameters}
\title{Extract the parameters of a `solarModel`}
\usage{
solarModel_parameters(model, as_tibble = FALSE)
}
\arguments{
\item{model}{object with the class `solarModel`. See the function \code{\link{solarModel}} for details.}

\item{as_tibble}{logical, when `TRUE` the output will be converted in a tibble.}
}
\value{
a named list with all the parameters
}
\description{
Extract the parameters of a `solarModel`
}
\examples{
model <- Bologna
solarModel_parameters(model)
solarModel_parameters(model, as_tibble = TRUE)

}
