% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-internals.R
\name{solarModel_PIT}
\alias{solarModel_PIT}
\title{Compute the Probability Integral Transform (PIT)}
\usage{
solarModel_PIT(model, moments)
}
\description{
Compute the Probability Integral Transform (PIT)
}
\note{
Version 1.0.3
}
\keyword{solarModel}
