% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-diagnostics.R
\name{solarModel_PIT_test}
\alias{solarModel_PIT_test}
\title{Evaluate a KS test on the PIT.}
\usage{
solarModel_PIT_test(
  model,
  type = "full",
  nyears,
  ci = 0.05,
  min_quantile = 0.015,
  max_quantile = 0.985
)
}
\arguments{
\item{model}{An object of the class \code{solarModel}}

\item{type}{character(1), type of test. Can be:
\describe{
\item{\code{"train"}}{the test is performed only on train data;}
\item{\code{"test"}}{the test is performed only on test data;}
\item{\code{"full"}}{the test is performed only on the full data.}
}}

\item{ci}{p.value for rejection.}

\item{min_quantile}{minimum quantile for the grid of values.}

\item{max_quantile}{maximum quantile for the grid of values.}
}
\description{
Evaluate a KS test on the PIT.
}
\details{
See the functions \code{\link{solarModel_PIT}} and \code{\link{ks_test}} for more details.
}
\note{
Version 1.0.0.
}
\keyword{solarModel}
