% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_internals.R
\name{solarModel_predict}
\alias{solarModel_predict}
\title{Produce a forecast from a solarModel object}
\usage{
solarModel_predict(model, moments, theta = 0, ci = 0.01)
}
\arguments{
\item{theta}{Esscher parameter}
}
\description{
Produce a forecast from a solarModel object
}
\examples{
model = solarModel$new(spec)
model$fit()
moments <- model$moments$conditional[14,]
object <- solarModel_predict(model, moments, ci = 0.01)
object
}
