% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{solarModel_scenario}
\alias{solarModel_scenario}
\title{Simulate multiple scenarios}
\usage{
solarModel_scenario(
  model,
  from = "2010-01-01",
  to = "2010-12-31",
  by = "1 month",
  nsim = 1,
  lambda = 0,
  seed = 1,
  quiet = FALSE
)
}
\arguments{
\item{model}{object with the class `solarModel`. See the function \code{\link{solarModel}} for details.}

\item{from}{character, start Date for simulations in the format `YYYY-MM-DD`.}

\item{to}{character, end Date for simulations in the format `YYYY-MM-DD`.}

\item{by}{character, steps for multiple scenarios, e.g. `1 day` (day-ahead simulations), `15 days`, `1 month`, `3 months`, ecc.
For each step are simulated `nsim` scenarios.}

\item{nsim}{integer, number of simulations.}

\item{lambda}{numeric, Esscher parameter.}

\item{seed}{scalar integer, starting random seed.}

\item{quiet}{logical}
}
\description{
Simulate multiple scenarios of solar radiation with a `solarModel` object.
}
\examples{
spec <- solarModel_spec("Ferrara", from="2005-01-01", to="2020-01-01")
model <- solarModel(spec)
solarModel_scenario(model, from = "2010-01-01", to = "2010-12-31", nsim = 2, by = "1 month")

}
