% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_internals.R
\name{solarModel_seasonal_data}
\alias{solarModel_seasonal_data}
\title{Extract a dataset with seasonal data from `solarModel`}
\usage{
solarModel_seasonal_data(
  model,
  monthly = TRUE,
  GM = TRUE,
  nmonths,
  ndays,
  date
)
}
\description{
Extract a dataset with seasonal data from `solarModel`
}
\examples{
model <- Bologna
# All seasonal and monthly data
solarModel_seasonal_data(model)
# Do not add monthly data
solarModel_seasonal_data(model, monthly = FALSE)
# Do not add Gaussian mixture data
solarModel_seasonal_data(model, monthly = FALSE, GM = FALSE)
# Extract onlu a particular day
solarModel_seasonal_data(model, date="2022-01-01")
}
