% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{solarModel_simulate}
\alias{solarModel_simulate}
\title{Simulate trajectories}
\usage{
solarModel_simulate(
  model,
  from = "2010-01-01",
  to = "2010-12-31",
  nsim = 1,
  lambda = 0,
  seed = 1,
  exclude_known = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{model}{object with the class `solarModel`. See the function \code{\link{solarModel}} for details.}

\item{from}{character, start Date for simulations in the format `YYYY-MM-DD`.}

\item{to}{character, end Date for simulations in the format `YYYY-MM-DD`.}

\item{nsim}{integer, number of simulations.}

\item{lambda}{numeric, Esscher parameter.}

\item{seed}{scalar integer, starting random seed.}

\item{quiet}{logical}
}
\description{
Simulate trajectories of solar radiation with a `solarModel` object.
}
\examples{
spec <- solarModel_spec("Ferrara", from="2005-01-01", to="2020-01-01")
model <- solarModel(spec)
solarModel_simulate(model, from = "2010-01-01", to = "2010-12-31", nsim = 1)

}
