% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{solarModel_spec}
\alias{solarModel_spec}
\title{Specification for a `solarModel` object}
\usage{
solarModel_spec(
  place,
  year_max = NULL,
  from = NULL,
  to = NULL,
  CAMS_data = solarr::CAMS_data,
  control = solarModel_control()
)
}
\arguments{
\item{place}{character, name for the selected location in `CAMS_data` list.}

\item{year_max}{integer, maximum year in the dataset}

\item{from}{character. Date in the format `YYYY-MM-DD`. Minimum date in the data in `CAMS_data`. If `NULL` will be used the maximum available.}

\item{to}{character. Date in the format `YYYY-MM-DD`. Maximum date in the data in `CAMS_data`. If `NULL` will be used the maximum available.}

\item{CAMS_data}{list with radiation data for different locations.}

\item{control}{list with control parameters, see `control_solarModel()`.}
}
\description{
Specification for a `solarModel` object
}
