% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-diagnostics.R
\name{solarModel_test_autocorr}
\alias{solarModel_test_autocorr}
\title{Autocorrelation test}
\usage{
solarModel_test_autocorr(
  model,
  nyears,
  method = "bg",
  type = "full",
  lag.max = 3,
  ci = 0.05
)
}
\arguments{
\item{model}{An object of the class \code{solarModel}}

\item{method}{character(1), method of test performed. Can be:
\describe{
\item{\code{"bg"}}{for Breush-Godfrey test;}
\item{\code{"bp"}}{for Box-pierce test;}
\item{\code{"lb"}}{for Ljung-Box test.}
}}

\item{type}{character(1), type of test. Can be:
\describe{
\item{\code{"train"}}{the test is performed only on train data;}
\item{\code{"test"}}{the test is performed only on test data;}
\item{\code{"full"}}{the test is performed only on the full data.}
}}

\item{lag.max}{integer(1), maximum lag to consider for the test.}

\item{ci}{p.value for rejection.}
}
\description{
Evaluate the autocorrelation in the components of a \code{solarModel} object.
}
\note{
Version 1.0.0.
}
\examples{
model = solarModel$new(spec)
model$fit()
solarModel_test_autocorr(model, method = "lb")

}
\keyword{solarModel}
