% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-diagnostics.R
\name{solarModel_test_distribution}
\alias{solarModel_test_distribution}
\title{Distribution test}
\usage{
solarModel_test_distribution(
  model,
  H0 = "gm",
  type = "full",
  ci = 0.05,
  min_quantile = 0.025,
  max_quantile = 0.985
)
}
\arguments{
\item{model}{An object of the class \code{solarModel}}

\item{H0}{Character, null hypothesis for the residuals distribution. Can be:
\describe{
\item{\code{"gm"}}{for a test against a gaussian mixture distribution;}
\item{\code{"norm"}}{for a test against a gaussian distribution.}
}}

\item{type}{character(1), type of test. Can be:
\describe{
\item{\code{"train"}}{the test is performed only on train data;}
\item{\code{"test"}}{the test is performed only on test data;}
\item{\code{"full"}}{the test is performed only on the full data.}
}}

\item{ci}{p.value for rejection.}

\item{min_quantile}{minimum quantile for the grid of values.}

\item{max_quantile}{maximum quantile for the grid of values.}
}
\description{
Evaluate a Kolmogorov-Smirnov test on the residuals of a \code{solarModel} model
object against the estimated Gaussian mixture distribution.
}
\note{
Version 1.0.0.
}
\examples{
model = solarModel$new(spec)
model$fit()
solarModel_test_distribution(model)

}
\keyword{solarModel}
