% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel-diagnostics.R
\name{solarModel_tests}
\alias{solarModel_tests}
\title{Autocorrelation and distribution tests}
\usage{
solarModel_tests(
  model,
  lags = c(7),
  method = "bg",
  type = "full",
  ci = 0.05,
  min_quantile = 0.025,
  max_quantile = 0.985
)
}
\arguments{
\item{model}{An object of the class \code{solarModel}}

\item{lags}{Numeric vector. Lags on which perform the autocorrelation tests.}

\item{method}{character(1), method of test performed. Can be:
\describe{
\item{\code{"bg"}}{for Breush-Godfrey test;}
\item{\code{"bp"}}{for Box-pierce test;}
\item{\code{"lb"}}{for Ljung-Box test.}
}}

\item{type}{character(1), type of test. Can be:
\describe{
\item{\code{"train"}}{the test is performed only on train data;}
\item{\code{"test"}}{the test is performed only on test data;}
\item{\code{"full"}}{the test is performed only on the full data.}
}}

\item{ci}{p.value for rejection.}

\item{min_quantile}{minimum quantile for the grid of values.}

\item{max_quantile}{maximum quantile for the grid of values.}
}
\description{
Evaluate a Kolmogorov-Smirnov test on the residuals of a \code{solarModel} model
object against the estimated Gaussian mixture distribution and a Breush-pagan or Box-pierce
test on the residuals.
}
\note{
Version 1.0.0.
}
\examples{
model = solarModel$new(spec)
model$fit()
solarModel_tests(model, train_data = TRUE)

}
\keyword{solarModel}
