% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_internals.R
\name{solarModel_unconditional_moments}
\alias{solarModel_unconditional_moments}
\title{Compute the unconditional moments}
\usage{
solarModel_unconditional_moments(seasonal_data, ARMA, GARCH)
}
\description{
Compute the unconditional moments
}
\keyword{internal}
\keyword{solarModel}
