% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{solarModel_update}
\alias{solarModel_update}
\title{Update the parameters of a `solarModel` object}
\usage{
solarModel_update(model, params)
}
\arguments{
\item{model}{`solarModel` object}

\item{params}{named list of parameters. See the function \code{\link{solarModel_parameters}} to structure the list of new parameters.}
}
\description{
Update the parameters of a `solarModel` object
}
