% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{solarModel_update_GM}
\alias{solarModel_update_GM}
\title{Update Gaussian Mixture parameters for a given month}
\usage{
solarModel_update_GM(model, params, nmonth)
}
\description{
Update Gaussian Mixture parameters for a given month
}
