% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModelCont.R
\name{solarModelCont}
\alias{solarModelCont}
\title{solarModelCont}
\description{
solarModelCont

solarModelCont
}
\section{Super class}{
\code{\link[solarr:solarModel]{solarr::solarModel}} -> \code{solarModelCont}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-solarModelCont-new}{\code{solarModelCont$new()}}
\item \href{#method-solarModelCont-Ct}{\code{solarModelCont$Ct()}}
\item \href{#method-solarModelCont-Yt_bar}{\code{solarModelCont$Yt_bar()}}
\item \href{#method-solarModelCont-Rt_bar}{\code{solarModelCont$Rt_bar()}}
\item \href{#method-solarModelCont-R_to_Y}{\code{solarModelCont$R_to_Y()}}
\item \href{#method-solarModelCont-Y_to_R}{\code{solarModelCont$Y_to_R()}}
\item \href{#method-solarModelCont-M_Y}{\code{solarModelCont$M_Y()}}
\item \href{#method-solarModelCont-S_Y}{\code{solarModelCont$S_Y()}}
\item \href{#method-solarModelCont-pdf_Y}{\code{solarModelCont$pdf_Y()}}
\item \href{#method-solarModelCont-clone}{\code{solarModelCont$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="solarr" data-topic="solarModel" data-id="compute_risk_drivers"><a href='../../solarr/html/solarModel.html#method-solarModel-compute_risk_drivers'><code>solarr::solarModel$compute_risk_drivers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="solarModel" data-id="conditional_moments"><a href='../../solarr/html/solarModel.html#method-solarModel-conditional_moments'><code>solarr::solarModel$conditional_moments()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="solarModel" data-id="corrective_monthly_mean"><a href='../../solarr/html/solarModel.html#method-solarModel-corrective_monthly_mean'><code>solarr::solarModel$corrective_monthly_mean()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="solarModel" data-id="corrective_monthly_variance"><a href='../../solarr/html/solarModel.html#method-solarModel-corrective_monthly_variance'><code>solarr::solarModel$corrective_monthly_variance()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="solarModel" data-id="detect_outliers_Yt"><a href='../../solarr/html/solarModel.html#method-solarModel-detect_outliers_Yt'><code>solarr::solarModel$detect_outliers_Yt()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="solarModel" data-id="filter"><a href='../../solarr/html/solarModel.html#method-solarModel-filter'><code>solarr::solarModel$filter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="solarModel" data-id="fit"><a href='../../solarr/html/solarModel.html#method-solarModel-fit'><code>solarr::solarModel$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="solarModel" data-id="fit_AR_model"><a href='../../solarr/html/solarModel.html#method-solarModel-fit_AR_model'><code>solarr::solarModel$fit_AR_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="solarModel" data-id="fit_GARCH_model"><a href='../../solarr/html/solarModel.html#method-solarModel-fit_GARCH_model'><code>solarr::solarModel$fit_GARCH_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="solarModel" data-id="fit_clearsky_model"><a href='../../solarr/html/solarModel.html#method-solarModel-fit_clearsky_model'><code>solarr::solarModel$fit_clearsky_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="solarModel" data-id="fit_mixture_model"><a href='../../solarr/html/solarModel.html#method-solarModel-fit_mixture_model'><code>solarr::solarModel$fit_mixture_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="solarModel" data-id="fit_seasonal_mean"><a href='../../solarr/html/solarModel.html#method-solarModel-fit_seasonal_mean'><code>solarr::solarModel$fit_seasonal_mean()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="solarModel" data-id="fit_seasonal_variance"><a href='../../solarr/html/solarModel.html#method-solarModel-fit_seasonal_variance'><code>solarr::solarModel$fit_seasonal_variance()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="solarModel" data-id="fit_solar_transform"><a href='../../solarr/html/solarModel.html#method-solarModel-fit_solar_transform'><code>solarr::solarModel$fit_solar_transform()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="solarModel" data-id="logLik"><a href='../../solarr/html/solarModel.html#method-solarModel-logLik'><code>solarr::solarModel$logLik()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="solarModel" data-id="print"><a href='../../solarr/html/solarModel.html#method-solarModel-print'><code>solarr::solarModel$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="solarModel" data-id="unconditional_moments"><a href='../../solarr/html/solarModel.html#method-solarModel-unconditional_moments'><code>solarr::solarModel$unconditional_moments()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="solarModel" data-id="update"><a href='../../solarr/html/solarModel.html#method-solarModel-update'><code>solarr::solarModel$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModelCont-new"></a>}}
\if{latex}{\out{\hypertarget{method-solarModelCont-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModelCont$new(place, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModelCont-Ct"></a>}}
\if{latex}{\out{\hypertarget{method-solarModelCont-Ct}{}}}
\subsection{Method \code{Ct()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModelCont$Ct(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModelCont-Yt_bar"></a>}}
\if{latex}{\out{\hypertarget{method-solarModelCont-Yt_bar}{}}}
\subsection{Method \code{Yt_bar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModelCont$Yt_bar(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModelCont-Rt_bar"></a>}}
\if{latex}{\out{\hypertarget{method-solarModelCont-Rt_bar}{}}}
\subsection{Method \code{Rt_bar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModelCont$Rt_bar(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModelCont-R_to_Y"></a>}}
\if{latex}{\out{\hypertarget{method-solarModelCont-R_to_Y}{}}}
\subsection{Method \code{R_to_Y()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModelCont$R_to_Y(x, n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModelCont-Y_to_R"></a>}}
\if{latex}{\out{\hypertarget{method-solarModelCont-Y_to_R}{}}}
\subsection{Method \code{Y_to_R()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModelCont$Y_to_R(x, n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModelCont-M_Y"></a>}}
\if{latex}{\out{\hypertarget{method-solarModelCont-M_Y}{}}}
\subsection{Method \code{M_Y()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModelCont$M_Y(Rt, n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModelCont-S_Y"></a>}}
\if{latex}{\out{\hypertarget{method-solarModelCont-S_Y}{}}}
\subsection{Method \code{S_Y()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModelCont$S_Y(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModelCont-pdf_Y"></a>}}
\if{latex}{\out{\hypertarget{method-solarModelCont-pdf_Y}{}}}
\subsection{Method \code{pdf_Y()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModelCont$pdf_Y(x, Rt, n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarModelCont-clone"></a>}}
\if{latex}{\out{\hypertarget{method-solarModelCont-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarModelCont$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
