% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_internals.R
\name{solarModels_fit_best_model}
\alias{solarModels_fit_best_model}
\title{Best models fit}
\usage{
solarModels_fit_best_model(
  place,
  lag.max = 5,
  ci = 0.01,
  control_model = control_solarModel(),
  ...
)
}
\description{
Best models fit
}
\examples{
solarModels_fit_best_model("Roma", 10, 0.05)
solarModels_fit_best_model("Bologna", 10, 0.05)
solarModels_fit_best_model("Palermo", 10, 0.05)

}
