% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarMoment-R6.R
\name{solarMoment}
\alias{solarMoment}
\title{Compute the generic conditional moments of a solarModel object}
\description{
Compute the generic conditional moments of a solarModel object

Compute the generic conditional moments of a solarModel object
}
\details{
Version 1.0.1.
}
\examples{
model <- solarModel$new(spec)
model$fit()
# Conditional moments
mom <- solarMoment$new(model, "2022-01-01", "2025-01-01")
mom$filter()
mom$data
solarOption_model(model, mom$data)

}
\keyword{solarMoments}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-solarMoment-new}{\code{solarMoment$new()}}
\item \href{#method-solarMoment-filter}{\code{solarMoment$filter()}}
\item \href{#method-solarMoment-filter_ARMA}{\code{solarMoment$filter_ARMA()}}
\item \href{#method-solarMoment-filter_NM}{\code{solarMoment$filter_NM()}}
\item \href{#method-solarMoment-filter_GARCH}{\code{solarMoment$filter_GARCH()}}
\item \href{#method-solarMoment-filter_weights}{\code{solarMoment$filter_weights()}}
\item \href{#method-solarMoment-print}{\code{solarMoment$print()}}
\item \href{#method-solarMoment-clone}{\code{solarMoment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-new"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$new(model, t_now, t_hor)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-filter"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-filter}{}}}
\subsection{Method \code{filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$filter(theta = 0, B = NULL, t_cond = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-filter_ARMA"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-filter_ARMA}{}}}
\subsection{Method \code{filter_ARMA()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$filter_ARMA()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-filter_NM"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-filter_NM}{}}}
\subsection{Method \code{filter_NM()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$filter_NM(theta = 0, B = NULL, t_cond = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-filter_GARCH"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-filter_GARCH}{}}}
\subsection{Method \code{filter_GARCH()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$filter_GARCH()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-filter_weights"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-filter_weights}{}}}
\subsection{Method \code{filter_weights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$filter_weights()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-print"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMoment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-solarMoment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMoment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
