% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption_internals.R
\name{solarOption_contracts}
\alias{solarOption_contracts}
\title{Optimal number of contracts}
\usage{
solarOption_contracts(
  payoff,
  type = "model",
  premium = "Q",
  put = TRUE,
  nyear = 2021,
  control = control_hedging()
)
}
\arguments{
\item{type}{character, method used for computing the premium. Can be \code{model} (Model with integral) or \code{sim} (Monte Carlo).}

\item{premium}{character, premium used. Can be \code{P}, \code{Qdw}, \code{Qup}, or \code{Q}.}

\item{nyear}{integer, actual year. The optimization will be performed excluding the year \code{nyear} and the following.}

\item{model}{object with the class \code{solarModel}. See the function \code{\link{solarModel}} for details.}

\item{control_hedging}{numeric, list of hedging parameters.}
}
\description{
Compute the optimal number of contracts given a particular setup.
}
\keyword{2beRevised}
