% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption-internals.R
\name{solarOption_discount}
\alias{solarOption_discount}
\title{Discount factor of a Solar Option}
\usage{
solarOption_discount(tau, P = 1, Gamma_h = 0, r = 0.03/365)
}
\arguments{
\item{tau}{Numeric, time to maturity in days.}

\item{P}{Numeric, price of the contract.}

\item{Gamma_h}{Numeric, hedged payoff.}

\item{r}{Numeric, daily risk-free rate.}
}
\description{
Discount factor of a Solar Option
}
\details{
The discount factor reads:
\deqn{B(\tau, P, \Gamma^h, r) = e^{-r \tau} + \frac{\Gamma^h}{P} (1 - e^{-r \tau})}
}
\note{
Version 1.0.0.
}
\examples{
solarOption_discount(365. 0.6, 2, 0.000008)
solarOption_discount(365, 0.3, 2, 0.00008)

}
\keyword{solarOption}
