% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{solarOption_historical_bootstrap}
\alias{solarOption_historical_bootstrap}
\title{Bootstrap a fair premium from historical data}
\usage{
solarOption_historical_bootstrap(
  model,
  put = TRUE,
  control_options = control_solarOption()
)
}
\arguments{
\item{model}{An object with the class \code{solarModel}. See the function \code{\link{solarModel}} for details.}

\item{put}{Logical. When \code{TRUE}, the default, will be computed the price for a \code{put} contract, otherwise for a \code{call} contract.}

\item{control_options}{Named list. Control parameters, see \code{\link{control_solarOption}} for more details.}
}
\value{
An object of the class \code{solarOptionBoot}.
}
\description{
Bootstrap a fair premium from historical data
}
\examples{
model <- Bologna
solarOption_historical_bootstrap(model, control_options = control_solarOption(ci = 0.4, nsim = 1000))

}
