% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{solarOption_implied_return}
\alias{solarOption_implied_return}
\title{Implied expected return at maturity}
\usage{
solarOption_implied_return(
  model,
  target_prices = NA,
  nmonths = 1:12,
  control_options = control_solarOption()
)
}
\description{
Implied expected return at maturity
}
