% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOptions.R
\name{solarOption_model}
\alias{solarOption_model}
\title{Pricing function under the solar model}
\usage{
solarOption_model(
  model,
  nmonths = 1:12,
  theta = 0,
  implvol = 1,
  control_options = control_solarOption()
)
}
\arguments{
\item{model}{object with the class `solarModel`. See the function \code{\link{solarModel}} for details.}

\item{nmonths}{numeric, months of which the payoff will be computed.}

\item{theta}{Esscher parameter}

\item{implvol}{implied unconditional GARCH variance, the default is `1`.}

\item{control_options}{control list, see \code{\link{control_solarOption}} for more details.}
}
\description{
Pricing function under the solar model
}
\examples{
control <- control_solarModel(outliers_quantile = 0.0005)
spec <- solarModel_spec("Berlino", from="2005-01-01", to="2024-01-01", control_model = control)
model <- solarModel(spec)
solarOption_model(model)
solarOption_historical(model)
}
