% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOptions.R
\name{solarOption_model_spatial}
\alias{solarOption_model_spatial}
\title{Pricing function under the solar model}
\usage{
solarOption_model_spatial(
  object,
  lat,
  lon,
  nmonths = 1:12,
  theta = 0,
  implvol = 1,
  control_options = control_solarOption()
)
}
\arguments{
\item{object}{a `spatialModel` object}

\item{lat}{numeric, latitude of the point.}

\item{lon}{numeric, longitude of the point.}

\item{nmonths}{numeric, months of which the payoff will be computed.}

\item{theta}{Esscher parameter}

\item{implvol}{implied unconditional GARCH variance, the default is `1`.}

\item{control_options}{control list, see \code{\link{control_solarOption}} for more details.}
}
\description{
Pricing function under the solar model
}
