% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption_internals.R
\name{solarOption_model_test}
\alias{solarOption_model_test}
\title{Test errors solar Option model}
\usage{
solarOption_model_test(
  model,
  nmonths = 1:12,
  put = TRUE,
  control_options = control_solarOption()
)
}
\description{
Test errors solar Option model
}
\examples{
model <- Bologna
solarOption_model_test(model)
solarOption_model_test(model, put = FALSE)
solarOption_model_test(model, nmonths = 6, put = FALSE)
solarOption_model_test(model, nmonths = 6, put = TRUE)
}
