% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModelCont.R
\name{solarOption_modelCont}
\alias{solarOption_modelCont}
\title{solarOption_modelCont}
\usage{
solarOption_modelCont(
  model,
  nmonths = 1:12,
  theta = 0,
  combinations = NA,
  implvol = 1,
  put = TRUE,
  target.Yt = TRUE,
  control_options = control_solarOption()
)
}
\description{
solarOption_modelCont
}
