% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption-internals.R
\name{solarOption_payoff}
\alias{solarOption_payoff}
\title{Payoff function of a Solar Option}
\usage{
solarOption_payoff(R, K = 0, put = TRUE)
}
\arguments{
\item{R}{Numeric, vector of values of solar radiation at maturity.}

\item{K}{Numeric, scalar or vector of strikes.}

\item{put}{Logical, when \code{TRUE}, the default, the function will return the output of a put payoff
otherwise a call payoff. See the details.}
}
\description{
Payoff function of a Solar Option
}
\details{
When \code{put = TRUE}, the function evaluate a put payoff:
\deqn{(K - R)^{+} = (K - R) 1_{K > R}}
When \code{put = TRUE}, the function evaluate a call payoff:
\deqn{(R-K)^{+} = (R - K) 1_{R \ge K}}
}
\note{
Version 1.0.0.
}
\examples{
# Put
solarOption_payoff(10, 9, put = TRUE)
solarOption_payoff(8, 9, put = TRUE)
# Call
solarOption_payoff(10, 9, put = FALSE)
solarOption_payoff(8, 9, put = FALSE)
}
\keyword{solarOption}
