% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{solarOption_scenario}
\alias{solarOption_scenario}
\title{Payoff on simulated Data}
\usage{
solarOption_scenario(
  scenario,
  nmonths = 1:12,
  put = TRUE,
  measure = "P",
  nsim,
  control_options = control_solarOption()
)
}
\arguments{
\item{scenario}{object with the class \code{solarModelScenario}. See the function \code{\link{solarModel_scenarios}} for details.}

\item{nmonths}{numeric vector of months in which the payoff is computed. Range from 1 to 12.}

\item{put}{logical, when \code{TRUE}, the default, the computations will consider a \code{put} contract. Otherwise a \code{call}.}

\item{nsim}{number of simulation to use for computation.}

\item{control_options}{control function, see \code{\link{control_solarOption}} for details.}
}
\value{
An object of the class \code{solarOptionPayoff}.
}
\description{
Payoff on simulated Data
}
\examples{
model <- Bologna
scenario <- solarScenario(model, from = "2011-01-01", to = "2012-01-01", by = "1 month", nsim = 1, seed = 3)
solarOption_scenario(scenario)
solarOption_historical(model)

}
