% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption_internals.R
\name{solarOptionPayoff}
\alias{solarOptionPayoff}
\title{solarOptionPayoff}
\usage{
solarOptionPayoff(data, leap_year = FALSE)
}
\arguments{
\item{data}{df_payoff}

\item{leap_year}{control params}
}
\value{
An object of the class \code{solarOptionPayoff}.
}
\description{
solarOptionPayoff
}
\keyword{internal}
