% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{solarOptionPayoffs}
\alias{solarOptionPayoffs}
\title{solarOptionPayoff}
\usage{
solarOptionPayoffs(model, control_options = control_solarOption())
}
\arguments{
\item{model}{solarModel}

\item{control_options}{control list, see \code{\link{control_solarOption}} for more details.}
}
\value{
An object of the class \code{solarOptionPayoffs}.
}
\description{
solarOptionPayoff
}
