% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarr-package.R
\docType{package}
\name{solarr-package}
\alias{solarr}
\alias{solarr-package}
\title{solarr: Stochastic models for solar radiation}
\description{
Implementation of stochastic models and option pricing on solar radiation data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/beniamino98/solarr}
  \item Report bugs at \url{https://github.com/beniamino98/solarr/issues}
}

}
\keyword{internal}
