% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarScenario-R6.R
\name{solarScenario}
\alias{solarScenario}
\title{R6 Class implementation for \code{solarScenario}}
\description{
Specification of scenarios of a \code{solarModel}
}
\note{
Version 1.0.3
}
\keyword{solarScenario}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-solarScenario-new}{\code{solarScenario$new()}}
\item \href{#method-solarScenario-simulate_residuals}{\code{solarScenario$simulate_residuals()}}
\item \href{#method-solarScenario-filter}{\code{solarScenario$filter()}}
\item \href{#method-solarScenario-clone}{\code{solarScenario$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarScenario-new"></a>}}
\if{latex}{\out{\hypertarget{method-solarScenario-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarScenario$new(simSpec, seed = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarScenario-simulate_residuals"></a>}}
\if{latex}{\out{\hypertarget{method-solarScenario-simulate_residuals}{}}}
\subsection{Method \code{simulate_residuals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarScenario$simulate_residuals(nsim = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarScenario-filter"></a>}}
\if{latex}{\out{\hypertarget{method-solarScenario-filter}{}}}
\subsection{Method \code{filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarScenario$filter(all = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarScenario-clone"></a>}}
\if{latex}{\out{\hypertarget{method-solarScenario-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarScenario$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
