% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarScenario-internals.R
\name{solarScenario_filter}
\alias{solarScenario_filter}
\title{Filter}
\usage{
solarScenario_filter(spec, residuals)
}
\description{
Specification of scenarios of a \code{solarModel}
}
\note{
Version 1.0.2
}
\examples{
model <- solarModel$new(spec)
model$fit()
# Specification
spec <- solarScenario_spec$new(model)
residuals <- solarScenario_residuals(spec, 4, 1)
solarScenario_filter(spec, residuals)

}
\keyword{solarScenario}
