% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarScenario-internals.R
\name{solarScenario_plot}
\alias{solarScenario_plot}
\title{Plot scenarios from a solarScenario object}
\usage{
solarScenario_plot(
  scenario,
  target = "GHI",
  nsim = 10,
  empiric = TRUE,
  ci = 0.05
)
}
\description{
Plot scenarios from a solarScenario object
}
\examples{
# solarModel
model <- solarModel$new(spec)
model$fit()
# Scenarios
scenario <- solarScenario(model, "2016-01-01", "2017-01-01", nsim = 10, by = "1 month")
# Plot
solarScenario_plot(scenario)
}
