% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarScenario-internals.R
\name{solarScenario_residuals}
\alias{solarScenario_residuals}
\title{Simulate}
\usage{
solarScenario_residuals(spec, nsim = 1, seed = 1)
}
\description{
Specification of scenarios of a \code{solarModel}
}
\note{
Version 1.0.2
}
\examples{
model <- solarModel$new(spec)
model$fit()
# Specification
spec <- solarScenario_spec$new(model)
solarScenario_residuals(spec, 4, 1)

}
\keyword{solarScenario}
