% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarScenario-R6.R
\name{solarScenario_spec}
\alias{solarScenario_spec}
\title{R6 Class implementation for \code{solarScenario_spec}}
\description{
R6 Class implementation for \code{solarScenario_spec}

R6 Class implementation for \code{solarScenario_spec}
}
\details{
Specification of scenarios of a \code{solarModel}
}
\note{
Version 1.0.3
}
\keyword{solarScenario}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{location}}{Tibble containing the name of the reference location and its coordinates.}

\item{\code{target}}{Character, target variable.}

\item{\code{quiet}}{Logical}

\item{\code{control}}{Named list}

\item{\code{simulations}}{List with simulations}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-solarScenario_spec-new}{\code{solarScenario_spec$new()}}
\item \href{#method-solarScenario_spec-prob}{\code{solarScenario_spec$prob()}}
\item \href{#method-solarScenario_spec-print}{\code{solarScenario_spec$print()}}
\item \href{#method-solarScenario_spec-clone}{\code{solarScenario_spec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarScenario_spec-new"></a>}}
\if{latex}{\out{\hypertarget{method-solarScenario_spec-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{solarScenario_spec}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarScenario_spec$new(
  model,
  from = "2010-01-01",
  to = "2010-12-31",
  theta = 0,
  exclude_known = FALSE,
  seed = 1,
  quiet = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarScenario_spec-prob"></a>}}
\if{latex}{\out{\hypertarget{method-solarScenario_spec-prob}{}}}
\subsection{Method \code{prob()}}{
Marginal mixture probabilities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarScenario_spec$prob(nmonth)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nmonth}}{Integer, number of month.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarScenario_spec-print"></a>}}
\if{latex}{\out{\hypertarget{method-solarScenario_spec-print}{}}}
\subsection{Method \code{print()}}{
Print method for the class \code{solarScenario_spec}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarScenario_spec$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarScenario_spec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-solarScenario_spec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarScenario_spec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
