% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarScenario-internals.R
\name{solarScenario_VaR}
\alias{solarScenario_VaR}
\title{Compute the Value at Risk from simulated values}
\usage{
solarScenario_VaR(scenario, alpha = 0.05)
}
\arguments{
\item{alpha}{Confidence level for the VaR}

\item{scenarios}{An object of the class \code{solarScenario}}
}
\description{
Compute the Value at Risk from simulated values
}
\note{
Version 1.0.0.
}
\examples{
# solarModel
model <- solarModel$new(spec)
model$fit()
# Scenarios
scenario <- solarScenario(model, "2016-01-01", "2017-01-01", nsim = 10, by = "1 month")
# VaR
solarScenario_VaR(scenario, 0.05)
}
\keyword{solarScenario}
