% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModelCont.R
\name{solarScenarioCont}
\alias{solarScenarioCont}
\title{solarScenarioCont}
\usage{
solarScenarioCont(
  model,
  from = "2010-01-01",
  to = "2011-01-01",
  by = "1 year",
  theta = 0,
  nsim = 1,
  seed = 1,
  quiet = FALSE,
  ...
)
}
\description{
solarScenarioCont
}
