% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarSimulation.R
\name{solarSimulation}
\alias{solarSimulation}
\title{Simulate trajectories}
\usage{
solarSimulation(simSpec, nsim = 1, seed = 1)
}
\arguments{
\item{nsim}{integer, number of simulations.}

\item{seed}{scalar integer, starting random seed.}
}
\description{
Simulate trajectories of solar radiation with a `solarModel` object.
}
\examples{
model <- Bologna
simSpec <- solarSimulation_spec(model)
sim <- solarSimulation(simSpec)
ggplot()+
geom_line(data = sim$emp, aes(date, GHI))+
geom_line(data = sim$sim[[1]], aes(date, GHI), color = "red")

}
