% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarSimulation.R
\name{solarSimulation_spec}
\alias{solarSimulation_spec}
\title{Simulate trajectories}
\usage{
solarSimulation_spec(
  model,
  from = "2010-01-01",
  to = "2010-12-31",
  theta = 0,
  exclude_known = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{model}{object with the class `solarModel`. See the function \code{\link{solarModel}} for details.}

\item{from}{character, start Date for simulations in the format `YYYY-MM-DD`.}

\item{to}{character, end Date for simulations in the format `YYYY-MM-DD`.}

\item{theta}{numeric, Esscher parameter.}

\item{exclude_known}{when true the two starting points (equals for all the simulations) will be excluded from the output.}

\item{quiet}{logical}
}
\description{
Simulate trajectories of solar radiation with a `solarModel` object.
}
\examples{
model <- Bologna
simSpec <- solarSimulation_spec(model)

}
