% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solarSpectra}
\alias{solarSpectra}
\title{Solar spectral data}
\format{
A tibble with 2,002 rows and 4 numeric columns:
\describe{
\item{Lambda}{Wavelength values.}
\item{Am0}{Numeric spectral values.}
\item{Am1}{Numeric spectral values.}
\item{Am2}{Numeric spectral values.}
}
}
\usage{
data(solarSpectra)
}
\description{
A tabular data set of spectral irradiance values.
}
\keyword{datasets}
