% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarTransform.R
\name{solarTransform}
\alias{solarTransform}
\title{Solar Model transformation functions}
\description{
Solar Model transformation functions

Solar Model transformation functions
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{solarTransform$new()}}
\item \href{#method-Yt}{\code{solarTransform$Yt()}}
\item \href{#method-GHI}{\code{solarTransform$GHI()}}
\item \href{#method-parameters}{\code{solarTransform$parameters()}}
\item \href{#method-update}{\code{solarTransform$update()}}
\item \href{#method-clone}{\code{solarTransform$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Solar Model transformation functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$new(alpha = 0.001, beta = 0.95)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{bound parameters.}

\item{\code{beta}}{bound parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Yt"></a>}}
\if{latex}{\out{\hypertarget{method-Yt}{}}}
\subsection{Method \code{Yt()}}{
Transformation from Xt to Yt and viceversa
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$Yt(x, inverse = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{numeric vector in \eqn{(\alpha, \alpha+\beta)}.}

\item{\code{inverse}}{when `TRUE` will compute the inverse transform.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes the transformation:

\deqn{Y_t = \log(\log(\beta) - \log(x - \alpha))}

In case in which `inverse = TRUE` it computes

\deqn{X_t = \alpha + \beta \exp(-\exp(Y_t))}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GHI"></a>}}
\if{latex}{\out{\hypertarget{method-GHI}{}}}
\subsection{Method \code{GHI()}}{
Solar radiation function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$GHI(x, Ct, inverse = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{numeric vector in \eqn{(\alpha, \alpha+\beta)}.}

\item{\code{Ct}}{clearsky radiation.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes the inverse of `GHI`

\deqn{GHI = C_t(1 - x)}

In case in which `inverse = TRUE` it computes

\deqn{X_t = 1 - \frac{x}{C_t}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parameters"></a>}}
\if{latex}{\out{\hypertarget{method-parameters}{}}}
\subsection{Method \code{parameters()}}{
Compute parameters from a time series
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$parameters(x, threshold = 0.01)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{time series of risk driver}

\item{\code{threshold}}{for minimum}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Update the parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$update(alpha, beta)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{bounds parameter.}

\item{\code{beta}}{bounds parameter.}

\item{\code{threshold}}{for minimum}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
