% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialClearsky-R6.R
\docType{class}
\name{spatialClearsky}
\alias{spatialClearsky}
\title{Spatial clear-sky model}
\description{
R6 implementation of a clear-sky seasonal model for solar radiation,
using extraterrestrial radiation and seasonal harmonics.
}
\note{
Version 1.0.1
}
\keyword{clearsky}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{coefficients}}{description}

\item{\code{spec}}{description}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-spatialClearsky-new}{\code{spatialClearsky$new()}}
\item \href{#method-spatialClearsky-fit}{\code{spatialClearsky$fit()}}
\item \href{#method-spatialClearsky-predict}{\code{spatialClearsky$predict()}}
\item \href{#method-spatialClearsky-parameters}{\code{spatialClearsky$parameters()}}
\item \href{#method-spatialClearsky-update}{\code{spatialClearsky$update()}}
\item \href{#method-spatialClearsky-print}{\code{spatialClearsky$print()}}
\item \href{#method-spatialClearsky-clone}{\code{spatialClearsky$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialClearsky-new"></a>}}
\if{latex}{\out{\hypertarget{method-spatialClearsky-new}{}}}
\subsection{Method \code{new()}}{
Initialize a spatialClearsky model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialClearsky$new(spec, control = control_spatialClearsky())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{control}}{description
Specification}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialClearsky-fit"></a>}}
\if{latex}{\out{\hypertarget{method-spatialClearsky-fit}{}}}
\subsection{Method \code{fit()}}{
Initialize a spatialClearsky model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialClearsky$fit(data_list, coords)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialClearsky-predict"></a>}}
\if{latex}{\out{\hypertarget{method-spatialClearsky-predict}{}}}
\subsection{Method \code{predict()}}{
Predict a clear-sky
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialClearsky$predict(n, lat, lon, alt = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialClearsky-parameters"></a>}}
\if{latex}{\out{\hypertarget{method-spatialClearsky-parameters}{}}}
\subsection{Method \code{parameters()}}{
Predict the clear-sky parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialClearsky$parameters(lat, lon)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialClearsky-update"></a>}}
\if{latex}{\out{\hypertarget{method-spatialClearsky-update}{}}}
\subsection{Method \code{update()}}{
Update the parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialClearsky$update(coefficients)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialClearsky-print"></a>}}
\if{latex}{\out{\hypertarget{method-spatialClearsky-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialClearsky$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatialClearsky-clone"></a>}}
\if{latex}{\out{\hypertarget{method-spatialClearsky-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialClearsky$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
