% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialClearsky-R6-internals.R
\name{spatialClearsky_optimizer_CLS}
\alias{spatialClearsky_optimizer_CLS}
\title{Fit the spatial model for clear sky radiation.}
\usage{
spatialClearsky_optimizer_CLS(
  formula_Ct,
  formula_params,
  eps = 0.01,
  data_list,
  coords,
  C_tilde_col = "clearsky",
  R_col = "GHI",
  lambda = 0,
  gamma = NA,
  normalize_loss = FALSE,
  penalty = c("none", "ridge"),
  mask_time = NULL,
  settings = list(verbose = TRUE, eps_abs = 1e-05, eps_rel = 1e-05, polish = TRUE,
    max_iter = 4000L)
)
}
\description{
Fit the spatial model for clear sky radiation.
}
\note{
Version 1.0.0
}
\keyword{spatialClearsky}
