% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialClearsky-R6-internals.R
\name{spatialClearsky_spec}
\alias{spatialClearsky_spec}
\title{Specification for \code{spatialClearsky} model}
\usage{
spatialClearsky_spec(
  formula_params,
  C_tilde_col = "clearsky",
  R_col = "GHI",
  order = 1,
  order_H0 = 1,
  method_H0 = "spencer",
  period = 365,
  include.intercept = TRUE,
  include.trend = FALSE
)
}
\arguments{
\item{formula_params}{Character, formula for spatial parameters.}

\item{C_tilde_col}{Integer scalar, number of combinations of sines and cosines.}

\item{R_col}{Name for GHI.}

\item{order}{Integer scalar, number of combinations of sines and cosines.}

\item{order_H0}{Integer scalar, order of the polynomial for extra-terrestrial radiation.}

\item{method_H0}{Character, method of computation of the extra-terrestrial radiation: can be:  \code{spencer}, \code{cooper}.}

\item{period}{Integer scalar, seasonality period. The default is 365.}

\item{include.intercept}{Logical, when \code{TRUE}, the default, the intercept will be included in the clear sky model.}
}
\value{
Named list of control parameters.
}
\description{
Specification for \code{spatialClearsky} model
}
\note{
Version 1.0.0
}
\examples{
spec_Ct = spatialClearsky_spec()
}
\keyword{spatialClearsky}
