% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialModel.R
\name{spatialModel_combinations}
\alias{spatialModel_combinations}
\title{Compute all possible states}
\usage{
spatialModel_combinations(object, lat, lon)
}
\arguments{
\item{object}{a `spatialModel` object}

\item{lat}{numeric, latitude of the point.}

\item{lon}{numeric, longitude of the point.}
}
\description{
Compute all possible states
}
