% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialModel.R
\name{spatialModel_interpolate}
\alias{spatialModel_interpolate}
\title{Compute a solar model for a location}
\usage{
spatialModel_interpolate(object, lat, lon, n = 4, quiet = FALSE, ...)
}
\arguments{
\item{object}{a `spatialModel` object}

\item{lat}{numeric, latitude of the point.}

\item{lon}{numeric, longitude of the point.}

\item{n}{number of neighborhoods}

\item{quiet}{logical}
}
\description{
Compute a solar model for a location
}
