% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialModel.R
\name{spatialModel_interpolate_GHI}
\alias{spatialModel_interpolate_GHI}
\title{Interpolate the solar radiation for a location}
\usage{
spatialModel_interpolate_GHI(
  object,
  lat,
  lon,
  n = 4,
  day_date,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{object}{a `spatialModel` object}

\item{lat}{numeric, latitude of the point.}

\item{lon}{numeric, longitude of the point.}

\item{n}{number of neighborhoods}

\item{day_date}{day date for interpolation. If missing all the available dates will be used.}

\item{quiet}{logical}
}
\description{
Interpolate the solar radiation for a location
}
