% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialModel.R
\name{spatialModel_neighborhoods}
\alias{spatialModel_neighborhoods}
\title{Find the n-closest neighborhoods of a point}
\usage{
spatialModel_neighborhoods(object, lat, lon, n = 4, beta = 2, d0)
}
\arguments{
\item{object}{a `spatialModel` object}

\item{lat}{numeric, latitude of the point.}

\item{lon}{numeric, longitude of the point.}

\item{n}{number of neighborhoods}

\item{beta}{parameter used in exponential and power functions.}

\item{d0}{parameter used only in exponential function.}
}
\description{
Find the n-closest neighborhoods of a point
}
