% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialParameters.R
\name{spatialParameters}
\alias{spatialParameters}
\title{Spatial kernel regression}
\usage{
spatialParameters(solarModels, quiet = FALSE)
}
\arguments{
\item{solarModels}{a list containing `solarModels` objects.}

\item{quiet}{logical}
}
\description{
Fit kernel regression on all the parameters of a list containing
`solarModels` at different coordinates.
}
