% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialParameters.R
\name{spatialParameters_predict}
\alias{spatialParameters_predict}
\title{Predict method}
\usage{
spatialParameters_predict(object, lat, lon, as_tibble = FALSE, quiet = FALSE)
}
\arguments{
\item{object}{an object of the class `spatialParameters`. See \code{\link{clearskyModel}}.}

\item{lat}{numeric latitude of the locations.}

\item{lon}{numeric longitude of the locations.}
}
\description{
Predict method for the class `spatialParameters`.
}
