% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_normality.R
\name{test_normality}
\alias{test_normality}
\title{Perform normality tests}
\usage{
test_normality(x = NULL, pvalue = 0.05)
}
\arguments{
\item{x}{numeric, a vector of observation.}

\item{pvalue}{numeric, the desiderd level of \code{p.value} at which the null hypothesis will be rejected.}
}
\value{
a tibble with the results of the normality tests.
}
\description{
Perform normality tests
}
\examples{
set.seed(1)
x <- rnorm(1000, 0, 1) + rchisq(1000, 1)
test_normality(x)
x <- rnorm(1000, 0, 1)
test_normality(x)

}
