% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{update.solarModel}
\alias{update.solarModel}
\title{Extract and update parameters for Solar Model}
\usage{
\method{update}{solarModel}(object, params = NULL)
}
\description{
Extract and update parameters for Solar Model
}
