% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAMS.R
\name{updateCAMS}
\alias{updateCAMS}
\title{update CAMS data}
\usage{
updateCAMS(
  place,
  lat,
  lon,
  alt,
  from = "2005-01-01",
  to = Sys.Date() - 1,
  CAMS_data = solarr::CAMS_data,
  quiet = FALSE
)
}
\description{
update CAMS data
}
