% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARCH-forecast.R
\name{v_sigma_h_mix}
\alias{v_sigma_h_mix}
\title{Iterative GARCH variance formula (approximated)}
\usage{
v_sigma_h_mix(h, omega, alpha, beta, e_x2 = 1, e_x4 = 3, sigma4_t)
}
\description{
Iterative GARCH variance formula (approximated)
}
