% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_tests.R
\name{VaR_viol}
\alias{VaR_viol}
\title{Violations of the VaR}
\usage{
VaR_viol(x, q)
}
\arguments{
\item{x}{vector, realized data}

\item{q}{vector or matrix, on the columns different VaR on the rows observations.}
}
\description{
Violations of the VaR
}
\keyword{diagnostic}
