% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMA_model-internals.R
\name{ARMA_expectation}
\alias{ARMA_expectation}
\title{Compute the conditional mean of an ARMA model}
\usage{
ARMA_expectation(h, X0, A, b, intercept = 0)
}
\arguments{
\item{h}{integer(1), number of steps ahead.}

\item{X0}{numeric vector of length p + q, state vector of past values.}

\item{A}{numeric matrix of size p+q x p+q. See the function \code{\link{ARMA_companion_matrix}} for more details.}

\item{b}{mumeric vector of length p+q. See the function \code{\link{ARMA_vector_b}} for more details.}

\item{intercept}{Numeric scalar, intercept parameter.}
}
\description{
Compute the conditional mean of an ARMA model
}
\note{
Version 1.0.2
}
\examples{
A <- ARMA_companion_matrix(c(0.4, 0.1), c(0.1, 0.05))
b <- ARMA_vector_b(2,2)
intercept <- 0.2
X0 <- c(0.9, 0.2, -0.1, 0.3)
ARMA_expectation(h = 3, X0, A, b, intercept)
}
\keyword{ARMA}
