% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMA_model-internals.R
\name{ARMA_vector_b}
\alias{ARMA_vector_b}
\title{Construct the vector b of an ARMA model}
\usage{
ARMA_vector_b(arOrder, maOrder)
}
\arguments{
\item{arOrder}{integer(1), order of AR model.}

\item{maOrder}{integer(1), order of MA model.}
}
\description{
Construct the vector b of an ARMA model
}
\note{
Version 1.0.2
}
\examples{
# ARMA(2,2)
ARMA_vector_b(2,2)
# AR(2)
ARMA_vector_b(2,0)
# MA(2)
ARMA_vector_b(0,2)
}
\keyword{ARMA}
