% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalClearsky.R
\name{clearsky_optimizer}
\alias{clearsky_optimizer}
\title{Optimizer for Solar Clear sky}
\usage{
clearsky_optimizer(x, Ct, lower = 0, upper = 3, by = 0.01, ntol = 30)
}
\arguments{
\item{x}{vector of realized solar radiation that will be not optimized.}

\item{Ct}{vector of clear sky values to be optimized.}

\item{lower}{numeric, lower bound for delta grid.}

\item{upper}{numeric, upper bound for delta grid.}

\item{by}{numeric, step for delta grid.}

\item{ntol}{integer, tolerance for `clearsky > GHI` condition. Maximum number of violations admitted.}
}
\value{
the optimal delta
}
\description{
Find the best parameter delta for fitting clear sky radiation.
}
\keyword{internal}
