% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalClearsky.R
\name{clearsky_optimizer}
\alias{clearsky_optimizer}
\title{Optimizer for Solar Clear sky}
\usage{
clearsky_optimizer(seasonal_model_Ct, data, ntol = 0)
}
\description{
Optimizer for Solar Clear sky
}
\keyword{internal}
