% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearskySeasonal.R
\name{clearsky_outliers}
\alias{clearsky_outliers}
\title{Detect outliers with respect to clearsky radiation}
\usage{
clearsky_outliers(x, Ct, date, quiet = FALSE)
}
\arguments{
\item{x}{time series of solar radiation}

\item{Ct}{time series of clearsky radiation}

\item{date}{time series of dates}
}
\description{
Detect outliers with respect to clearsky radiation
}
