% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalClearsky.R
\name{control_seasonalClearsky}
\alias{control_seasonalClearsky}
\title{Control parameters for a `seasonalClearsky` object}
\usage{
control_seasonalClearsky(
  method = "II",
  include.intercept = TRUE,
  order = 1,
  period = 365,
  delta0 = 1.4,
  lower = 0,
  upper = 3,
  by = 0.001,
  ntol = 30,
  quiet = FALSE
)
}
\arguments{
\item{method}{character, method for clear sky estimate, can be `I` or `II`.}

\item{include.intercept}{logical. When `TRUE`, the default, the intercept will be included in the model.}

\item{order}{numeric, of sine and cosine elements.}

\item{period}{numeric, periodicity. The default is `365`.}

\item{delta0}{Value for delta init in the clear sky model.}

\item{lower}{numeric, lower bound for delta grid.}

\item{upper}{numeric, upper bound for delta grid.}

\item{by}{numeric, step for delta grid.}

\item{ntol}{integer, tolerance for `clearsky > GHI` condition. Maximum number of violations admitted.}

\item{quiet}{logical. When `FALSE`, the default, the functions displays warning or messages.}
}
\description{
Control parameters for a `seasonalClearsky` object
}
\details{
The parameters `ntol`, `lower`, `upper` and `by` are used exclusively in \code{\link{clearsky_optimizer}}.
}
\examples{
control = control_seasonalClearsky()
}
