% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarEsscher.R
\name{control_solarEsscher}
\alias{control_solarEsscher}
\title{Control for Esscher calibration.}
\usage{
control_solarEsscher(
  nsim = 200,
  ci = 0.05,
  seed = 1,
  n_key_points = 15,
  init_lambda = 0,
  lower_lambda = -1,
  upper_lambda = 1,
  quiet = FALSE
)
}
\arguments{
\item{nsim}{integer, number of simulations used to bootstrap the premium bounds.}

\item{ci}{numeric, confidence interval for bootstrapping. See `solar_option_payoff_bootstrap()`.}

\item{seed}{integer, random seed for reproducibility.}

\item{n_key_points}{integer, number of key points for interpolation.}

\item{init_lambda}{numeric, initial value for the Esscher parameter.}

\item{lower_lambda}{numeric, lower value for the Esscher parameter.}

\item{upper_lambda}{numeric, upper value for the Esscher parameter.}

\item{quiet}{logical}
}
\description{
Control parameters for calibration of Esscher parameters.
}
