% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{control_solarModel}
\alias{control_solarModel}
\title{Control parameters for a `solarModel` object}
\usage{
control_solarModel(
  clearsky.model = control_clearskyModel(),
  mean.model = list(seasonalOrder = 1, arOrder = 2, include.intercept = FALSE,
    monthly.mean = TRUE),
  variance.model = list(seasonalOrder = 1, unconditional_variance = NA, match_moments =
    FALSE, monthly.mean = TRUE, abstol = 1e-20, maxit = 100),
  threshold = 0.01,
  outliers_quantile = 0,
  quiet = FALSE
)
}
\arguments{
\item{clearsky.model}{list with control parameters, see \code{\link{control_clearskyModel}} for details.}

\item{mean.model}{a list of parameters. Available choices are:
\describe{
 \item{`seasonalOrder`}{An integer specifying the order of the seasonal component in the model. The default is `1`.}
 \item{`arOrder`}{An integer specifying the order of the autoregressive component in the model. The default is `2`.}
 \item{`include.intercept`}{When `TRUE` the intercept will be included in the AR model. The dafault if `FALSE`.}
 \item{`monthly.mean`}{When `TRUE` a set of 12 monthly means parameters will be computed from the deseasonalized time series to center it perfectly around zero.}
}}

\item{variance.model}{a list of parameters.}

\item{threshold}{numeric, threshold for the estimation of alpha and beta.}

\item{outliers_quantile}{quantile for outliers detection. If different from 0, the observations that are below the quantile at confidence levels `outliers_quantile` and
the observation above the quantile at confidence level 1-`outliers_quantile` will have a weight equal to zero and will be excluded from estimations.}

\item{quiet}{logical, when `TRUE` the function will not display any message.}
}
\description{
Control function for a solarModel
}
