% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_internals.R
\name{control_solarModel}
\alias{control_solarModel}
\title{Control parameters for a \code{solarModel} object}
\usage{
control_solarModel(
  clearsky = control_seasonalClearsky(),
  stochastic_clearsky = FALSE,
  seasonal.mean = list(seasonalOrder = 1, include.H0 = FALSE, include.intercept = TRUE,
    monthly.mean = TRUE),
  mean.model = list(arOrder = 2, include.intercept = FALSE),
  seasonal.variance = list(seasonalOrder = 1, correction = TRUE, monthly.mean = TRUE),
  variance.model = rugarch::ugarchspec(variance.model = list(garchOrder = c(1, 1)),
    mean.model = list(armaOrder = c(0, 0), include.mean = FALSE)),
  mixture.model = list(abstol = 0.001, maxit = 150),
  threshold = 0.01,
  outliers_quantile = 0,
  garch_variance = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{clearsky}{list with control parameters for the clear sky seasonal model. See \code{\link{control_seasonalClearsky}} for details.}

\item{seasonal.mean}{a list of parameters. Available choices are:
\describe{
\item{\code{seasonalOrder}}{An integer specifying the order of the seasonal component. The default is \code{1}.}
\item{\code{include.intercept}}{Logical, when \code{TRUE} the intercept will be included in the seasonal model, otherwise will be omitted. The dafault if \code{TRUE}.}
\item{\code{include.H0}}{Logical, when \code{TRUE} the extraterrestrial radiation will be included in the seasonal model, otherwise will be omitted. The dafault if \code{FALSE}.}
\item{\code{monthly.mean}}{Logical, when \code{TRUE} a vector of 12 monthly means is computed on the deseasonalized series and it is subtracted to ensure that it is perfectly centered around zero.}
}}

\item{mean.model}{a list of parameters for the AR model.
\describe{
\item{\code{arOrder}}{An integer specifying the order of the AR component. The default is \code{2}.}
\item{\code{include.intercept}}{When \code{TRUE} an intercept will be included in the AR equation, otherwise is omitted. The dafault if \code{FALSE}.}
}}

\item{seasonal.variance}{a list of parameters for the seasonal variance. Available choices are:
\describe{
\item{\code{seasonalOrder}}{Integer, it specify the order of the seasonal component in the model. The default is \code{1}.}
\item{\code{correction}}{Logical, when \code{TRUE} the parameters of seasonal variance are corrected to ensure that the standardize the residuals have exactly a unitary variance.}
\item{\code{monthly.mean}}{Logical, when \code{TRUE} a vector of 12 monthly std. deviations is computed on the GARCH residuals. Then, they are divided by this quantity to ensure that the monthly variance is one.}
}}

\item{variance.model}{an \code{ugarchspec} object for GARCH variance. Default is \code{GARCH(1,1)} specification.}

\item{mixture.model}{a list of parameters for the monthly Gaussian mixture model. Available choices are:
\describe{
\item{\code{abstol}}{Numeric, absolute level for convergence. The default is \code{1e-3}.}
\item{\code{maxit}}{Integer, maximum number of iterations. The default is \code{200}.}
}}

\item{threshold}{numeric, threshold used to estimate the transformation parameters alpha and beta. See \code{\link{solarTransform}} for details.}

\item{outliers_quantile}{quantile for outliers detection. If different from 0, the observations that are below the quantile at confidence levels \code{outliers_quantile} and
the observation above the quantile at confidence level 1-\code{outliers_quantile} will have a weight equal to zero and will be excluded from estimations.}

\item{quiet}{logical, when \code{TRUE} the function will not display any message.}
}
\description{
Control function for a solarModel
}
\examples{
control <- control_solarModel()

}
