% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{control_solarOption}
\alias{control_solarOption}
\title{Control parameters for a solar option}
\usage{
control_solarOption(
  nyears = c(2005, 2023),
  K = 0,
  leap_year = FALSE,
  nsim = 200,
  ci = 0.05,
  seed = 1,
  B = discountFactor()
)
}
\arguments{
\item{nyears}{numeric vector. Interval of years considered. The first element will be the minimum and the second the maximum years used in
the computation of the fair payoff.}

\item{K}{numeric, level for the strike with respect to the seasonal mean. The seasonal mean is multiplied by `exp(K)`.}

\item{leap_year}{logical, when `FALSE`, the default, the year will be considered of 365 days, otherwise 366.}

\item{nsim}{integer, number of simulations used to bootstrap the premium's bounds. See \code{\link{solarOption_historical_bootstrap}}.}

\item{ci}{numeric, confidence interval for bootstrapping. See \code{\link{solarOption_historical_bootstrap}}.}

\item{seed}{integer, random seed for reproducibility. See \code{\link{solarOption_historical_bootstrap}}.}

\item{B}{function. Discount factor function. Should take as input a number (in years) and return a discount factor.}

\item{put}{logical, when `TRUE`, the default, the computations will consider a `put` contract. Otherwise a `call`.}

\item{target.Yt}{logical, when `TRUE`, the default, the computations will consider the pdf of `Yt` otherwise the pdf of solar radiation.}
}
\description{
Control parameters for a solar option
}
\examples{
control_options <- control_solarOption()
}
