% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinvgumbel.R
\name{dinvgumbel}
\alias{dinvgumbel}
\alias{pinvgumbel}
\alias{qinvgumbel}
\alias{rinvgumbel}
\title{Inverted Gumbel random variable}
\usage{
dinvgumbel(x, location = 0, scale = 1, log = FALSE)

pinvgumbel(x, location = 0, scale = 1, log.p = FALSE, lower.tail = TRUE)

qinvgumbel(p, location = 0, scale = 1, log.p = FALSE, lower.tail = TRUE)

rinvgumbel(n, location = 0, scale = 1)
}
\arguments{
\item{x}{vector of quantiles.}

\item{location}{location parameter.}

\item{scale}{scale parameter.}

\item{log}{logical; if `TRUE`, probabilities are returned as `log(p)`.}

\item{log.p}{logical; if `TRUE`, probabilities p are given as `log(p)`.}

\item{lower.tail}{logical; if TRUE (default), probabilities are `P[X < x]` otherwise, `P[X > x]`.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If `length(n) > 1`, the length is taken to be the number required.}
}
\description{
Inverted Gumbel density, distribution, quantile and random generator.
}
\examples{
# Grid
x <- seq(-5, 5, 0.01)

# Density function
p <- dinvgumbel(x, location = 0, scale = 1)
plot(x, p, type = "l")

# Distribution function
p <- pinvgumbel(x, location = 0, scale = 1)
plot(x, p, type = "l")

# Quantile function
qgumbel(0.1)
pinvgumbel(qinvgumbel(0.1))

# Random Numbers
rinvgumbel(1000)
plot(rinvgumbel(1000), type = "l")

}
