% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkumaraswamy.R
\name{dkumaraswamy}
\alias{dkumaraswamy}
\alias{pkumaraswamy}
\alias{qkumaraswamy}
\alias{rkumaraswamy}
\title{Kumaraswamy random variable}
\usage{
dkumaraswamy(x, a = 1, b = 1, log = FALSE)

pkumaraswamy(x, a = 1, b = 1, log.p = FALSE, lower.tail = TRUE)

qkumaraswamy(p, a = 1, b = 1, log.p = FALSE, lower.tail = TRUE)

rkumaraswamy(n, a = 1, b = 1)
}
\arguments{
\item{x}{vector of quantiles.}

\item{a}{parameter `a > 0`.}

\item{b}{parameter `b > 0`.}

\item{log}{logical; if `TRUE`, probabilities are returned as `log(p)`.}

\item{log.p}{logical; if `TRUE`, probabilities p are given as `log(p)`.}

\item{lower.tail}{logical; if `TRUE`, the default, the computed probabilities are `P[X < x]`. Otherwise, `P[X > x]`.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If `length(n) > 1`, the length is taken to be the number required.}
}
\description{
Kumaraswamy density, distribution, quantile and random generator.
}
\examples{
x <- seq(0, 1, 0.01)
# Density function
plot(x, dkumaraswamy(x, 0.2, 0.3), type = "l")
plot(x, dkumaraswamy(x, 2, 1.1), type = "l")
# Distribution function
plot(x, pkumaraswamy(x, 2, 1.1), type = "l")
# Quantile function
qkumaraswamy(0.2, 0.4, 1.4)
# Random generator
rkumaraswamy(20, 0.4, 1.4)

}
\references{
Kumaraswamy Distribution [\href{https://en.wikipedia.org/wiki/Kumaraswamy_distribution}{W}].
}
