% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmix2norm.R
\name{pmix2norm}
\alias{pmix2norm}
\alias{qmix2norm}
\title{Distribution of a bivariate Gaussian mixture with 2 components each}
\usage{
pmix2norm(
  x,
  means = matrix(0, 4, 2),
  Sigma = rep(list(diag(1, 2)), 4),
  probs = rep(1/4, 4),
  log = FALSE
)

qmix2norm(
  p,
  means = matrix(0, 4, 2),
  Sigma = rep(list(diag(1, 2)), 4),
  probs = rep(1/4, 4),
  v = NULL,
  x0 = c(-1, 1),
  range_s = c(-100, 100),
  log.p = FALSE,
  lower.tail = TRUE
)
}
\description{
Distribution of a bivariate Gaussian mixture with 2 components each

Quantile of a bivariate Gaussian mixture with 2 components each
}
