% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmixnorm.R
\name{dmixnorm}
\alias{dmixnorm}
\alias{pmixnorm}
\alias{qmixnorm}
\alias{rmixnorm}
\title{Gaussian mixture random variable}
\usage{
dmixnorm(x, means = rep(0, 2), sd = rep(1, 2), p = rep(1/2, 2), log = FALSE)

pmixnorm(
  x,
  means = rep(0, 2),
  sd = rep(1, 2),
  p = rep(1/2, 2),
  lower.tail = TRUE,
  log.p = FALSE
)

qmixnorm(
  x,
  means = rep(0, 2),
  sd = rep(1, 2),
  p = rep(1/2, 2),
  lower.tail = TRUE,
  log.p = FALSE
)

rmixnorm(n, means = rep(0, 3), sd = rep(1, 3), p = rep(1/3, 3))
}
\arguments{
\item{x}{vector of quantiles or probabilities.}

\item{means}{vector of means parameters.}

\item{sd}{vector of std. deviation parameters.}

\item{p}{vector of probability parameters.}

\item{log}{logical; if `TRUE`, probabilities are returned as `log(p)`.}

\item{lower.tail}{logical; if TRUE (default), probabilities are `P[X < x]` otherwise, `P[X > x]`.}

\item{log.p}{logical; if `TRUE`, probabilities p are given as `log(p)`.}

\item{n}{number of observations. If `length(n) > 1`, the length is taken to be the number required.}
}
\description{
Gaussian mixture density, distribution, quantile and random generator.
}
\examples{
# Parameters
means = c(-3,0,3)
sd = rep(1, 3)
p = c(0.2, 0.3, 0.5)
# Density function
dmixnorm(3, means, sd, p)
# Distribution function
dmixnorm(c(1.2, -3), means, sd, p)
# Quantile function
qmixnorm(0.2, means, sd, p)
# Random generator
rmixnorm(1000, means, sd, p)

}
\references{
Mixture Models [\href{https://en.wikipedia.org/wiki/Mixture_model}{W}].
}
