% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmvsolarGHI.R
\name{dmvsolarGHI}
\alias{dmvsolarGHI}
\title{Bivariate PDF GHI}
\usage{
dmvsolarGHI(x, Ct, alpha, beta, joint_pdf_Yt)
}
\arguments{
\item{x}{vector of quantiles.}

\item{Ct}{clear sky radiation}

\item{alpha}{parameters `alpha > 0`.}

\item{beta}{parameters `beta > 0` and `alpha + beta < 1`.}

\item{joint_pdf_Yt}{joint density of Y1_t, Y2_t.}
}
\description{
Bivariate PDF GHI
}
