% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsolar.R
\name{psolarGHI}
\alias{psolarGHI}
\alias{qsolarGHI}
\alias{rsolarGHI}
\title{Distribution function for the GHI}
\usage{
psolarGHI(x, Ct, alpha, beta, pdf_Yt, log.p = FALSE, lower.tail = TRUE)

qsolarGHI(p, Ct, alpha, beta, pdf_Yt, log.p = FALSE, lower.tail = TRUE)

rsolarGHI(n, Ct, alpha, beta, pdf_Yt)
}
\description{
Distribution function for the GHI

Quantile function for the GHI

Random generator function for the GHI
}
