% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsolarGHI.R
\name{dsolarGHI}
\alias{dsolarGHI}
\alias{psolarGHI}
\alias{qsolarGHI}
\alias{rsolarGHI}
\title{Solar radiation random variable}
\usage{
dsolarGHI(x, Ct, alpha, beta, pdf_Yt)

psolarGHI(x, Ct, alpha, beta, pdf_Yt)

qsolarGHI(p, Ct, alpha, beta, pdf_Yt)

rsolarGHI(n, Ct, alpha, beta, pdf_Yt)
}
\arguments{
\item{x}{vector of quantiles.}

\item{Ct}{clear sky radiation}

\item{alpha}{parameter `alpha > 0`.}

\item{beta}{parameter `beta > 0` and `alpha + beta < 1`.}

\item{pdf_Yt}{density of Yt.}

\item{p}{vector of probabilities.}
}
\description{
Solar radiation density, distribution, quantile and random generator.
}
\examples{
# Density
dsolarGHI(5, 7, 0.001, 0.9, function(x) dnorm(x))
dsolarGHI(6.993, 7, 0.001, 0.9, function(x) dnorm(x))
# Distribution
psolarGHI(6.993, 7, 0.001, 0.9, function(x) dnorm(x))
# Quantile
qsolarGHI(1, 7, 0.001, 0.9, function(x) dnorm(x))
qsolarGHI(c(0.05, 0.95), 7, 0.001, 0.9, function(x) dnorm(x))
# Random generator
rsolarGHI(10, 7, 0.001, 0.9, function(x) dnorm(x))
Ct <- Bologna$seasonal_data$Ct
GHI <- purrr::map(Ct, ~rsolarGHI(1, .x, 0.001, 0.9, function(x) dsnorm(x, shape = -4)))
plot(1:366, GHI, type="l")
}
