% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsugeno.R
\name{dsugeno}
\alias{dsugeno}
\alias{psugeno}
\title{Sugeno-distorted density and distribution}
\usage{
dsugeno(pdf, cdf, lambda = 0)

psugeno(cdf, lambda = 0)
}
\arguments{
\item{pdf}{Function, density function.}

\item{cdf}{Function, distribution function.}

\item{lambda}{Numeric, distortion parameter.}
}
\value{
\itemize{
\item \code{dsugeno()} returns a density function.
\item \code{psugeno()} returns a distribution function.
}
}
\description{
Compute the Sugeno-distorted distribution for a given distribution or density.
}
\details{
Version 1.0.0.
}
\examples{
cdf <- function(x) pnorm(x)
pdf <- function(x) dnorm(x)
x <- c(-1, 0, 1)
psugeno(cdf, lambda = -0.2)(x)
dsugeno(pdf, cdf, lambda = -0.2)(x)

}
\keyword{distributions}
