% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desscher_mix.R
\name{desscher_mix}
\alias{desscher_mix}
\alias{pesscher_mix}
\title{Esscher transform of a Gaussian Mixture}
\usage{
desscher_mix(params = c(0, 1, 0, 1, 0.5))

pesscher_mix(params = c(0, 1, 0, 1, 0.5))
}
\arguments{
\item{params}{Gaussian Mixture parameters, mu1, sigma1, mu2, sigma2, p.}
}
\description{
Esscher transform of a Gaussian Mixture
}
