% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianMixture.R
\name{gaussianMixture}
\alias{gaussianMixture}
\title{Gaussian mixture}
\usage{
gaussianMixture(
  x,
  means,
  sd,
  p,
  components = 2,
  prior_p = rep(NA, components),
  weights,
  maxit = 100,
  abstol = 1e-14,
  na.rm = FALSE
)
}
\arguments{
\item{x}{vector}

\item{means}{vector of initial means parameters.}

\item{sd}{vector of initial std. deviation parameters.}

\item{p}{vector of initial probability parameters.}

\item{components}{number of components.}

\item{prior_p}{prior probability for the k-state. If the k-component is not `NA` the probability will be considered as given and
the parameter `p[k]` will be equal to `prior_p[k]`.}

\item{weights}{observations weights, if a weight is equal to zero the observation is excluded, otherwise is included with unitary weight.
When `missing` all the available observations will be used.}

\item{maxit}{maximum number of iterations.}

\item{na.rm}{logical. When `TRUE`, the default, `NA` values will be excluded from the computations.}

\item{match_moments}{logical. When `TRUE`, the parameters of the second distribution will be estimated such that
the empirical first two moments of `x` matches the theoretical Gaussian mixture moments.}

\item{absotol}{absolute level for convergence.}
}
\value{
list with clustered components and the optimal parameters.
}
\description{
Fit the parameters of a gaussian mixture with k-components.
}
\examples{
means = c(-3,0,3)
sd = rep(1, 3)
p = c(0.2, 0.3, 0.5)
# Density function
pdf <- dmixnorm(means, sd, p)
# Distribution function
cdf <- pmixnorm(means, sd, p)
# Random numbers
x <- rgaussianMixture(1000, means, sd, p)
gaussianMixture(x$X, means, sd, p, components = 3)
gaussianMixture(x$X, means, sd, prior_p = p, components = 3)
}
