% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarFunctions.R
\name{number_of_day}
\alias{number_of_day}
\title{Number of day}
\usage{
number_of_day(day_date = NULL)
}
\arguments{
\item{day_date}{dates vector in the format `YYYY-MM-DD`.}
}
\value{
Numeric vector with the number of the day during the year.
}
\description{
Compute the number of day of the year given a vector of dates.
}
\examples{
number_of_day("2040-01-31")
number_of_day(c("2040-01-31", "2023-04-01", "2015-09-02"))
}
