% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{number_of_day}
\alias{number_of_day}
\title{Number of day}
\usage{
number_of_day(x)
}
\arguments{
\item{x}{dates vector in the format \code{YYYY-MM-DD}.}
}
\value{
Numeric vector with the number of the day during the year.
}
\description{
Compute the number of day of the year given a vector of dates.
}
\examples{
number_of_day("2040-01-31")
number_of_day(c("2040-01-31", "2023-04-01", "2015-09-02"))
number_of_day(c("2029-02-28", "2029-03-01", "2020-12-31"))
number_of_day(c("2020-02-29", "2020-03-01", "2020-12-31"))
}
