% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDF.R
\name{PDF}
\alias{PDF}
\alias{CDF}
\alias{numericQuantile}
\title{Density, distribution and quantile function}
\usage{
PDF(.f, ...)

CDF(.f, lower = -Inf, ...)

numericQuantile(cdf, lower = -Inf, x0 = 0)
}
\arguments{
\item{.f}{density function}

\item{...}{other parameters to be passed to `.f`.}

\item{lower}{lower bound for integration (domain).}

\item{cdf}{cumulative distribution function.}
}
\description{
Return a function of `x` given the specification of a function of `x`.
}
\examples{
# Density
pdf <- PDF(dnorm, mean = 0.3, sd = 1.3)
pdf(3)
dnorm(3, mean = 0.3, sd = 1.3)
# Distribution
cdf <- CDF(dnorm, mean = 0.3, sd = 1.3)
cdf(3)
pnorm(3, mean = 0.3, sd = 1.3)
# Numeric quantile function
pnorm(numericQuantile(dnorm)(0.9))
}
