% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalClearsky.R
\name{seasonalClearsky}
\alias{seasonalClearsky}
\title{Clear sky seasonal model}
\description{
Clear sky seasonal model

Clear sky seasonal model
}
\examples{
library(ggplot2)
# Arguments
place <- "Palermo"
# solarModel specification
spec <- solarModel_spec(place, target = "GHI")
# Extract the required elements
x <- spec$data$GHI
date <- spec$data$date
lat <- spec$coords$lat
clearsky <- spec$data$clearsky

# Initialize the model
model <- seasonalClearsky$new()
# Fit the model
model$fit(x, date, lat, clearsky)
# Predict the seasonal values
spec$data$Ct <- model$predict(spec$data$n)

}
\section{Super class}{
\code{\link[solarr:seasonalModel]{solarr::seasonalModel}} -> \code{seasonalClearsky}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{control}}{Named list of control parameters. See the function \code{\link{control_seasonalClearsky}} for details.}

\item{\code{lat}}{latitude of the place considered.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-seasonalClearsky-new}{\code{seasonalClearsky$new()}}
\item \href{#method-seasonalClearsky-fit}{\code{seasonalClearsky$fit()}}
\item \href{#method-seasonalClearsky-updateH0}{\code{seasonalClearsky$updateH0()}}
\item \href{#method-seasonalClearsky-print}{\code{seasonalClearsky$print()}}
\item \href{#method-seasonalClearsky-clone}{\code{seasonalClearsky$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="solarr" data-topic="seasonalModel" data-id="predict"><a href='../../solarr/html/seasonalModel.html#method-seasonalModel-predict'><code>solarr::seasonalModel$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="solarr" data-topic="seasonalModel" data-id="update"><a href='../../solarr/html/seasonalModel.html#method-seasonalModel-update'><code>solarr::seasonalModel$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalClearsky-new"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalClearsky-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{seasonalClearsky} model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalClearsky$new(control = control_seasonalClearsky())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{control}}{Named list of control parameters. See the function \code{\link{control_seasonalClearsky}} for details.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalClearsky-fit"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalClearsky-fit}{}}}
\subsection{Method \code{fit()}}{
Fit a seasonal model for clear sky radiation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalClearsky$fit(x, date, lat, clearsky)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{time series of solar radiation.}

\item{\code{date}}{time series of dates.}

\item{\code{lat}}{Numeric, reference latitude.}

\item{\code{clearsky}}{Numeric, optional time series of observed clerasky radiation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalClearsky-updateH0"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalClearsky-updateH0}{}}}
\subsection{Method \code{updateH0()}}{
Update the time series of extraterrestrial radiation for a given latitude.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalClearsky$updateH0(lat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lat}}{reference latitude}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalClearsky-print"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalClearsky-print}{}}}
\subsection{Method \code{print()}}{
Print method for the class \code{seasonalClearsky}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalClearsky$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalClearsky-clone"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalClearsky-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalClearsky$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
