% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalClearsky.R
\name{seasonalClearsky}
\alias{seasonalClearsky}
\title{Clear sky seasonal model}
\usage{
seasonalClearsky(x, date, lat, clearsky, control = control_seasonalClearsky())
}
\arguments{
\item{x}{time series of solar radiation}

\item{date}{time series of dates}

\item{lat}{reference latitude}

\item{clearsky}{optional time series of observed clerasky radiation.}

\item{control}{See the function \code{\link{control_seasonalClearsky}} for details.}
}
\description{
Fit a seasonal model for clear sky radiation
}
\examples{
library(ggplot2)
place <- "Palermo"
# Model for GHI
spec <- solarModel_spec(place, target = "GHI")
x <- spec$data$GHI
date <- spec$data$date
lat <- spec$coords$lat
clearsky <- spec$data$clearsky
model <- seasonalClearsky(x, date, lat, clearsky)
spec$data$Ct <- model$predict(spec$data$n)
ggplot(spec$data)+
 geom_line(aes(n, Ct), color = "blue")+
 geom_line(aes(n, GHI))

# Model for clear sky
spec <- solarModel_spec(place, target = "clearsky")
x <- spec$data$clearsky
date <- spec$data$date
lat <- spec$coords$lat
clearsky <- spec$data$clearsky
model <- seasonalClearsky(x, date, lat, clearsky)
spec$data$Ct <- model$predict(spec$data$n)
ggplot(spec$data)+
 geom_line(aes(n, Ct), color = "blue")+
 geom_line(aes(n, clearsky))

}
