% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalModel.R
\name{seasonalModel}
\alias{seasonalModel}
\title{Seasonal Model Object}
\description{
The `seasonalModel` class implements a seasonal regression model as a linear combination of sine and cosine functions.
This model is designed to capture periodic effects in time series data, particularly for applications involving seasonal trends.
}
\details{
The seasonal model is fitted using a specified formula, which allows for the inclusion of external regressors along with sine and cosine terms to model seasonal variations.
The periodicity can be customized, and the model can be updated with new coefficients after fitting.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{seasonal_data}}{a dataset containing external seasonal data.}

\item{\code{extra_params}}{a slot for containing extra parameters.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{coefficients}}{Extract the coefficients}

\item{\code{model}}{Extract the model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{seasonalModel$new()}}
\item \href{#method-predict}{\code{seasonalModel$predict()}}
\item \href{#method-update}{\code{seasonalModel$update()}}
\item \href{#method-clone}{\code{seasonalModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Fit a seasonal model as a linear combination of sine and cosine functions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalModel$new(formula, order = 1, period = 365, data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{formula, an object of class `formula` (or one that can be coerced to that class).
It is a symbolic description of the model to be fitted and can be used to include or exclude the intercept or external regressors in `data`.}

\item{\code{order}}{numeric, of sine and cosine elements.}

\item{\code{period}}{numeric, periodicity. The default is \eqn{\frac{2\pi}{365}}.}

\item{\code{data}}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.
If not found in data, the variables are taken from environment(formula), typically the environment from which `lm` is called.}

\item{\code{...}}{other parameters to be passed to the function lm.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Predict method for a `seasonalModel`,
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalModel$predict(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{integer, number of day of the year.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Update the parameters of a `seasonalModel`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalModel$update(coefficients)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coefficients}}{vector of parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
