% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalModel.R
\name{seasonalModel}
\alias{seasonalModel}
\title{Seasonal Model}
\description{
The \code{seasonalModel} class implements a seasonal regression model as a linear combination of sine and cosine functions.
This model is designed to capture periodic effects in time series data, particularly for applications involving seasonal trends.
}
\details{
The seasonal model is fitted using a specified formula, which allows for the inclusion of external regressors along with sine and cosine terms
to model seasonal variations. The periodicity can be customized, and the model can be updated with new coefficients after the initial fit.
}
\note{
Version 1.0.0
}
\examples{
sm <- seasonalModel$new(1, 365)
formula <- "Yt ~ 1"
data = data.frame(Yt = rnorm(1000), n = 1:1000)
sm$fit(formula, data = data)
sm
sm$coefficients
sm$update(sm$coefficients*3)
sm$predict(20)

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{extra_params}}{List to contain custom extra parameters.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{coefficients}}{A named vector with the fitted coefficients.}

\item{\code{coefficients2}}{A named vector with the coefficients reparametrized
to obtain a linear combination of only shifted sine functions.}

\item{\code{model}}{A slot with the fitted \code{lm} object.}

\item{\code{period}}{Integer, the seasonality period.}

\item{\code{order}}{Integer, number of combinations of sines and cosines.}

\item{\code{std.errors}}{Numeric vector. Parameters std. errors}

\item{\code{tidy}}{Method tidy for the estimated parameters}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-seasonalModel-new}{\code{seasonalModel$new()}}
\item \href{#method-seasonalModel-fit}{\code{seasonalModel$fit()}}
\item \href{#method-seasonalModel-fit_differential}{\code{seasonalModel$fit_differential()}}
\item \href{#method-seasonalModel-predict}{\code{seasonalModel$predict()}}
\item \href{#method-seasonalModel-differential}{\code{seasonalModel$differential()}}
\item \href{#method-seasonalModel-update}{\code{seasonalModel$update()}}
\item \href{#method-seasonalModel-update_std.errors}{\code{seasonalModel$update_std.errors()}}
\item \href{#method-seasonalModel-print}{\code{seasonalModel$print()}}
\item \href{#method-seasonalModel-clone}{\code{seasonalModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{seasonalModel} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalModel$new(order = 1, period = 365)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{order}}{Integer, number of combinations of sines and cosines.}

\item{\code{period}}{Integer, seasonality period. The default is 365.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalModel-fit"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalModel-fit}{}}}
\subsection{Method \code{fit()}}{
Fit a seasonal model as a linear combination of sine and cosine functions and
eventual external regressors specified in the formula. The external regressors used should
have the same periodicity, i.e. not stochastic regressors are allowed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalModel$fit(formula, data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{formula, an object of class \code{formula} (or one that can be coerced to that class).
It is a symbolic description of the model to be fitted and can be used to include or exclude the intercept or external regressors in \code{data}.}

\item{\code{data}}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.
If not found in data, the variables are taken from environment(formula), typically the environment from which \code{lm} is called.}

\item{\code{...}}{other parameters to be passed to the function \code{lm}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalModel-fit_differential"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalModel-fit_differential}{}}}
\subsection{Method \code{fit_differential()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalModel$fit_differential(formula, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalModel-predict"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalModel-predict}{}}}
\subsection{Method \code{predict()}}{
Predict method for the class \code{seasonalModel}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalModel$predict(n, newdata, dt = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{integer, number of day of the year.}

\item{\code{newdata}}{An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.}

\item{\code{dt}}{Numeric, time step.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalModel-differential"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalModel-differential}{}}}
\subsection{Method \code{differential()}}{
Compute the differential of the sinusoidal function. It do not consider the differential of
eventual external regressors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalModel$differential(n, newdata, dt = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Integer, number of day of the year.}

\item{\code{newdata}}{An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.}

\item{\code{dt}}{Numeric, time step.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalModel-update"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalModel-update}{}}}
\subsection{Method \code{update()}}{
Update the parameters inside the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalModel$update(coefficients)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coefficients}}{A named vector with coefficients.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalModel-update_std.errors"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalModel-update_std.errors}{}}}
\subsection{Method \code{update_std.errors()}}{
Update the std. errors of the parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalModel$update_std.errors(std.errors)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalModel-print"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalModel-print}{}}}
\subsection{Method \code{print()}}{
Print method for the class \code{seasonalModel}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalModel$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seasonalModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-seasonalModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
