% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalModel.R
\name{seasonalModel}
\alias{seasonalModel}
\title{Seasonal model object}
\description{
Seasonal model object

Seasonal model object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{seasonalModel$new()}}
\item \href{#method-from_date_to_n}{\code{seasonalModel$from_date_to_n()}}
\item \href{#method-predict}{\code{seasonalModel$predict()}}
\item \href{#method-update}{\code{seasonalModel$update()}}
\item \href{#method-clone}{\code{seasonalModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Fit a seasonal model as a linear combination of sine and cosine functions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalModel$new(formula = "Yt ~ 1", order = 1, period = 365, data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{formula, an object of class `formula` (or one that can be coerced to that class).
It is a symbolic description of the model to be fitted and can be used to include or exclude the intercept or external regressors in `data`.}

\item{\code{order}}{numeric, of sine and cosine elements.}

\item{\code{period}}{numeric, periodicity. The default is `2*base::pi/365`.}

\item{\code{data}}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.
If not found in data, the variables are taken from environment(formula), typically the environment from which `lm` is called.}

\item{\code{...}}{other parameters to be passed to the function lm.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_date_to_n"></a>}}
\if{latex}{\out{\hypertarget{method-from_date_to_n}{}}}
\subsection{Method \code{from_date_to_n()}}{
Convert a date into the respective number of day of the year.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalModel$from_date_to_n(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{character or Date object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Predict method for a `seasonalModel`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalModel$predict(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{integer, number of day of the year.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Update the parameters of a `seasonalModel`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalModel$update(coefficients)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{integer, number of day of the year.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
