% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalModel.R
\name{seasonalModel}
\alias{seasonalModel}
\title{Seasonal Model}
\usage{
seasonalModel(formula = "Yt ~ 1", order = 1, period = 365, data)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{order}{number of sine/cosine expansions.}

\item{period}{periodicity period 2pi/365.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.
If not found in data, the variables are taken from environment(formula), typically the environment from which `lm` is called.}
}
\description{
Seasonal Model
}
\examples{
formula = "GHI ~ 1"
order = 1
period = 365
data = model$data
}
