% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalSolarFunctions.R
\name{seasonalSolarFunctions}
\alias{seasonalSolarFunctions}
\title{Solar seasonal functions}
\description{
Solar seasonal functions

Solar seasonal functions
}
\examples{
dates <- seq.Date(as.Date("2022-01-01"), as.Date("2022-12-31"), 1)
# Seasonal functions object
sf <- seasonalSolarFunctions$new()

# Adjustment parameter
sf$B(number_of_day(dates))

# Time adjustment in minutes
sf$E(dates)

# Declination
sf$declination(dates)

# Solar constant
sf$G0

# Solar constant adjusted
sf$G0n(dates)

# Extraterrestrial radiation
sf$H0(dates, 43)

# Number of hours of sun
sf$sun_hours(dates, 43)

# Sunset hour angle
sf$sunset_hour_angle(dates, 43)

}
\references{
Duffie, Solar Engineering of Thermal Processes Fourth Edition.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{legal_hour}}{Logical, when `TRUE` the clock time will be corrected for the legal hour.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{G0}}{solar constant, i,e, `1367`.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{seasonalSolarFunctions$new()}}
\item \href{#method-update_method}{\code{seasonalSolarFunctions$update_method()}}
\item \href{#method-B}{\code{seasonalSolarFunctions$B()}}
\item \href{#method-degree}{\code{seasonalSolarFunctions$degree()}}
\item \href{#method-radiant}{\code{seasonalSolarFunctions$radiant()}}
\item \href{#method-E}{\code{seasonalSolarFunctions$E()}}
\item \href{#method-solar_time}{\code{seasonalSolarFunctions$solar_time()}}
\item \href{#method-hour_angle}{\code{seasonalSolarFunctions$hour_angle()}}
\item \href{#method-incidence_angle}{\code{seasonalSolarFunctions$incidence_angle()}}
\item \href{#method-azimut_angle}{\code{seasonalSolarFunctions$azimut_angle()}}
\item \href{#method-G0n}{\code{seasonalSolarFunctions$G0n()}}
\item \href{#method-declination}{\code{seasonalSolarFunctions$declination()}}
\item \href{#method-H0}{\code{seasonalSolarFunctions$H0()}}
\item \href{#method-sunset_hour_angle}{\code{seasonalSolarFunctions$sunset_hour_angle()}}
\item \href{#method-sun_hours}{\code{seasonalSolarFunctions$sun_hours()}}
\item \href{#method-solar_altitude}{\code{seasonalSolarFunctions$solar_altitude()}}
\item \href{#method-solar_angles}{\code{seasonalSolarFunctions$solar_angles()}}
\item \href{#method-clearsky}{\code{seasonalSolarFunctions$clearsky()}}
\item \href{#method-clone}{\code{seasonalSolarFunctions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a `seasonalSolarFunctions` object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$new(method = "spencer", legal_hour = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{character, method type for computations. Can be `cooper` or `spencer`.}

\item{\code{legal_hour}}{Logical, when `TRUE` the clock time will be corrected for the legal hour.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_method"></a>}}
\if{latex}{\out{\hypertarget{method-update_method}{}}}
\subsection{Method \code{update_method()}}{
Extract or update the method used for computations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$update_method(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{character, method type. Can be `cooper` or `spencer`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
When `x` is missing it return a character containing the method that
is actually used.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-B"></a>}}
\if{latex}{\out{\hypertarget{method-B}{}}}
\subsection{Method \code{B()}}{
Seasonal adjustment parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$B(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes
\deqn{B(n) = \frac{2\pi}{365} n}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-degree"></a>}}
\if{latex}{\out{\hypertarget{method-degree}{}}}
\subsection{Method \code{degree()}}{
Convert angles in radiant into an angles in degrees.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$degree(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{numeric vector, angles in radiant.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes:
\deqn{\frac{x 180}{\pi}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-radiant"></a>}}
\if{latex}{\out{\hypertarget{method-radiant}{}}}
\subsection{Method \code{radiant()}}{
Convert angles in degrees into an angles in radiant
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$radiant(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{numeric vector, angles in degrees.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes:
\deqn{\frac{x \pi}{180}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-E"></a>}}
\if{latex}{\out{\hypertarget{method-E}{}}}
\subsection{Method \code{E()}}{
Compute the time adjustment in minutes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$E(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function implement Eq. 1.5.3 from Duffie (4th edition), i.e.
\deqn{E = 229.2(0.000075 + 0.001868 \cos(B) - 0.032077\sin(B) - 0.014615\cos(2B) - 0.04089\sin(2B))}
}

\subsection{Returns}{
The time adjustment in minutes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solar_time"></a>}}
\if{latex}{\out{\hypertarget{method-solar_time}{}}}
\subsection{Method \code{solar_time()}}{
Compute the solar time from a clock time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$solar_time(x, lon, lon_sd = 15)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{datetime, clock hour.}

\item{\code{lon}}{longitude of interest in degrees.}

\item{\code{lon_sd}}{longitude of the Local standard meridian in degrees.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function implement Eq. 1.5.2 from Duffie (4th edition), i.e.
\deqn{solartime = clocktime + 4 (lon_s - lon) + E(n)}
}

\subsection{Returns}{
A datetime object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hour_angle"></a>}}
\if{latex}{\out{\hypertarget{method-hour_angle}{}}}
\subsection{Method \code{hour_angle()}}{
Compute the solar angle for a specific hour of the day.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$hour_angle(x, lon, lon_sd = 15)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{datetime, clock hour.}

\item{\code{lon}}{longitude of interest in degrees.}

\item{\code{lon_sd}}{longitude of the Local standard meridian in degrees.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An angle in degrees
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-incidence_angle"></a>}}
\if{latex}{\out{\hypertarget{method-incidence_angle}{}}}
\subsection{Method \code{incidence_angle()}}{
Compute the incidence angle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$incidence_angle(
  x,
  lat,
  lon,
  lon_sd = 15,
  beta = 0,
  gamma = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{datetime, clock hour.}

\item{\code{lat}}{latitude of interest in degrees.}

\item{\code{lon}}{longitude of interest in degrees.}

\item{\code{lon_sd}}{longitude of the Local standard meridian in degrees.}

\item{\code{beta}}{altitude}

\item{\code{gamma}}{orientation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An angle in degrees
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-azimut_angle"></a>}}
\if{latex}{\out{\hypertarget{method-azimut_angle}{}}}
\subsection{Method \code{azimut_angle()}}{
Compute the solar azimuth angle for a specific time of the day.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$azimut_angle(x, lat, lon, lon_sd = 15)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{datetime, clock hour.}

\item{\code{lat}}{latitude of interest in degrees.}

\item{\code{lon}}{longitude of interest in degrees.}

\item{\code{lon_sd}}{longitude of the Local standard meridian in degrees.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function implement Eq. 1.6.6 from Duffie (4th edition), i.e.
\deqn{\gamma_s = sign(\omega) \left|\cos^{-1}\left( \frac{\cos \theta_z \sin \phi - \sin \delta}{\sin \theta_z \cos \phi} \right) \right|}
}

\subsection{Returns}{
The solar azimut angle in degrees
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-G0n"></a>}}
\if{latex}{\out{\hypertarget{method-G0n}{}}}
\subsection{Method \code{G0n()}}{
Compute the solar constant adjusted for the day of the year.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$G0n(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
When method is `cooper` the function implement Eq. 1.4.1a from Duffie (4th edition), i.e.
\deqn{G_{0,n} = G_0 (1 + 0.033\cos(B))}
otherwise when it is `spencer` it implement Eq. 1.4.1b from Duffie (4th edition):
\deqn{G_{0,n} = G_0 (1.000110 + 0.034221\cos(B) + 0.001280\sin(B) + 0.000719\cos(2B) + 0.000077\sin(2B))}
}

\subsection{Returns}{
The solar constant in \eqn{W/m^2} for the day n.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-declination"></a>}}
\if{latex}{\out{\hypertarget{method-declination}{}}}
\subsection{Method \code{declination()}}{
Compute solar declination in degrees.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$declination(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
When method is `cooper` the function implement Eq. 1.6.1a from Duffie (4th edition), i.e.
\deqn{\delta(n) = 23.45 \sin \left(\frac{2 \pi (284 + n)}{365}\right)}
otherwise when it is `spencer` it implement Eq. 1.6.1b from Duffie (4th edition):
\deqn{\delta(n) = \frac{180}{\pi}(0.006918 - 0.399912\cos(B) + 0.070257\sin(B) - 0.006758\cos(2B))}
}

\subsection{Returns}{
The solar declination in degrees.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-H0"></a>}}
\if{latex}{\out{\hypertarget{method-H0}{}}}
\subsection{Method \code{H0()}}{
Compute the solar extraterrestrial radiation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$H0(n, lat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}

\item{\code{lat}}{latitude of interest in degrees.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Extraterrestrial radiation on an horizontal surface in kilowatt hour for metres squared for day.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sunset_hour_angle"></a>}}
\if{latex}{\out{\hypertarget{method-sunset_hour_angle}{}}}
\subsection{Method \code{sunset_hour_angle()}}{
Compute solar angle at sunset in degrees
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$sunset_hour_angle(n, lat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}

\item{\code{lat}}{Numeric, latitude of interest in degrees.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function implement Eq. 1.6.10 from Duffie (4th edition), i.e.
\deqn{\omega_s = \cos^{-1}(-\tan(\delta(n))\tan(\phi))}
}

\subsection{Returns}{
The sunset hour angle in degrees.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sun_hours"></a>}}
\if{latex}{\out{\hypertarget{method-sun_hours}{}}}
\subsection{Method \code{sun_hours()}}{
Compute number of sun hours for a day n.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$sun_hours(n, lat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year.}

\item{\code{lat}}{Numeric, latitude of interest in degrees.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function implement Eq. 1.6.11 from Duffie (4th edition), i.e.
\deqn{\frac{2}{15} \omega_s}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solar_altitude"></a>}}
\if{latex}{\out{\hypertarget{method-solar_altitude}{}}}
\subsection{Method \code{solar_altitude()}}{
Compute solar altitude in degrees
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$solar_altitude(n, lat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of the day of the year}

\item{\code{lat}}{Numeric, latitude of interest in degrees.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes
\deqn{\sin^{-1}(-\sin(\delta(n))\sin(\phi) + \cos(\delta(n))\cos(\phi))}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solar_angles"></a>}}
\if{latex}{\out{\hypertarget{method-solar_angles}{}}}
\subsection{Method \code{solar_angles()}}{
Compute the solar angle for a latitude in different dates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$solar_angles(x, lat, lon, lon_sd, by = "1 min")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{datetime, clock hour.}

\item{\code{lat}}{Numeric, latitude of interest in degrees.}

\item{\code{lon}}{Numeric, longitude of interest in degrees.}

\item{\code{lon_sd}}{Numeric, longitude of the Local standard meridian in degrees.}

\item{\code{by}}{Character, time step. Default is `1 min`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clearsky"></a>}}
\if{latex}{\out{\hypertarget{method-clearsky}{}}}
\subsection{Method \code{clearsky()}}{
Hottel clearsky
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$clearsky(
  cosZ = NULL,
  G0 = NULL,
  altitude = 2.5,
  clime = "No Correction"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cosZ}}{solar incidence angle}

\item{\code{G0}}{solar constant}

\item{\code{altitude}}{altitude in km}

\item{\code{clime}}{clime correction}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seasonalSolarFunctions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
