% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarFunctions.R
\name{solar_clearsky_hourly}
\alias{solar_clearsky_hourly}
\title{Solar clear sky hourly}
\usage{
solar_clearsky_hourly(
  cosZ = NULL,
  G0 = NULL,
  altitude = 2.5,
  clime = "No Correction"
)
}
\arguments{
\item{cosZ}{cosine angle of incidence}

\item{G0}{extraterrestrial radiation}

\item{altitude}{altitude in meters.}

\item{clime}{correction for different climes, can be `No Correction`, `Summer`, `Winter`, `Subartic Summer`, `Tropical`.}
}
\value{
a numeric vector containing the time adjustment in minutes.
}
\description{
Compute the clear sky radiation for hourly data.
}
\examples{
solar_clearsky_hourly(cosZ = 0.4, G0 = 4, altitude = 2.5, clime = "No Correction")

}
