% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_option.R
\name{solar_option_esscher_calibrator}
\alias{solar_option_esscher_calibrator}
\title{Calibrate Esscher Bounds and parameters}
\usage{
solar_option_esscher_calibrator(
  model,
  sim,
  control_options = control.solarOption(),
  control = control.solarEsscher()
)
}
\description{
Calibrate Esscher Bounds and parameters
}
\examples{
model
sim
control_options = control.solarOption()
control = control.solarEsscher()
}
