% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_option.R
\name{solar_option_payoff_bootstrap}
\alias{solar_option_payoff_bootstrap}
\title{Bootstrap a fair price from historical data}
\usage{
solar_option_payoff_bootstrap(
  model,
  nsim = 500,
  ci = 0.05,
  seed = 1,
  control = control.solarOption()
)
}
\description{
Bootstrap a fair price from historical data
}
\examples{
model <- Location$model
nsim = 500
ci = 0.05
seed = 1
control = control.solarOption()
}
