% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_option.R
\name{solar_option_payoff_historical}
\alias{solar_option_payoff_historical}
\title{Payoff on Historical Data}
\usage{
solar_option_payoff_historical(
  data,
  nmonth = 1:12,
  control = control_solarOption()
)
}
\arguments{
\item{data}{slot `data` from `solarModel` object.}

\item{nmonth}{index for the months.}

\item{control}{control function, see `control_solarOption`.}
}
\description{
Payoff on Historical Data
}
