% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_option.R
\name{solar_option_payoff_scenarios}
\alias{solar_option_payoff_scenarios}
\title{Payoff on Simulated Data}
\usage{
solar_option_payoff_scenarios(
  sim,
  nmonth = 1:12,
  nsim = NULL,
  control = control_solarOption()
)
}
\arguments{
\item{sim}{slot `sim` from `solarModel` object.}

\item{nmonth}{index for the months.}

\item{nsim}{number of simulation to use.}

\item{control}{control function, see `control_solarOption`.}
}
\description{
Payoff on Simulated Data
}
