% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_option.R
\name{solar_option_payoff_structure}
\alias{solar_option_payoff_structure}
\title{Structure payoffs}
\usage{
solar_option_payoff_structure(model, type = "sim", exact_daily_premium = TRUE)
}
\description{
Structure payoffs
}
