% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarEsscher.R
\name{solarEsscher_calibrator}
\alias{solarEsscher_calibrator}
\title{Calibrator for Esscher parameter}
\usage{
solarEsscher_calibrator(
  model,
  target_price,
  nmonths = 1:12,
  control_esscher = control_solarEsscher(),
  control_options = control_solarOption()
)
}
\arguments{
\item{model}{object with the class `solarModel`. See the function \code{\link{solarModel}} for details.}

\item{target_price}{target price for the calibration.}

\item{nmonths}{months used in the model computation.}

\item{control_esscher}{control function. See \code{\link{control_solarEsscher}} for details.}

\item{control_options}{control function. See \code{\link{control_solarOption}} for details.}
}
\description{
Calibrator for Esscher parameter
}
