% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarMixture.R
\name{solarMixture}
\alias{solarMixture}
\title{Monthly Gaussian mixture with two components}
\description{
Monthly Gaussian mixture with two components

Monthly Gaussian mixture with two components
}
\note{
Version 1.0.0
}
\examples{
# Model fit
model <- solarModel$new(spec)
model$fit()
# Inputs
x <- model$data$u_tilde
w <- model$data$weights
date <- model$data$date
# Solar Mixture object
sm <- solarMixture$new()
sm$fit(x, date, w)
params <- sm$parameters
sm$std.errors
# params[1,]$mu1 <- params[1,]$mu1*0.9
# sm$update(means = params[,c(2,3)])
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{maxit}}{Integer, maximum number of iterations.}

\item{\code{abstol}}{Numeric, absolute level for convergence.}

\item{\code{components}}{Integer, number of components.}

\item{\code{mu1}}{Function, see \code{\link{monthlyParams}}.}

\item{\code{mu2}}{Function, see \code{\link{monthlyParams}}.}

\item{\code{sd1}}{Function, see \code{\link{monthlyParams}}.}

\item{\code{sd2}}{Function, see \code{\link{monthlyParams}}.}

\item{\code{prob}}{Function, see \code{\link{monthlyParams}}.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{A tibble with the following columns:
\describe{
\item{date}{Time series of dates.}
\item{Month}{Vector of Month.}
\item{x}{Time series used for fitting.}
\item{w}{Time series of weights.}}}

\item{\code{means}}{Matrix of means where a row represents a month and a column a mixture component.}

\item{\code{sd}}{Matrix of std. deviations where a row represents a month and a column a mixture component.}

\item{\code{p}}{Matrix of probabilities where a row represents a month and a column a mixture component.}

\item{\code{model}}{Named List with 12 \code{\link{gaussianMixture}} objects.}

\item{\code{use_empiric}}{logical to denote if empiric parameters are currently used}

\item{\code{loglik}}{Numeric, total log-likelihood.}

\item{\code{fitted}}{A \code{tibble} with the classified series}

\item{\code{moments}}{A \code{tibble} with the theoric moments. It contains:
\describe{
\item{Month}{Month of the year.}
\item{mean}{Theoric monthly expected value of the mixture model.}
\item{variance}{Theoric monthly variance of the mixture model.}
\item{skewness}{Theoric monthly skewness.}
\item{kurtosis}{Theoric monthly kurtosis.}
\item{nobs}{Number of observations used for fitting.}
\item{loglik}{Monthly log-likelihood.}}}

\item{\code{coefficients}}{A \code{tibble} with the fitted parameters.}

\item{\code{std.errors}}{A \code{tibble} with the fitted std.errors}

\item{\code{summary}}{A \code{tibble} with the fitted std.errors}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-solarMixture-new}{\code{solarMixture$new()}}
\item \href{#method-solarMixture-fit}{\code{solarMixture$fit()}}
\item \href{#method-solarMixture-update}{\code{solarMixture$update()}}
\item \href{#method-solarMixture-update_logLik}{\code{solarMixture$update_logLik()}}
\item \href{#method-solarMixture-update_empiric_parameters}{\code{solarMixture$update_empiric_parameters()}}
\item \href{#method-solarMixture-filter}{\code{solarMixture$filter()}}
\item \href{#method-solarMixture-Hessian}{\code{solarMixture$Hessian()}}
\item \href{#method-solarMixture-use_empiric_parameters}{\code{solarMixture$use_empiric_parameters()}}
\item \href{#method-solarMixture-print}{\code{solarMixture$print()}}
\item \href{#method-solarMixture-clone}{\code{solarMixture$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMixture-new"></a>}}
\if{latex}{\out{\hypertarget{method-solarMixture-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{solarMixture} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMixture$new(components = 2, maxit = 5000, abstol = 1e-08)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{components}}{Integer, number of components.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMixture-fit"></a>}}
\if{latex}{\out{\hypertarget{method-solarMixture-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the parameters with mclust package
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMixture$fit(x, date, weights, B = 50, method = "mixtools")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector}

\item{\code{date}}{date vector}

\item{\code{weights}}{observations weights, if a weight is equal to zero the observation is excluded, otherwise is included with unitary weight.
When \code{missing} all the available observations will be used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMixture-update"></a>}}
\if{latex}{\out{\hypertarget{method-solarMixture-update}{}}}
\subsection{Method \code{update()}}{
Update means, sd, p .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMixture$update(means, sd, p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{means}}{Numeric matrix of means parameters.}

\item{\code{sd}}{Numeric matrix of std. deviation parameters.}

\item{\code{p}}{Numeric matrix of probability parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMixture-update_logLik"></a>}}
\if{latex}{\out{\hypertarget{method-solarMixture-update_logLik}{}}}
\subsection{Method \code{update_logLik()}}{
Apply the \verb{$update_logLik()} method to all the \code{gaussianMixture} models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMixture$update_logLik()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMixture-update_empiric_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-solarMixture-update_empiric_parameters}{}}}
\subsection{Method \code{update_empiric_parameters()}}{
Apply the \verb{$update_empiric_parameters()} method to all the \code{gaussianMixture} models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMixture$update_empiric_parameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMixture-filter"></a>}}
\if{latex}{\out{\hypertarget{method-solarMixture-filter}{}}}
\subsection{Method \code{filter()}}{
Apply the \verb{$filter()} method to all the \code{gaussianMixture} models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMixture$filter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMixture-Hessian"></a>}}
\if{latex}{\out{\hypertarget{method-solarMixture-Hessian}{}}}
\subsection{Method \code{Hessian()}}{
Apply the \verb{$Hessian()} method to all the \code{gaussianMixture} models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMixture$Hessian()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMixture-use_empiric_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-solarMixture-use_empiric_parameters}{}}}
\subsection{Method \code{use_empiric_parameters()}}{
Substitute the empiric parameters with EM parameters. If evaluated again
the EM parameters will be substituted back.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMixture$use_empiric_parameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMixture-print"></a>}}
\if{latex}{\out{\hypertarget{method-solarMixture-print}{}}}
\subsection{Method \code{print()}}{
Print method for \code{solarMixture} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMixture$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solarMixture-clone"></a>}}
\if{latex}{\out{\hypertarget{method-solarMixture-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarMixture$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
