% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{solarModel}
\alias{solarModel}
\title{Solar Model}
\usage{
solarModel(object, control = control.solarModel())
}
\arguments{
\item{object}{Location object, `CAMS("Bologna")`}

\item{control}{control settings, `control.solarModel()`.}
}
\description{
Solar Model
}
