% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{solarModel}
\alias{solarModel}
\title{Fit a model for solar radiation}
\usage{
solarModel(spec)
}
\arguments{
\item{spec}{an object with class `solarModelSpec`. See the function \code{\link{solarModel_spec}} for details.}
}
\description{
Fit a model for solar radiation
}
\examples{
control <- control_solarModel(outliers_quantile = 0.0005)
spec <- solarModel_spec("Palermo", from="2005-01-01", to="2024-01-01", control_model = control)
model <- solarModel(spec)
}
