% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_internals.R
\name{solarModel_conditional_moments}
\alias{solarModel_conditional_moments}
\title{Compute conditional moments from a \code{solarModel} object}
\usage{
solarModel_conditional_moments(model, date)
}
\description{
Compute conditional moments from a \code{solarModel} object
}
\examples{
model <- Bologna
solarModel_conditional_moments(model)
solarModel_conditional_moments(model, date = "2022-01-01")

}
