% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_internals.R
\name{solarModel_forecast}
\alias{solarModel_forecast}
\title{Iterate the forecast on multiple dates}
\usage{
solarModel_forecast(model, date, ci = 0.1, unconditional = FALSE)
}
\description{
Iterate the forecast on multiple dates
}
\examples{
model <- Bologna
dates <- seq.Date(as.Date("2020-01-01"), as.Date("2020-01-31"), 1)
solarModel_forecast(model, date = dates)
}
