% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{solarModel_loglik}
\alias{solarModel_loglik}
\title{Compute the log-likelihood of a `solarModel` object}
\usage{
solarModel_loglik(model, target = c("Yt", "GHI"), nmonths = 1:12)
}
\arguments{
\item{model}{`solarModel` object}

\item{nmonths}{months to consider}
}
\description{
Compute the log-likelihood of a `solarModel` object
}
