% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{solarModel_scenario}
\alias{solarModel_scenario}
\title{Simulate multiple scenarios}
\usage{
solarModel_scenario(
  object,
  from = "2010-01-01",
  to = "2010-12-31",
  by = "1 month",
  nsim = 1,
  lambda = 0,
  vol = NA,
  rf = FALSE,
  seed = 1,
  quiet = FALSE
)
}
\arguments{
\item{from}{character, start Date for simulations in the format `YYYY-MM-DD`.}

\item{to}{character, end Date for simulations in the format `YYYY-MM-DD`.}

\item{by}{character, steps for multiple scenarios, e.g. `1 day` (day-ahead simulations), `15 days`, `1 month`, `3 months`, ecc.
For each step are simulated `nsim` scenarios.}

\item{nsim}{integer, number of simulations.}

\item{lambda}{numeric, Esscher parameter. When `rf = FALSE`, the input parameter `lambda` will be transformed in negative.}

\item{vol}{numeric, unconditional mean of GARCH(1,1) standard deviation. If `NA` will be used the estimated one.}

\item{rf}{logical. When `TRUE` the AR(2) component will be set to zero.}

\item{seed}{scalar integer, starting random seed.}

\item{quiet}{logical}
}
\description{
Simulate multiple scenarios of solar radiation with a `solarModel` object.
}
