% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{solarModel_simulate}
\alias{solarModel_simulate}
\title{Simulate trajectories}
\usage{
solarModel_simulate(
  object,
  from = "2010-01-01",
  to = "2010-12-31",
  nsim = 1,
  lambda = 0,
  vol = NA,
  rf = FALSE,
  seed = 1,
  quiet = FALSE
)
}
\arguments{
\item{from}{date, starting date for simulations.}

\item{to}{date, end date for simulations.}

\item{nsim}{integer, number of simulations.}

\item{lambda}{numeric, Esscher parameter. When `rf = FALSE`, the input parameter `lambda` will be transformed in negative.}

\item{vol}{numeric, unconditional mean of GARCH(1,1) standard deviation. If `NA` will be used the estimated one.}

\item{rf}{logical. When `TRUE` the AR(2) component will be set to zero.}

\item{seed}{scalar integer, starting random seed.}

\item{quiet}{logical}
}
\description{
Simulate trajectories of solar radiation with a `solarModel` object.
}
