% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel.R
\name{solarModel_spec}
\alias{solarModel_spec}
\title{Specification function for a `solarModel`}
\usage{
solarModel_spec(
  place,
  min_date,
  max_date,
  from,
  to,
  CAMS_data = solarr::CAMS_data,
  control_model = control_solarModel()
)
}
\arguments{
\item{place}{character, name of an element in the `CAMS_data` list.}

\item{min_date}{character. Date in the format `YYYY-MM-DD`. Minimum date for the complete data. If `missing` will be used the minimum data available.}

\item{max_date}{character. Date in the format `YYYY-MM-DD`. Maximum date for the complete data. If `missing` will be used the maximum data available.}

\item{from}{character. Date in the format `YYYY-MM-DD`. Starting date to use for training data.
If `missing` will be used the minimum data available after filtering for `min_date`.}

\item{to}{character. Date in the format `YYYY-MM-DD`. Ending date to use for training data.
If `missing` will be used the maximum data available after filtering for `max_date`.}

\item{CAMS_data}{named list with radiation data for different locations.}

\item{control_model}{list with control parameters, see \code{\link{control_solarModel}} for details.}
}
\description{
Specification function for a `solarModel`
}
