% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_internals.R
\name{solarModel_spec}
\alias{solarModel_spec}
\title{Specification function for a \code{solarModel}}
\usage{
solarModel_spec(
  place,
  target = "GHI",
  min_date,
  max_date,
  from,
  to,
  CAMS_data = solarr::CAMS_data,
  control_model = control_solarModel()
)
}
\arguments{
\item{place}{Character, name of an element in the \code{CAMS_data} list.}

\item{target}{Character, target variable to model. Can be \code{GHI} or \code{clearsky}.}

\item{min_date}{Character. Date in the format \code{YYYY-MM-DD}. Minimum date for the complete data. If \code{missing} will be used the minimum data available.}

\item{max_date}{Character. Date in the format \code{YYYY-MM-DD}. Maximum date for the complete data. If \code{missing} will be used the maximum data available.}

\item{from}{Character. Date in the format \code{YYYY-MM-DD}. Starting date to use for training data.
If \code{missing} will be used the minimum data available after filtering for \code{min_date}.}

\item{to}{character. Date in the format \code{YYYY-MM-DD}. Ending date to use for training data.
If \code{missing} will be used the maximum data available after filtering for \code{max_date}.}

\item{CAMS_data}{named list with radiation data for different locations.}

\item{control_model}{list with control parameters, see \code{\link{control_solarModel}} for details.}
}
\description{
Specification function for a \code{solarModel}
}
\examples{
control <- control_solarModel(outliers_quantile = 0)
spec <- solarModel_spec("Bologna", from="2005-01-01", to="2022-01-01", control_model = control)

}
