% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_internals.R
\name{solarModel_test_autocorr}
\alias{solarModel_test_autocorr}
\title{Autocorrelation test}
\usage{
solarModel_test_autocorr(model, lag.max = 3, ci = 0.05, method = "bp")
}
\arguments{
\item{model}{An object of the class \code{solarModel}}

\item{lag.max}{Numeric, scalar. Maximum lag to consider for the test.}

\item{ci}{Numeric, scalar. Minimum p-value to consider the test \code{"passed"}.}

\item{method}{Character. Type of test. Can be \code{"bp"} for breush-pagan or \code{"lb"} for Box-pierce.}
}
\description{
Evaluate the autocorrelation in the components of a \code{solarModel} object.
}
\examples{
model = solarModel$new(spec)
model$fit()
solarModel_test_autocorr(model, method = "lb")
}
