% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_internals.R
\name{solarModel_test_distribution}
\alias{solarModel_test_distribution}
\title{Distribution test}
\usage{
solarModel_test_distribution(
  model,
  ci = 0.05,
  min_quantile = 0.025,
  max_quantile = 0.985
)
}
\arguments{
\item{model}{An object of the class \code{solarModel}}

\item{ci}{p.value for rejection.}

\item{min_quantile}{minimum quantile for the grid of values.}

\item{max_quantile}{maximum quantile for the grid of values.}
}
\description{
Evaluate a Kolmogorov-smirnov test on the residuals of a \code{solarModel} model
object against the estimated Gaussian mixture distribution.
}
\examples{
model = solarModel$new(spec)
model$fit()
solarModel_test_distribution(model)
}
