% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_internals.R
\name{solarModel_test_residuals}
\alias{solarModel_test_residuals}
\title{Stationarity and distribution test (Gaussian mixture) for a \code{solarModel}}
\usage{
solarModel_test_residuals(
  model,
  nrep = 50,
  ci = 0.05,
  min_quantile = 0.015,
  max_quantile = 0.985,
  seed = 1
)
}
\description{
Stationarity and distribution test (Gaussian mixture) for a \code{solarModel}
}
\examples{
model <- Bologna
solarModel_test_residuals(model)
}
