% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_internals.R
\name{solarModel_tests}
\alias{solarModel_tests}
\title{Autocorrelation and Distribution tests}
\usage{
solarModel_tests(
  model,
  lag.max = 3,
  ci = 0.05,
  min_quantile = 0.025,
  max_quantile = 0.985,
  method = "bp"
)
}
\arguments{
\item{model}{An object of the class \code{solarModel}}

\item{lag.max}{Numeric, scalar. Maximum lag to consider for the test.}

\item{ci}{p.value for rejection.}

\item{min_quantile}{minimum quantile for the grid of values.}

\item{max_quantile}{maximum quantile for the grid of values.}

\item{method}{Character. Type of test. Can be \code{"bp"} for breush-pagan or \code{"lb"} for Box-pierce.}
}
\description{
Evaluate a Kolmogorov-Smirnov test on the residuals of a \code{solarModel} model
object against the estimated Gaussian mixture distribution and a Breush-pagan or Box-pierce
test on the residuals.
}
\examples{
model = solarModel$new(spec)
model$fit()
solarModel_tests(model)
}
