% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarModel_internals.R
\name{solarModel_unconditional_moments}
\alias{solarModel_unconditional_moments}
\title{Compute conditional moments from a \code{solarModel} object}
\usage{
solarModel_unconditional_moments(model, nmonths, ndays, date)
}
\description{
Compute conditional moments from a \code{solarModel} object
}
\examples{
model <- Bologna
solarModel_unconditional_moments(model)
solarModel_unconditional_moments(model, nmonths = 1)
solarModel_unconditional_moments(model, nmonths = 1, ndays = 1)
solarModel_unconditional_moments(model, date = "2022-01-01")

}
