% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{solarOption_calibrator}
\alias{solarOption_calibrator}
\title{Calibrator for solar Options}
\usage{
solarOption_calibrator(
  model,
  nmonths = 1:12,
  abstol = 0.001,
  reltol = 0.01,
  control_options = control_solarOption()
)
}
\description{
Calibrator for solar Options
}
\examples{
model <- Bologna
model_cal <- solarOption_calibrator(model, nmonths = 8, reltol=1e-3)
solarModel_loglik(model)
solarModel_loglik(model_cal)

}
