% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{solarOption_calibrator}
\alias{solarOption_calibrator}
\title{Calibrator for solar Options}
\usage{
solarOption_calibrator(
  model,
  nmonths = 1:12,
  abstol = 1e-04,
  reltol = 1e-04,
  control_options = control_solarOption()
)
}
\arguments{
\item{model}{object with the class `solarModel`. See the function \code{\link{solarModel}} for details.}

\item{nmonths}{numeric vector of months in which the payoff is computed. Range from 1 to 12.}

\item{abstol}{The absolute convergence tolerance. Only useful for non-negative functions, as a tolerance for reaching zero.}

\item{reltol}{Relative convergence tolerance. The algorithm stops if it is unable to reduce the value by a factor of reltol * (abs(val) + reltol) at a step.
Defaults to `sqrt(.Machine$double.eps)`, typically about 1e-8.}

\item{control_options}{control list, see \code{\link{control_solarOption}} for more details.}
}
\description{
Calibrator for solar Options
}
\examples{
model <- Bologna
model_cal <- solarOption_calibrator(model, nmonths = 8, reltol=1e-3)
solarModel_loglik(model)
solarModel_loglik(model_cal)
}
