% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{solarOption_contracts}
\alias{solarOption_contracts}
\title{Optimal number of contracts}
\usage{
solarOption_contracts(
  model,
  type = "model",
  premium = "Q",
  nyear = 2021,
  tick = 0.06,
  efficiency = 0.2,
  n_panels = 2000,
  pun = 0.06
)
}
\arguments{
\item{model}{object with the class `solarModel`. See the function \code{\link{solarModel}} for details.}

\item{type}{character, method used for computing the premium. Can be `model` (Model with integral) or `sim` (Monte Carlo).}

\item{premium}{character, premium used. Can be `P`, `Qdw`, `Qup`, or `Q`.}

\item{nyear}{integer, actual year. The optimization will be performed excluding the year `nyear` and the following.}

\item{tick}{numeric, conversion tick for the monetary payoff of a contract.}

\item{efficiency}{numeric, mean efficiency of the solar panels.}

\item{n_panels}{numeric, number of meters squared of solar panels.}

\item{pun}{numeric, reference electricity price at which the energy is sold for computing the cash-flows.}
}
\description{
Compute the optimal number of contracts given a particular setup.
}
