% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption_internals.R
\name{solarOption_contracts}
\alias{solarOption_contracts}
\title{Optimal number of contracts}
\usage{
solarOption_contracts(
  payoff,
  type = "model",
  premium = "Q",
  put = TRUE,
  nyear = 2021,
  tick = 0.06,
  efficiency = 0.2,
  n_panels = 2000,
  pun = 0.06
)
}
\arguments{
\item{type}{character, method used for computing the premium. Can be \code{model} (Model with integral) or \code{sim} (Monte Carlo).}

\item{premium}{character, premium used. Can be \code{P}, \code{Qdw}, \code{Qup}, or \code{Q}.}

\item{nyear}{integer, actual year. The optimization will be performed excluding the year \code{nyear} and the following.}

\item{tick}{numeric, conversion tick for the monetary payoff of a contract.}

\item{efficiency}{numeric, mean efficiency of the solar panels.}

\item{n_panels}{numeric, number of meters squared of solar panels.}

\item{pun}{numeric, reference electricity price at which the energy is sold for computing the cash-flows.}

\item{model}{object with the class \code{solarModel}. See the function \code{\link{solarModel}} for details.}
}
\description{
Compute the optimal number of contracts given a particular setup.
}
\keyword{2beRevised}
