% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{solarOption_historical}
\alias{solarOption_historical}
\title{Payoff on Historical Data}
\usage{
solarOption_historical(
  model,
  nmonths = 1:12,
  put = TRUE,
  control_options = control_solarOption()
)
}
\arguments{
\item{model}{object with the class `solarModel`. See the function \code{\link{solarModel}} for details.}

\item{nmonths}{numeric vector of months in which the payoff is computed. Range from 1 to 12.}

\item{put}{logical, when `TRUE`, the default, the computations will consider a `put` contract. Otherwise a `call`.}

\item{control_options}{control list, see \code{\link{control_solarOption}} for more details.}
}
\value{
An object of the class `solarOptionPayoff`.
}
\description{
Payoff on Historical Data
}
\examples{
model <- Bologna
solarOption_historical(model, put=TRUE)
solarOption_historical(model, put=FALSE)

}
