% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{solarOption_historical}
\alias{solarOption_historical}
\title{Payoff on Historical Data}
\usage{
solarOption_historical(
  model,
  nmonths = 1:12,
  put = TRUE,
  control_options = control_solarOption()
)
}
\arguments{
\item{model}{An object with the class \code{solarModel}. See the function \code{\link{solarModel}} for details.}

\item{nmonths}{Numeric vector. Months in which the payoff should be computed. Can vary from 1 (January) to 12 (December).}

\item{put}{Logical. When \code{TRUE}, the default, will be computed the price for a \code{put} contract, otherwise for a \code{call} contract.}

\item{control_options}{Named list. Control parameters, see \code{\link{control_solarOption}} for more details.}
}
\value{
An object of the class \code{solarOptionPayoff}.
}
\description{
Payoff on Historical Data
}
\examples{
model <- Bologna
solarOption_historical(model, put=TRUE)
solarOption_historical(model, put=FALSE)

}
