% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOptions.R
\name{solarOption_historical}
\alias{solarOption_historical}
\title{Payoff on Historical Data}
\usage{
solarOption_historical(
  model,
  nmonths = 1:12,
  control_options = control_solarOption()
)
}
\arguments{
\item{model}{object with the class `solarModel`. See the function \code{\link{solarModel}} for details.}

\item{nmonths}{numeric, months of which the payoff will be computed.}

\item{control_options}{control list, see \code{\link{control_solarOption}} for more details.}
}
\description{
Payoff on Historical Data
}
\examples{
model <- Bologna
solarOption_historical(model)
}
