% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{solarOption_historical_bootstrap}
\alias{solarOption_historical_bootstrap}
\title{Bootstrap a fair premium from historical data}
\usage{
solarOption_historical_bootstrap(
  model,
  put = TRUE,
  control_options = control_solarOption()
)
}
\arguments{
\item{model}{object with the class `solarModel`. See the function \code{\link{solarModel}} for details.}

\item{put}{logical, when `TRUE`, the default, the computations will consider a `put` contract. Otherwise a `call`.}

\item{control_options}{control list, see \code{\link{control_solarOption}} for more details.}
}
\value{
An object of the class `solarOptionBoot`.
}
\description{
Bootstrap a fair premium from historical data
}
\examples{
model <- Bologna
solarOption_historical_bootstrap(model)

}
