% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{solarOption_historical_bootstrap}
\alias{solarOption_historical_bootstrap}
\title{Bootstrap a fair premium from historical data}
\usage{
solarOption_historical_bootstrap(
  model,
  put = TRUE,
  control_options = control_solarOption()
)
}
\arguments{
\item{model}{object with the class `solarModel`. See the function \code{\link{solarModel}} for details.}

\item{control_options}{control function, see \code{\link{control_solarOption}} for details.}

\item{nsim}{number of simulation to bootstrap.}

\item{ci}{confidence interval for quantile}

\item{seed}{random seed.}
}
\value{
An object of the class `solarOptionBoot`.
}
\description{
Bootstrap a fair premium from historical data
}
\examples{
model <- Bologna
solarOption_historical_bootstrap(model)

}
