% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{solarOption_model}
\alias{solarOption_model}
\title{Pricing function under the solar model}
\usage{
solarOption_model(
  model,
  nmonths = 1:12,
  theta = 0,
  combinations = NA,
  implvol = 1,
  put = TRUE,
  target.Yt = TRUE,
  control_options = control_solarOption()
)
}
\arguments{
\item{model}{object with the class \code{solarModel}. See the function \code{\link{solarModel}} for details.}

\item{nmonths}{numeric vector of months in which the payoff is computed. Range from 1 to 12.}

\item{theta}{Esscher parameter}

\item{combinations}{list of 12 elements with gaussian mixture components.}

\item{implvol}{implied unconditional GARCH variance, the default is \code{1}.}

\item{put}{logical, when \code{TRUE}, the default, the computations will consider a \code{put} contract. Otherwise a \code{call}.}

\item{target.Yt}{logical, when \code{TRUE}, the default, the computations will consider the pdf of \code{Yt} otherwise the pdf of solar radiation.}

\item{control_options}{control list, see \code{\link{control_solarOption}} for more details.}
}
\value{
An object of the class \code{solarOptionPayoff}.
}
\description{
Pricing function under the solar model
}
\examples{
model <- Bologna
solarOption_model(model, put=FALSE)
solarOption_model(model, put=TRUE)

}
