% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOptions.R
\name{solarOption_model}
\alias{solarOption_model}
\title{Pricing function under the solar model}
\usage{
solarOption_model(
  model,
  nmonths = 1:12,
  theta = 0,
  combinations = NA,
  implvol = 1,
  control_options = control_solarOption()
)
}
\arguments{
\item{model}{object with the class `solarModel`. See the function \code{\link{solarModel}} for details.}

\item{nmonths}{numeric, months of which the payoff will be computed.}

\item{theta}{Esscher parameter}

\item{combinations}{list of 12 elements with gaussian mixture components.}

\item{implvol}{implied unconditional GARCH variance, the default is `1`.}

\item{control_options}{control list, see \code{\link{control_solarOption}} for more details.}

\item{target.Yt}{pdf to use for expectation}
}
\description{
Pricing function under the solar model
}
\examples{
model <- Bologna
control_options <- control_solarOption(put = FALSE)
df_call <- solarOption_model(model, control_options = control_options)
control_options <- control_solarOption(put = TRUE)
df_put <- solarOption_model(model, control_options = control_options)
}
