% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOptions.R
\name{solarOption_scenario}
\alias{solarOption_scenario}
\title{Payoff on Simulated Data}
\usage{
solarOption_scenario(
  scenario,
  nmonths = 1:12,
  nsim,
  control_options = control_solarOption()
)
}
\arguments{
\item{scenario}{object with the class `solarModelScenario`. See the function \code{\link{solarModel_scenarios}} for details.}

\item{nmonths}{numeric, months of which the payoff will be computed.}

\item{nsim}{number of simulation to use for computation.}

\item{control_options}{control function, see \code{\link{control_solarOption}} for details.}
}
\description{
Payoff on Simulated Data
}
