% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{solarOption_scenario}
\alias{solarOption_scenario}
\title{Payoff on simulated Data}
\usage{
solarOption_scenario(
  model,
  scenario,
  nmonths = 1:12,
  put = TRUE,
  nsim,
  control_options = control_solarOption()
)
}
\arguments{
\item{model}{An object with the class \code{solarModel}. See the function \code{\link{solarModel}} for details.}

\item{scenario}{object with the class \code{solarModelScenario}. See the function \code{\link{solarModel_scenarios}} for details.}

\item{nmonths}{Numeric vector. Months in which the payoff should be computed. Can vary from 1 (January) to 12 (December).}

\item{put}{Logical. When \code{TRUE}, the default, will be computed the price for a \code{put} contract, otherwise for a \code{call} contract.}

\item{nsim}{number of simulation to use for computation.}

\item{control_options}{control function, see \code{\link{control_solarOption}} for details.}
}
\value{
An object of the class \code{solarOptionPayoff}.
}
\description{
Payoff on simulated Data
}
\examples{
model <- Bologna
scenario <- solarScenario(model, from = "2011-01-01", to = "2012-01-01", by = "1 month", nsim = 10, seed = 3)
solarOption_scenario(model, scenario)
solarOption_historical(model)

}
