% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarOption.R
\name{solarOption_structure}
\alias{solarOption_structure}
\title{Structure payoffs}
\usage{
solarOption_structure(
  payoffs,
  type = "model",
  put = TRUE,
  exact_daily_premium = TRUE
)
}
\arguments{
\item{payoffs}{object with the class \code{solarOptionPayoffs}. See the function \code{\link{solarOptionPayoffs}} for details.}

\item{type}{method used for computing the premium. If \code{model}, the default will be used the analytic model,
otherwise with \code{scenarios} the monte carlo scenarios stored inside the \code{model$scenarios$P}.}

\item{exact_daily_premium}{when \code{TRUE} the historical premium is computed as daily average among all the years.
Otherwise the monthly premium is computed and then divided by the number of days of the month.}
}
\value{
The object \code{payoffs} with class \code{solarOptionPayoffs}.
}
\description{
Structure payoffs
}
