% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarScenario.R
\name{solarScenario}
\alias{solarScenario}
\title{Simulate multiple scenarios}
\usage{
solarScenario(
  model,
  from = "2010-01-01",
  to = "2011-01-01",
  by = "1 year",
  theta = 0,
  nsim = 1,
  seed = 1,
  quiet = FALSE
)
}
\arguments{
\item{model}{object with the class \code{solarModel}. See the function \code{\link{solarModel}} for details.}

\item{from}{character, start Date for simulations in the format \code{YYYY-MM-DD}.}

\item{to}{character, end Date for simulations in the format \code{YYYY-MM-DD}.}

\item{by}{character, steps for multiple scenarios, e.g. \verb{1 day} (day-ahead simulations), \verb{15 days}, \verb{1 month}, \verb{3 months}, ecc.
For each step are simulated \code{nsim} scenarios.}

\item{theta}{numeric, Esscher parameter.}

\item{nsim}{integer, number of simulations.}

\item{seed}{scalar integer, starting random seed.}

\item{quiet}{logical}
}
\description{
Simulate multiple scenarios of solar radiation with a \code{solarModel} object.
}
\examples{
model <- Bologna
scen <- solarScenario(model, "2016-01-01", "2017-01-01", nsim = 10, by = "1 month")
# Plot
solarScenario_plot(scen, nsim = 3)
# Solar Option
solarOption_scenario(model, scen)

}
