% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarScenario.R
\name{solarScenario}
\alias{solarScenario}
\title{Simulate multiple scenarios}
\usage{
solarScenario(
  model,
  from = "2010-01-01",
  to = "2010-12-31",
  by = "1 month",
  theta = 0,
  nsim = 1,
  seed = 1,
  quiet = FALSE
)
}
\arguments{
\item{model}{object with the class `solarModel`. See the function \code{\link{solarModel}} for details.}

\item{from}{character, start Date for simulations in the format `YYYY-MM-DD`.}

\item{to}{character, end Date for simulations in the format `YYYY-MM-DD`.}

\item{by}{character, steps for multiple scenarios, e.g. `1 day` (day-ahead simulations), `15 days`, `1 month`, `3 months`, ecc.
For each step are simulated `nsim` scenarios.}

\item{theta}{numeric, Esscher parameter.}

\item{nsim}{integer, number of simulations.}

\item{seed}{scalar integer, starting random seed.}

\item{quiet}{logical}
}
\description{
Simulate multiple scenarios of solar radiation with a `solarModel` object.
}
\examples{
model <- Bologna
scen <- solarScenario(model, "2010-01-01", to = "2020-12-31", nsim = 10)
scen <- solarScenario(model, to = "2010-02-01", by = "1 day")
}
