% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarScenario.R
\name{solarScenario_residuals}
\alias{solarScenario_residuals}
\title{Simulate residuals for a \code{solarScenario_spec}}
\usage{
solarScenario_residuals(simSpec, nsim = 1, seed = 1)
}
\arguments{
\item{simSpec}{object with the class \code{solarScenario_spec}. See the function \code{\link{solarScenario_spec}} for details.}

\item{nsim}{integer, number of simulations.}

\item{seed}{scalar integer, starting random seed.}
}
\description{
Simulate residuals for a \code{solarScenario_spec}
}
\examples{
model <- Bologna
simSpec <- solarScenario_spec(model)
simSpec <- solarScenario_residuals(simSpec, nsim = 10)
}
