% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarTransform.R
\name{solarTransform}
\alias{solarTransform}
\title{solarTransform
Solar functions}
\description{
Solar Model transformation functions
}
\examples{
st <- solarTransform$new()
st$GHI(0.4, 3)
st$GHI(st$iGHI(0.4, 3), 3)
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{alpha}}{Numeric, the first transformation parameter.}

\item{\code{beta}}{Numeric, the second transformation parameter.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{solarTransform$new()}}
\item \href{#method-GHI}{\code{solarTransform$GHI()}}
\item \href{#method-GHI_y}{\code{solarTransform$GHI_y()}}
\item \href{#method-iGHI}{\code{solarTransform$iGHI()}}
\item \href{#method-Y}{\code{solarTransform$Y()}}
\item \href{#method-iY}{\code{solarTransform$iY()}}
\item \href{#method-fit}{\code{solarTransform$fit()}}
\item \href{#method-bounds}{\code{solarTransform$bounds()}}
\item \href{#method-update}{\code{solarTransform$update()}}
\item \href{#method-clone}{\code{solarTransform$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a solarTransform object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$new(alpha = 0, beta = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{Numeric, transformation parameter.}

\item{\code{beta}}{Numeric, transformation parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GHI"></a>}}
\if{latex}{\out{\hypertarget{method-GHI}{}}}
\subsection{Method \code{GHI()}}{
Solar radiation function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$GHI(x, Ct)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Numeric values in \eqn{(\alpha, \alpha+\beta)}.}

\item{\code{Ct}}{Numeric, clear sky radiation.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes:
\deqn{GHI(x) = C_t(1 - x)}
}

\subsection{Returns}{
Numeric values in \eqn{C_t(1-\alpha-\beta, 1-\alpha)}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GHI_y"></a>}}
\if{latex}{\out{\hypertarget{method-GHI_y}{}}}
\subsection{Method \code{GHI_y()}}{
Solar radiation function in terms of y
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$GHI_y(y, Ct)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{Numeric values in \eqn{(-\infty, \infty)}.}

\item{\code{Ct}}{Numeric, clear sky radiation.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes:
\deqn{GHI(y) = C_t(1 - \alpha-\beta \exp(-\exp(x)))}
}

\subsection{Returns}{
Numeric values in \eqn{[C_t(1-\alpha-\beta), C_t(1-\alpha)]}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iGHI"></a>}}
\if{latex}{\out{\hypertarget{method-iGHI}{}}}
\subsection{Method \code{iGHI()}}{
Compute the risk driver process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$iGHI(x, Ct)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Numeric values in \eqn{[C_t(1-\alpha-\beta), C_t(1-\alpha)]}.}

\item{\code{Ct}}{Numeric, clear sky radiation.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes the inverse of the `GHI`funcion
\deqn{iGHI(x) = 1 - \frac{x}{C_t}}
}

\subsection{Returns}{
Numeric values in \eqn{[\alpha,\alpha+\beta]}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Y"></a>}}
\if{latex}{\out{\hypertarget{method-Y}{}}}
\subsection{Method \code{Y()}}{
Transformation function from X to Y
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$Y(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{numeric vector in \eqn{[\alpha, \alpha+\beta]}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes:
\deqn{Y(x) = \log(\log(\beta) - \log(x - \alpha))}
}

\subsection{Returns}{
Numeric values in \eqn{[-\infty, \infty]}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iY"></a>}}
\if{latex}{\out{\hypertarget{method-iY}{}}}
\subsection{Method \code{iY()}}{
Inverse transformation from Y to X.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$iY(y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{numeric vector in \eqn{[-\infty, \infty]}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function computes:
\deqn{iY(y) = \alpha + \beta \exp(-\exp(y))}
}

\subsection{Returns}{
Numeric values in \eqn{[\alpha, \alpha + \beta]}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the best parameters from a time series
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$fit(x, threshold = 0.01)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{time series of solar risk drivers in \eqn{(0, 1)}.}

\item{\code{threshold}}{for minimum}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Return a list that contains:
\describe{
 \item{alpha}{Numeric, first transformation parameter.}
 \item{beta}{Numeric, second transformation parameter.}
 \item{epsilon}{Numeric, threshold used for fitting.}
 \item{Xt_min}{Numeric, minimum value of the supplied time series.}
 \item{Xt_min}{Numeric, maximum value of the supplied time series.}
}
}

\subsection{Returns}{
A named list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bounds"></a>}}
\if{latex}{\out{\hypertarget{method-bounds}{}}}
\subsection{Method \code{bounds()}}{
Compute the bounds for the transformed variables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$bounds(target = "Xt")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target}}{target variable. Available choices are:
\describe{
 \item{Xt}{Solar risk driver, the bounds returned are \eqn{[\alpha, \alpha + \beta]}.}
 \item{Kt}{Clearness index, the bounds returned are \eqn{[1-\alpha-\beta, 1-\alpha]}.}
 \item{Yt}{Solar transform, the bounds returned are \eqn{[-\infty, \infty]}.}
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector where the first element is the lower bound and the second the upper bound.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Update the transformation parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$update(alpha, beta)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{Numeric, transformation parameter.}

\item{\code{beta}}{Numeric, transformation parameter.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Update the slots `$alpha` and `$beta`.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{solarTransform$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
