% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCorrelations.R
\name{spatialCorrelation}
\alias{spatialCorrelation}
\title{spatialCorrelation object}
\description{
spatialCorrelation object

spatialCorrelation object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{spatialCorrelation$new()}}
\item \href{#method-get_sigma_B}{\code{spatialCorrelation$get_sigma_B()}}
\item \href{#method-get_margprob}{\code{spatialCorrelation$get_margprob()}}
\item \href{#method-get_cr_X}{\code{spatialCorrelation$get_cr_X()}}
\item \href{#method-get}{\code{spatialCorrelation$get()}}
\item \href{#method-clone}{\code{spatialCorrelation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialCorrelation$new(binprobs, mixture_cr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sigma_B"></a>}}
\if{latex}{\out{\hypertarget{method-get_sigma_B}{}}}
\subsection{Method \code{get_sigma_B()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialCorrelation$get_sigma_B(places, nmonth = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_margprob"></a>}}
\if{latex}{\out{\hypertarget{method-get_margprob}{}}}
\subsection{Method \code{get_margprob()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialCorrelation$get_margprob(places, nmonth = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_cr_X"></a>}}
\if{latex}{\out{\hypertarget{method-get_cr_X}{}}}
\subsection{Method \code{get_cr_X()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialCorrelation$get_cr_X(places, nmonth = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialCorrelation$get(places, nmonth = 1, date)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialCorrelation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
