% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialFunctions.R
\name{spatialGrid}
\alias{spatialGrid}
\title{Spatial Grid}
\usage{
spatialGrid(lat = c(43.7, 45.1), lon = c(9.2, 12.7), by = c(0.1, 0.1))
}
\arguments{
\item{by}{step for longitudes and latitudes. If two values are specified the first will be used for
latitudes and the second for longitudes}

\item{range_lat}{vector with latitudes. Only the minimum and maximum values are considered.}

\item{range_lon}{vector with longitudes. Only the minimum and maximum values are considered.}
}
\value{
a tibble with two columns `lat` and `lon`.
}
\description{
Create a grid from a range of latitudes and longitudes.
}
\examples{
spatialGrid(lat = c(43.7, 43.8), lon = c(12.5, 12.7), by = 0.1)
spatialGrid(lat = c(43.7, 43.8), lon = c(12.6, 12.7), by = c(0.05, 0.01))

}
