% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialModel.R
\name{spatialModel}
\alias{spatialModel}
\title{Spatial model object}
\usage{
spatialModel(locations, solarModels)
}
\arguments{
\item{locations}{grid of locations}

\item{solarModels}{list of `solarModel` objects}
}
\description{
Spatial model object
}
