% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralDistribution.R
\name{spectralDistribution}
\alias{spectralDistribution}
\title{Compute the spectral distribution for a black body}
\usage{
spectralDistribution(lambda = NULL, measure = "nanometer")
}
\arguments{
\item{lambda}{numeric, wave length in micrometers.}

\item{measure}{character, measure of the irradiated energy. If `nanometer` the final energy will be in W/m2 x nanometer,
otherwise if `micrometer` the final energy will be in W/m2 x micrometer.}
}
\description{
Compute the spectral distribution for a black body
}
